% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{adlFileInputStreamCheck}
\alias{adlFileInputStreamCheck}
\title{Check for proper adlFileInputStream.}
\usage{
adlFileInputStreamCheck(adlFileInputStream)
}
\arguments{
\item{adlFileInputStream}{the adlFileInputStream object to check}
}
\description{
Check for proper adlFileInputStream.
}
\seealso{
Other Azure resource functions: \code{\link{adlFileOutputStreamCheck}},
  \code{\link{azureAuthenticateOnAuthType}},
  \code{\link{azureAuthenticate}},
  \code{\link{azureCheckToken}},
  \code{\link{azureGetTokenClientCredential}},
  \code{\link{azureGetTokenDeviceCodeFetch}},
  \code{\link{azureGetTokenDeviceCode}},
  \code{\link{azureGetTokenRefreshToken}},
  \code{\link{showDeviceCodeMessageToUser}}
}
