% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureDataLake.R
\name{adlFileInputStreamClose}
\alias{adlFileInputStreamClose}
\title{Close an adlFileInputStream.}
\usage{
adlFileInputStreamClose(adlFileInputStream, verbose = FALSE)
}
\arguments{
\item{adlFileInputStream}{adlFileInputStream of the file}

\item{verbose}{Print tracing information (default FALSE)}
}
\value{
NULL (void)
}
\description{
Close an adlFileInputStream.
}
\seealso{
Other Azure Data Lake Store functions: \code{\link{adlFileInputStreamRead}},
  \code{\link{adlFileOutputStreamClose}},
  \code{\link{adlFileOutputStreamFlush}},
  \code{\link{adlFileOutputStreamWrite}},
  \code{\link{azureDataLakeAppendBOS}},
  \code{\link{azureDataLakeAppendCore}},
  \code{\link{azureDataLakeAppend}},
  \code{\link{azureDataLakeCreate}},
  \code{\link{azureDataLakeDelete}},
  \code{\link{azureDataLakeGetFileStatus}},
  \code{\link{azureDataLakeListStatus}},
  \code{\link{azureDataLakeMkdirs}},
  \code{\link{azureDataLakeOpenBIS}},
  \code{\link{azureDataLakeReadCore}},
  \code{\link{azureDataLakeRead}},
  \code{\link{createAdlFileInputStream}},
  \code{\link{createAdlFileOutputStream}}
}
