% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{adlFileOutputStreamCheck}
\alias{adlFileOutputStreamCheck}
\title{Check for proper adlFileOutputStream.}
\usage{
adlFileOutputStreamCheck(adlFileOutputStream)
}
\arguments{
\item{adlFileOutputStream}{the adlFileOutputStream object to check}
}
\description{
Check for proper adlFileOutputStream.
}
\seealso{
Other Azure resource functions: \code{\link{adlFileInputStreamCheck}},
  \code{\link{azureAuthenticateOnAuthType}},
  \code{\link{azureAuthenticate}},
  \code{\link{azureCheckToken}},
  \code{\link{azureGetTokenClientCredential}},
  \code{\link{azureGetTokenDeviceCodeFetch}},
  \code{\link{azureGetTokenDeviceCode}},
  \code{\link{azureGetTokenRefreshToken}},
  \code{\link{showDeviceCodeMessageToUser}}
}
