% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureDataLake.R
\name{adlFileOutputStreamClose}
\alias{adlFileOutputStreamClose}
\title{Close an adlFileOutputStream.}
\usage{
adlFileOutputStreamClose(adlFileOutputStream, verbose = FALSE)
}
\arguments{
\item{adlFileOutputStream}{adlFileOutputStream of the file}

\item{verbose}{Print tracing information (default FALSE)}
}
\value{
NULL (void)
}
\description{
Close an adlFileOutputStream.
}
\seealso{
Other Azure Data Lake Store functions: \code{\link{adlFileInputStreamAvailable}},
  \code{\link{adlFileInputStreamClose}},
  \code{\link{adlFileInputStreamGetPos}},
  \code{\link{adlFileInputStreamLength}},
  \code{\link{adlFileInputStreamMarkSupported}},
  \code{\link{adlFileInputStreamMark}},
  \code{\link{adlFileInputStreamReadBuffered}},
  \code{\link{adlFileInputStreamRead}},
  \code{\link{adlFileInputStreamReset}},
  \code{\link{adlFileInputStreamSeek}},
  \code{\link{adlFileInputStreamSkip}},
  \code{\link{adlFileOutputStreamFlush}},
  \code{\link{adlFileOutputStreamWrite}},
  \code{\link{azureDataLakeAppendBOS}},
  \code{\link{azureDataLakeAppendCore}},
  \code{\link{azureDataLakeAppend}},
  \code{\link{azureDataLakeCreate}},
  \code{\link{azureDataLakeDelete}},
  \code{\link{azureDataLakeGetFileStatus}},
  \code{\link{azureDataLakeListStatus}},
  \code{\link{azureDataLakeMkdirs}},
  \code{\link{azureDataLakeOpenBIS}},
  \code{\link{azureDataLakeReadCore}},
  \code{\link{azureDataLakeRead}},
  \code{\link{createAdlFileInputStream}},
  \code{\link{createAdlFileOutputStream}},
  \code{\link{readFromService}}
}
