% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureDataLake.R
\name{adlFileOutputStreamFlush}
\alias{adlFileOutputStreamFlush}
\title{Flush an adlFileOutputStream.}
\usage{
adlFileOutputStreamFlush(adlFileOutputStream, syncFlag = "METADATA",
  verbose = FALSE)
}
\arguments{
\item{adlFileOutputStream}{adlFileOutputStream of the file}

\item{syncFlag}{type of sync (DATA, METADATA, CLOSE)}

\item{verbose}{Print tracing information (default FALSE)}
}
\value{
NULL (void)
}
\description{
Flush an adlFileOutputStream.
}
\seealso{
Other Azure Data Lake Store functions: \code{\link{adlFileOutputStreamClose}},
  \code{\link{adlFileOutputStreamWrite}},
  \code{\link{azureDataLakeAppendBOS}},
  \code{\link{azureDataLakeAppend}},
  \code{\link{azureDataLakeCreate}},
  \code{\link{azureDataLakeDelete}},
  \code{\link{azureDataLakeGetFileStatus}},
  \code{\link{azureDataLakeListStatus}},
  \code{\link{azureDataLakeMkdirs}},
  \code{\link{azureDataLakeRead}},
  \code{\link{createAdlFileOutputStream}}
}
