% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureDataLake.R
\name{adlFileOutputStreamWrite}
\alias{adlFileOutputStreamWrite}
\title{Write to an adlFileOutputStream.}
\usage{
adlFileOutputStreamWrite(adlFileOutputStream, contents, off, len)
}
\arguments{
\item{adlFileOutputStream}{adlFileOutputStream of the file}

\item{contents}{contents to write to the file}

\item{off}{the start offset in the data}

\item{len}{the number of bytes to write}
}
\value{
NULL (void)
}
\description{
Write to an adlFileOutputStream.
}
\seealso{
Other Azure Data Lake Store functions: \code{\link{azureDataLakeAppend}},
  \code{\link{azureDataLakeCreate}},
  \code{\link{azureDataLakeDelete}},
  \code{\link{azureDataLakeGetFileStatus}},
  \code{\link{azureDataLakeListStatus}},
  \code{\link{azureDataLakeMkdirs}},
  \code{\link{azureDataLakeRead}},
  \code{\link{createAdlFileOutputStream}}
}
