% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{adls.append.core}
\alias{adls.append.core}
\title{The Core Append API.}
\usage{
adls.append.core(azureActiveContext, adlFileOutputStream = NULL,
  azureDataLakeAccount, relativePath, bufferSize, contents, contentSize = -1L,
  leaseId = NULL, sessionId = NULL, syncFlag = NULL,
  offsetToAppendTo = -1, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{adlFileOutputStream}{The adlFileOutputStream object to operate with.}

\item{azureDataLakeAccount}{Name of the Azure Data Lake account.}

\item{relativePath}{Relative path of a file.}

\item{bufferSize}{Size of the buffer to be used.}

\item{contents}{raw contents to be written to the file.}

\item{contentSize}{size of \code{contents} to be written to the file.}

\item{leaseId}{a String containing the lease ID (generated by client). Can be null.}

\item{sessionId}{a String containing the session ID (generated by client). Can be null.}

\item{syncFlag}{Use \code{DATA} when writing more bytes to same file path. Most performant operation.
Use \code{METADATA} when metadata for the
file also needs to be updated especially file length
retrieved from \code{adls.file.info} or \code{adls.ls} API call.
Has an overhead of updating metadata operation.
Use \code{CLOSE} when no more data is
expected to be written in this path. Adl backend would
update metadata, close the stream handle and
release the lease on the
path if valid leaseId is passed.
Expensive operation and should be used only when last
bytes are written.}

\item{offsetToAppendTo}{offset at which to append to to file.
To let the server choose offset, pass \code{-1}.}

\item{verbose}{Print tracing information (default FALSE).}
}
\value{
response object
}
\description{
The Core Append API.
}
\details{
Exceptions - IOException
}
\references{
\url{https://docs.microsoft.com/en-us/azure/data-lake-store/data-lake-store-data-operations-rest-api#upload-data}
}
\seealso{
\url{https://hadoop.apache.org/docs/current/hadoop-project-dist/hadoop-hdfs/WebHDFS.html#Append_to_a_File}

\url{https://hadoop.apache.org/docs/current/hadoop-project-dist/hadoop-hdfs/WebHDFS.html#Buffer_Size}

\url{https://hadoop.apache.org/docs/current/api/org/apache/hadoop/fs/FileSystem.html#append-org.apache.hadoop.fs.Path-int-org.apache.hadoop.util.Progressable-}

Other Azure Data Lake Store functions: \code{\link{adls.append.direct}},
  \code{\link{adls.append}}, \code{\link{adls.concat}},
  \code{\link{adls.create}}, \code{\link{adls.delete}},
  \code{\link{adls.file.info}},
  \code{\link{adls.fileinputstream.available}},
  \code{\link{adls.fileinputstream.close}},
  \code{\link{adls.fileinputstream.create}},
  \code{\link{adls.fileinputstream.getpos}},
  \code{\link{adls.fileinputstream.length}},
  \code{\link{adls.fileinputstream.marksupported}},
  \code{\link{adls.fileinputstream.mark}},
  \code{\link{adls.fileinputstream.readfromservice}},
  \code{\link{adls.fileinputstream.readfully}},
  \code{\link{adls.fileinputstream.read}},
  \code{\link{adls.fileinputstream.reset}},
  \code{\link{adls.fileinputstream.seek}},
  \code{\link{adls.fileinputstream.skip}},
  \code{\link{adls.fileoutputstream.close}},
  \code{\link{adls.fileoutputstream.create}},
  \code{\link{adls.fileoutputstream.flush}},
  \code{\link{adls.fileoutputstream.write}},
  \code{\link{adls.ls}}, \code{\link{adls.mkdir}},
  \code{\link{adls.read.core}},
  \code{\link{adls.read.direct}}, \code{\link{adls.read}},
  \code{\link{adls.rename}},
  \code{\link{createAdlExponentialBackoffRetryPolicy}},
  \code{\link{createAdlNonIdempotentRetryPolicy}},
  \code{\link{createAdlRetryPolicy}},
  \code{\link{shouldRetry.adlExponentialBackoffRetryPolicy}},
  \code{\link{shouldRetry.adlNonIdempotentRetryPolicy}},
  \code{\link{shouldRetry}}
}
