% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureDataLake.R
\name{adls.concat}
\alias{adls.concat}
\title{Concat existing files together. (MSCONCAT).}
\usage{
adls.concat(azureActiveContext, azureDataLakeAccount, relativePath,
  sourceRelativePaths, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{azureDataLakeAccount}{Name of the Azure Data Lake account.}

\item{relativePath}{Relative path of a destination file.}

\item{sourceRelativePaths}{Simple vector of relative paths of source files to be concatenated.}

\item{verbose}{Print tracing information (default FALSE).}
}
\value{
NULL (void)
}
\description{
Concat existing files together. (MSCONCAT).
}
\details{
Exceptions - IOException, UnsupportedOperationException
}
\references{
\url{https://docs.microsoft.com/en-us/azure/data-lake-store/data-lake-store-data-operations-rest-api}
}
\seealso{
\url{https://hadoop.apache.org/docs/current/hadoop-project-dist/hadoop-hdfs/WebHDFS.html#Concat_Files}

\url{https://hadoop.apache.org/docs/current/hadoop-project-dist/hadoop-hdfs/WebHDFS.html#Sources}

Other Azure Data Lake Store functions: \code{\link{adls.append.core}},
  \code{\link{adls.append.direct}},
  \code{\link{adls.append}}, \code{\link{adls.create}},
  \code{\link{adls.delete}}, \code{\link{adls.file.info}},
  \code{\link{adls.fileinputstream.available}},
  \code{\link{adls.fileinputstream.close}},
  \code{\link{adls.fileinputstream.create}},
  \code{\link{adls.fileinputstream.getpos}},
  \code{\link{adls.fileinputstream.length}},
  \code{\link{adls.fileinputstream.marksupported}},
  \code{\link{adls.fileinputstream.mark}},
  \code{\link{adls.fileinputstream.readfromservice}},
  \code{\link{adls.fileinputstream.readfully}},
  \code{\link{adls.fileinputstream.read}},
  \code{\link{adls.fileinputstream.reset}},
  \code{\link{adls.fileinputstream.seek}},
  \code{\link{adls.fileinputstream.skip}},
  \code{\link{adls.fileoutputstream.close}},
  \code{\link{adls.fileoutputstream.create}},
  \code{\link{adls.fileoutputstream.flush}},
  \code{\link{adls.fileoutputstream.write}},
  \code{\link{adls.ls}}, \code{\link{adls.mkdir}},
  \code{\link{adls.read.core}},
  \code{\link{adls.read.direct}}, \code{\link{adls.read}},
  \code{\link{adls.rename}},
  \code{\link{createAdlExponentialBackoffRetryPolicy}},
  \code{\link{createAdlNonIdempotentRetryPolicy}},
  \code{\link{createAdlRetryPolicy}},
  \code{\link{shouldRetry.adlExponentialBackoffRetryPolicy}},
  \code{\link{shouldRetry.adlNonIdempotentRetryPolicy}},
  \code{\link{shouldRetry}}
}
