% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureDataLake.R
\name{adls.fileinputstream.skip}
\alias{adls.fileinputstream.skip}
\title{Skip to given position in stream.}
\usage{
adls.fileinputstream.skip(adlFileInputStream, n)
}
\arguments{
\item{adlFileInputStream}{adlFileInputStream of the file}

\item{n}{position to seek to}
}
\value{
Skips over and discards n bytes of data from the input stream.
The skip method may, for a variety of reasons, end up skipping over some smaller
number of bytes, possibly 0. The actual number of bytes skipped is returned.
}
\description{
Skip to given position in stream.
}
\details{
Exceptions - IOException, EOFException
}
\seealso{
Other Azure Data Lake Store functions: \code{\link{adls.append.core}},
  \code{\link{adls.append.direct}},
  \code{\link{adls.append}}, \code{\link{adls.concat}},
  \code{\link{adls.create}}, \code{\link{adls.delete}},
  \code{\link{adls.file.info}},
  \code{\link{adls.fileinputstream.available}},
  \code{\link{adls.fileinputstream.close}},
  \code{\link{adls.fileinputstream.create}},
  \code{\link{adls.fileinputstream.getpos}},
  \code{\link{adls.fileinputstream.length}},
  \code{\link{adls.fileinputstream.marksupported}},
  \code{\link{adls.fileinputstream.mark}},
  \code{\link{adls.fileinputstream.readfromservice}},
  \code{\link{adls.fileinputstream.readfully}},
  \code{\link{adls.fileinputstream.read}},
  \code{\link{adls.fileinputstream.reset}},
  \code{\link{adls.fileinputstream.seek}},
  \code{\link{adls.fileoutputstream.close}},
  \code{\link{adls.fileoutputstream.create}},
  \code{\link{adls.fileoutputstream.flush}},
  \code{\link{adls.fileoutputstream.write}},
  \code{\link{adls.ls}}, \code{\link{adls.mkdir}},
  \code{\link{adls.read.core}},
  \code{\link{adls.read.direct}}, \code{\link{adls.read}},
  \code{\link{adls.rename}},
  \code{\link{createAdlExponentialBackoffRetryPolicy}},
  \code{\link{createAdlNonIdempotentRetryPolicy}},
  \code{\link{createAdlRetryPolicy}},
  \code{\link{shouldRetry.adlExponentialBackoffRetryPolicy}},
  \code{\link{shouldRetry.adlNonIdempotentRetryPolicy}},
  \code{\link{shouldRetry}}
}
