% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzureBlobCD}
\alias{AzureBlobCD}
\title{Azure Blob change current Directory.}
\usage{
AzureBlobCD(AzureActiveContext, Directory, Container, File, StorageAccount,
  StorageKey, ResourceGroup, SubscriptionID, AzToken, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{Directory}{Set Directory to list}

\item{Container}{Container Object}

\item{File}{- Local FileName to Store in Azure Blob}

\item{StorageAccount}{StorageAccount Object}

\item{StorageKey}{StorageKey Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Azure Blob change current Directory.
}
\seealso{
Other Blob store functions: \code{\link{AzureBlobFind}},
  \code{\link{AzureBlobLS}}, \code{\link{AzureGetBlob}},
  \code{\link{AzureListSABlobs}},
  \code{\link{AzurePutBlob}}
}

