% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzureSM: AzureBlobFind}
\alias{AzureBlobFind}
\alias{AzureSM: AzureBlobFind}
\title{Find File in a Storage account directory}
\usage{
AzureBlobFind(AzureActiveContext, File, StorageAccount, StorageKey, Container,
  ResourceGroup, SubscriptionID, AzToken, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{StorageAccount}{StorageAccount}

\item{StorageKey}{StorageKey}

\item{Container}{Storage Container}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{verbose}{Print Tracing information (Default False)}

\item{Token}{Token Object (or use AzureActiveContext)}
}

