% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azureblobLS}
\alias{azureblobLS}
\title{List blob files in a Storage account directory.}
\usage{
azureblobLS(azureActiveContext, directory, recursive = FALSE, storageAccount,
  storageKey, container, resourceGroup, subscriptionID, azToken,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{directory}{Set directory to list}

\item{recursive}{List directories recursively (Default FALSE)}

\item{storageAccount}{storageAccount Object}

\item{storageKey}{storageKey Object}

\item{container}{container Object}

\item{resourceGroup}{resourceGroup Object}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List blob files in a Storage account directory.
}
\seealso{
Other blob store functions: \code{\link{azureBlobCD}},
  \code{\link{azureBlobFind}}, \code{\link{azureGetBlob}},
  \code{\link{azureListStorageBlobs}},
  \code{\link{azurePutBlob}}
}

