% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{azureCancelDeploy}
\alias{azureCancelDeploy}
\title{Cancel Template Deployment.}
\usage{
azureCancelDeploy(azureActiveContext, deplname, resourceGroup, subscriptionID,
  azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{deplname}{deplname}

\item{resourceGroup}{resourceGroup Object}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Cancel Template Deployment.
}
\seealso{
Other Template functions: \code{\link{azureDeleteDeploy}},
  \code{\link{azureDeployStatus}},
  \code{\link{azureDeployTemplate}}
}

