% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureAuthenticate.R
\name{azureCheckToken}
\alias{azureCheckToken}
\title{Check the timestamp of a token and renew if needed.}
\usage{
azureCheckToken(azureActiveContext)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}
}
\description{
Check the timestamp of a token and renew if needed.
}
\seealso{
Other Azure resource functions: \code{\link{azureAuthenticate}}
}
