% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBatch.R
\name{azureCreateBatchAccount}
\alias{azureCreateBatchAccount}
\title{Create an azure batch account.}
\usage{
azureCreateBatchAccount(azureActiveContext, batchAccount,
  location = "northeurope", resourceGroup, subscriptionID,
  asynchronous = FALSE, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{location}{A string for the location to create batch account}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{asynchronous}{If TRUE, submits asynchronous request to Azure. Otherwise waits until batch account is created.}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Create an azure batch account.
}
\references{
https://docs.microsoft.com/en-us/rest/api/batchmanagement/batchaccount
}
\seealso{
Other Batch account functions: \code{\link{azureBatchGetKey}},
  \code{\link{azureDeleteBatchAccount}},
  \code{\link{azureListBatchAccounts}}
}
