% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{AzureSM:}
\alias{AzureCreateHDI}
\alias{AzureRunScriptAction}
\alias{AzureSM:}
\alias{AzureSM: AzureRunScriptAction}
\title{Create Specifed HDInsight Cluster}
\usage{
AzureCreateHDI(AzureActiveContext, ClusterName, Location, Kind = "spark",
  StorageAcc, SKey, Version = "3.4", Workers = 2, AdminUser, AdminPassword,
  SSHUser, SSHPassword, HiveServer, HiveDB, HiveUser, HivePassword,
  ResourceGroup, AzToken, SubscriptionID, Mode = "Sync", verbose = FALSE)

AzureRunScriptAction(AzureActiveContext, ScriptName = "script1", ScriptURL,
  HeadNode = TRUE, WorkerNode = FALSE, EdgeNode = FALSE, ClusterName,
  ResourceGroup, Parameters = "", AzToken, SubscriptionID, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{- Azure Context Object}

\item{ClusterName}{- ResourceGroup Object (or use AzureActiveContext)}

\item{Location}{- Location String}

\item{Kind}{- Kind (spark/hadoop) DEFAULT{spark}}

\item{StorageAcc}{- Storage Account Name}

\item{SKey}{- Storage Key}

\item{Workers}{- # of Workers (Default 2)}

\item{ResourceGroup}{- ResourceGroup Object (or use AzureActiveContext)}

\item{AzToken}{- Token Object (or use AzureActiveContext)}

\item{SubscriptionID}{- SubscriptionID Object (or use AzureActiveContext)}

\item{verbose}{- Print Tracing information (Default False)}

\item{ScriptName}{- Identifier for Custom action scrript operation}

\item{ScriptURL}{- URL to custom action script (Sring)}

\item{HeadNode}{- install on head nodes (default FALSE)}

\item{WorkerNode}{- install on worker nodes (default FALSE)}

\item{EdgeNode}{- install on worker nodes (default FALSE)}

\item{AzureActiveContext}{- Azure Context Object}

\item{Roles}{- Specificy the Roles to apply action string (workernode,headnode,edgenode)}

\item{ClusterName}{- ResourceGroup Object (or use AzureActiveContext)}

\item{ResourceGroup}{- ResourceGroup Object (or use AzureActiveContext)}

\item{AzToken}{- Token Object (or use AzureActiveContext)}

\item{SubscriptionID}{- SubscriptionID Object (or use AzureActiveContext)}

\item{verbose}{- Print Tracing information (Default False)}
}
\value{
Returns Success Message

Returns Success Message
}

