% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{AzureCreateResourceGroup}
\alias{AzureCreateResourceGroup}
\title{Create a ResourceGroup.}
\usage{
AzureCreateResourceGroup(AzureActiveContext, ResourceGroup, Location, SUBID, AT,
  verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ResourceGroup}{ResourceGroup Object}
}
\value{
Returns Dataframe of Resources
}
\description{
Create a ResourceGroup.
}
\seealso{
Other Resources: \code{\link{AzureAuthenticate}},
  \code{\link{AzureCheckToken}},
  \code{\link{AzureDeleteResourceGroup}},
  \code{\link{AzureListAllRecources}},
  \code{\link{AzureListRG}},
  \code{\link{AzureListSubscriptions}}

Other Resources: \code{\link{AzureAuthenticate}},
  \code{\link{AzureCheckToken}},
  \code{\link{AzureDeleteResourceGroup}},
  \code{\link{AzureListAllRecources}},
  \code{\link{AzureListRG}},
  \code{\link{AzureListSubscriptions}}
}

