% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureStorageAccount.R
\name{azureCreateStorageAccount}
\alias{azureCreateStorageAccount}
\title{Create an Azure Storage Account.}
\usage{
azureCreateStorageAccount(azureActiveContext, storageAccount, azToken,
  resourceGroup, subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{storageAccount}{Set the name of the azure storage account}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{resourceGroup}{Set the name of the Resource Group}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Create an Azure Storage Account.
}
\seealso{
Other Storage account functions: \code{\link{azureDeletestorageAccount}},
  \code{\link{azureListSA}}, \code{\link{azureSAGetKey}}
}

