% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContainer.R
\name{azureCreateStorageContainer}
\alias{azureCreateStorageContainer}
\title{Create Storage containers in a specified Storage Account.}
\usage{
azureCreateStorageContainer(azureActiveContext, container, storageAccount,
  storageKey, resourceGroup, azToken, subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{container}{Storage container name. See \code{\link{azureListStorageContainers}}}

\item{storageAccount}{Name of the azure storage account}

\item{storageKey}{Storage key associated with storage account}

\item{resourceGroup}{Name of the resource group}

\item{azToken}{Azure authentication token, obtained by \code{\link{azureAuthenticate}}}

\item{subscriptionID}{Set the subscriptionID.  This is obtained automatically by \code{\link{azureAuthenticate}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Create Storage containers in a specified Storage Account.
}
\seealso{
Other container functions: \code{\link{azureDeleteStorageContainer}},
  \code{\link{azureListStorageContainers}}
}

