% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureDataLake.R
\name{azureDataLakeDelete}
\alias{azureDataLakeDelete}
\title{Azure Data Lake DELETE for specified relativePath of an azure data lake account.}
\usage{
azureDataLakeDelete(azureActiveContext, azureDataLakeAccount, relativePath,
  recursive = FALSE, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Provide an \code{azureActiveContext} object used for authentication.}

\item{azureDataLakeAccount}{Provide the name of the Azure Data Lake account.}

\item{relativePath}{Provide a relative path of the directory.}

\item{recursive}{Provide recursive delete option.}

\item{verbose}{Print tracing information (default FALSE).}
}
\value{
Returns a boolean.
}
\description{
Azure Data Lake DELETE for specified relativePath of an azure data lake account.
}
\seealso{
Other Azure Data Lake Store functions: \code{\link{azureDataLakeAppend}},
  \code{\link{azureDataLakeCreate}},
  \code{\link{azureDataLakeGetFileStatus}},
  \code{\link{azureDataLakeListStatus}},
  \code{\link{azureDataLakeMkdirs}},
  \code{\link{azureDataLakeOpen}}
}
