% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureDataLake.R
\name{azureDataLakeListStatus}
\alias{azureDataLakeListStatus}
\title{Azure Data Lake LISTSTATUS for specified storage account.}
\usage{
azureDataLakeListStatus(azureActiveContext, storageAccount, relativePath = "",
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Provide an \code{azureActiveContext} object used for authentication.}

\item{verbose}{Print Tracing information (Default False)}

\item{azureDataLakeStoreAccount}{Provide the name of the Azure Data Lake Store account.}
}
\value{
Returns a data frame. This data frame has an attribute called \code{marker} that can be used with the \code{marker} argument to return the next set of values.
}
\description{
Azure Data Lake LISTSTATUS for specified storage account.
}
