% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureDataLake.R
\name{azureDataLakeOpen}
\alias{azureDataLakeOpen}
\title{Azure Data Lake OPEN for specified relativePath of an azure data lake account.}
\usage{
azureDataLakeOpen(azureActiveContext, azureDataLakeAccount, relativePath,
  offset, length, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Provide an \code{azureActiveContext} object used for authentication.}

\item{azureDataLakeAccount}{Provide the name of the Azure Data Lake account.}

\item{relativePath}{Provide a relative path of the directory.}

\item{offset}{Provide the offset to read from.}

\item{length}{Provide length of data to read.}

\item{verbose}{Print tracing information (default FALSE).}
}
\value{
Returns a data frame.
}
\description{
Azure Data Lake OPEN for specified relativePath of an azure data lake account.
}
\seealso{
Other Azure Data Lake Store functions: \code{\link{azureDataLakeAppend}},
  \code{\link{azureDataLakeCreate}},
  \code{\link{azureDataLakeDelete}},
  \code{\link{azureDataLakeGetFileStatus}},
  \code{\link{azureDataLakeListStatus}},
  \code{\link{azureDataLakeMkdirs}}
}
