% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBatch.R
\name{azureDeleteBatchAccount}
\alias{azureDeleteBatchAccount}
\title{Delete an Azure Batch Account.}
\usage{
azureDeleteBatchAccount(azureActiveContext, batchAccount, resourceGroup,
  subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Delete an Azure Batch Account.
}
\seealso{
Other Batch account functions: \code{\link{azureBatchGetKey}},
  \code{\link{azureCreateBatchAccount}}
}
