% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{AzureDeleteResourceGroup: AzureListAllRecources}
\alias{AzureDeleteResourceGroup}
\alias{AzureDeleteResourceGroup: AzureListAllRecources}
\title{Delete a ResourceGroup with all Resources}
\usage{
AzureDeleteResourceGroup(AzureActiveContext, ResourceGroup, SUBID, AT, Type,
  verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{Token}{Token Object (or use AzureActiveContext)}

\item{Verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of Resources
}

