% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{AzureSM: AzureDeployTemplate}
\alias{AzureDeployTemplate}
\alias{AzureSM: AzureDeployTemplate}
\title{Submit a Azure Resource ManagerTemplate}
\usage{
AzureDeployTemplate(DeplName, TemplateURL, ParamURL, ParamJSON, Mode = "Sync",
  ResourceGroup, SubscriptionID, AzToken, verbose = FALSE)
}
\arguments{
\item{DeplName}{DeplName}

\item{TemplateURL}{TemplateURL}

\item{ParamURL}{ParamURL}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{AzureActiveContext}{Azure Context Object}

\item{Token}{Token Object (or use AzureActiveContext)}

\item{Verbose}{Print Tracing information (Default False)}
}

