% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureGetAllVMstatus}
\alias{azureGetAllVMstatus}
\title{List the status of every virtual machine in the azure active context.}
\usage{
azureGetAllVMstatus(azureActiveContext)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}
}
\description{
First queries the azure active context for all visible resources, then sequentially queries the status of all virtuam machines.
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureListScaleSetNetwork}},
  \code{\link{azureListScaleSetVM}},
  \code{\link{azureListScaleSets}},
  \code{\link{azureListVM}}, \code{\link{azureStartVM}},
  \code{\link{azureStopVM}}, \code{\link{azureVMInfo}},
  \code{\link{azureVMStatus}}
}
