% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBatch.R
\name{azureGetBatchAccount}
\alias{azureGetBatchAccount}
\title{Get the specified Batch Account.}
\usage{
azureGetBatchAccount(azureActiveContext, batchAccount, resourceGroup,
  subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{If TRUE, prints verbose messages}
}
\description{
Get the specified Batch Account.
}
\seealso{
Other Batch account functions: \code{\link{azureBatchGetKey}},
  \code{\link{azureCreateBatchAccount}},
  \code{\link{azureDeleteBatchAccount}},
  \code{\link{azureListBatchAccounts}}
}
