% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureStorageAccount.R
\name{azureGetStorageAccount}
\alias{azureGetStorageAccount}
\title{Get specified storage account}
\usage{
azureGetStorageAccount(azureActiveContext, resourceGroup, subscriptionID,
  storageAccount, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{storageAccount}{Name of the azure storage account. Storage account names must be between 3 and 24 characters in length and use numbers and lower-case letters only.}

\item{verbose}{If TRUE, prints verbose messages}
}
\description{
Get specified storage account
}
\seealso{
Other Storage account functions: \code{\link{azureCreateStorageAccount}},
  \code{\link{azureDeletestorageAccount}},
  \code{\link{azureListSA}}, \code{\link{azureSAGetKey}}
}
