% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureAuthenticate.R
\name{azureGetTokenRefreshToken}
\alias{azureGetTokenRefreshToken}
\title{Get Azure token using RefreshToken}
\usage{
azureGetTokenRefreshToken(azureActiveContext, tenantID, refreshToken,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{tenantID}{The tenant ID provided during creation of the Active Directory application / service principal}

\item{refreshToken}{Provide the previously obtained refreshToken}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
If successful, returns TRUE
}
\description{
Get Azure token using RefreshToken
}
\note{
See \url{https://azure.microsoft.com/en-us/documentation/articles/resource-group-create-service-principal-portal/} for instructions to set up an Active Directory application
}
\references{
\url{https://azure.microsoft.com/en-us/documentation/articles/resource-group-create-service-principal-portal/}
}
\seealso{
Other Azure resource functions: \code{\link{azureAuthenticateOnAuthType}},
  \code{\link{azureAuthenticate}},
  \code{\link{azureCheckToken}},
  \code{\link{azureGetTokenClientCredential}},
  \code{\link{azureGetTokenDeviceCodeFetch}},
  \code{\link{azureGetTokenDeviceCode}},
  \code{\link{showDeviceCodeMessageToUser}}
}
