% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{azureHDIConf}
\alias{azureHDIConf}
\title{Get Configuration Information for a specified clustername.}
\usage{
azureHDIConf(azureActiveContext, clustername, resourceGroup, subscriptionID,
  azToken, name, type, location, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{clustername}{clustername Object}

\item{resourceGroup}{resourceGroup Object}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{name}{name}

\item{type}{type}

\item{location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of HDInsight Clusters information
}
\description{
Get Configuration Information for a specified clustername.
}
\seealso{
Other HDInsight functions: \code{\link{azureCreateHDI}},
  \code{\link{azureDeleteHDI}}, \code{\link{azureListHDI}},
  \code{\link{azureResizeHDI}},
  \code{\link{azureRunScriptAction}},
  \code{\link{azureScriptActionHistory}}
}

