% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHive.R
\name{AzureHiveSQL}
\alias{AzureHiveSQL}
\title{Submit SQL command to Hive Service.}
\usage{
AzureHiveSQL(AzureActiveContext, CMD, ClusterName, HDIAdmin, HDIPassword,
  Path = "wasb:///tmp/", verbose = FALSE)
}
\description{
Submit SQL command to Hive Service.
}

