% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHive.R
\name{AzureHiveStatus}
\alias{AzureHiveStatus}
\title{Get Status of a HDI Hive Service/Version.}
\usage{
AzureHiveStatus(AzureActiveContext, ClusterName, HDIAdmin, HDIPassword,
  verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ClusterName}{ClusterName Object}

\item{HDIAdmin}{HDIAdmin Object}

\item{HDIPassword}{HDIPassword Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Get Status of a HDI Hive Service/Version.
}
\seealso{
Other Hive functions: \code{\link{AzureHiveSQL}}
}

