% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{azureListHDI}
\alias{azureListHDI}
\title{Get all HDInsight Clusters in default Subscription or details for a specified clustername.}
\usage{
azureListHDI(azureActiveContext, resourceGroup, clustername = "*",
  subscriptionID, azToken, name, type, location, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{resourceGroup}{resourceGroup Object}

\item{clustername}{Cluster name}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{name}{name}

\item{type}{type}

\item{location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of HDInsight Clusters
}
\description{
Get all HDInsight Clusters in default Subscription or details for a specified clustername.
}
\seealso{
Other HDInsight functions: \code{\link{azureCreateHDI}},
  \code{\link{azureDeleteHDI}}, \code{\link{azureHDIConf}},
  \code{\link{azureResizeHDI}},
  \code{\link{azureRunScriptAction}},
  \code{\link{azureScriptActionHistory}}
}

