% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzureListSAContainers}
\alias{AzureListSAContainers}
\title{List Storage Containers for Specified Storage Account.}
\usage{
AzureListSAContainers(AzureActiveContext, StorageAccount, StorageKey,
  ResourceGroup, AzToken, SubscriptionID, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{StorageAccount}{StorageAccount Object}

\item{StorageKey}{StorageKey Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{AzToken}{Token Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object}
}
\description{
List Storage Containers for Specified Storage Account.
}
\seealso{
Other Storage account: \code{\link{AzureBlobCD}},
  \code{\link{AzureBlobFind}}, \code{\link{AzureBlobLS}},
  \code{\link{AzureCreateSAContainer}},
  \code{\link{AzureCreateStorageAccount}},
  \code{\link{AzureDeleteSAContainer}},
  \code{\link{AzureDeleteStorageAccount}},
  \code{\link{AzureGetBlob}},
  \code{\link{AzureListSABlobs}},
  \code{\link{AzureListSA}}, \code{\link{AzurePutBlob}},
  \code{\link{AzureSAGetKey}}
}

