% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureListScaleSetNetwork}
\alias{azureListScaleSetNetwork}
\title{List scale set network information with a ResourceGroup}
\usage{
azureListScaleSetNetwork(azureActiveContext, resourceGroup, location,
  subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{resourceGroup}{Name of the resource group}

\item{location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{subscriptionID}{Set the subscriptionID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List scale set network information with a ResourceGroup
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureListScaleSetVM}},
  \code{\link{azureListScaleSets}},
  \code{\link{azureListVM}}, \code{\link{azureStartVM}},
  \code{\link{azureStopVM}}, \code{\link{azureVMStatus}}
}
