% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azureScalesets.R
\name{azureListScaleSets}
\alias{azureListScaleSets}
\title{List scale sets within a resource group.}
\usage{
azureListScaleSets(azureActiveContext, resourceGroup, location, subscriptionID,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{resourceGroup}{Name of the resource group}

\item{location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List scale sets within a resource group.
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureGetAllVMstatus}},
  \code{\link{azureListScaleSetNetwork}},
  \code{\link{azureListScaleSetVM}},
  \code{\link{azureListVM}}, \code{\link{azureStartVM}},
  \code{\link{azureStopVM}}, \code{\link{azureVMStatus}}
}
