% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azureScalesets.R
\name{azureListScaleSetVM}
\alias{azureListScaleSetVM}
\title{List VMs within a scale set}
\usage{
azureListScaleSetVM(azureActiveContext, scaleSet, resourceGroup, location,
  subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}}

\item{scaleSet}{Character vector with name of the scaleset (see also \code{\link[=azureListScaleSets]{azureListScaleSets()}})}

\item{resourceGroup}{Name of the resource group}

\item{location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List VMs within a scale set
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureGetAllVMstatus}},
  \code{\link{azureListScaleSetNetwork}},
  \code{\link{azureListScaleSets}},
  \code{\link{azureListVM}}, \code{\link{azureStartVM}},
  \code{\link{azureStopVM}}, \code{\link{azureVMStatus}}
}
