% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azureListStorageBlobs}
\alias{azureListStorageBlobs}
\title{List Storage containers for Specified Storage Account.}
\usage{
azureListStorageBlobs(azureActiveContext, storageAccount, storageKey, container,
  resourceGroup, subscriptionID, azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{storageAccount}{Set the name of the azure storage account}

\item{storageKey}{Set the Storage Key associated with storage account}

\item{container}{Set the storage container}

\item{resourceGroup}{Set the name of the Resource Group}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List Storage containers for Specified Storage Account.
}
\seealso{
Other blob store functions: \code{\link{azureBlobCD}},
  \code{\link{azureBlobFind}}, \code{\link{azureGetBlob}},
  \code{\link{azurePutBlob}}, \code{\link{azureblobLS}}
}

