% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{AzureSM: AzureListSubscriptions}
\alias{AzureListSubscriptions}
\alias{AzureSM: AzureListSubscriptions}
\title{Get available Subscriptions}
\usage{
AzureListSubscriptions(AzureActiveContext, ATI, verbose = FALSE)
}
\arguments{
\item{Azure}{Context Object}

\item{Token}{Token Object (or use AzureActiveContext)}

\item{Verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of SubscriptionID sets AzureContext SubscriptionID
}
\description{
Get available Subscriptions
}

