% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{AzureSM: AzureCheckToken}
\alias{AzureCheckToken}
\alias{AzureSM: AzureCheckToken}
\title{Check the timestamp of a Token and Renew if needed.}
\usage{
AzureCheckToken(AzureActiveContext)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}
}

