% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{AzureSM: AzureDeleteDeploy}
\alias{AzureDeleteDeploy}
\alias{AzureSM: AzureDeleteDeploy}
\title{Delete Template Deployment}
\usage{
AzureDeleteDeploy(SubscriptionID, AzToken, ResourceGroup, DeplName,
  verbose = FALSE)
}
\arguments{
\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{DeplName}{DeplName}

\item{AzureActiveContext}{Azure Context Object}

\item{Token}{Token Object (or use AzureActiveContext)}

\item{Verbose}{Print Tracing information (Default False)}
}

