% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSM-package.R
\docType{package}
\name{AzureSM-package}
\alias{AzureSM}
\alias{AzureSM-package}
\title{AzureSM}
\description{
The AzureSM package connects R to the Azure Service Manager API.
}
\details{
This enables you to use and change many Azure resources, including:

\itemize{
\item Storage Blobs
\item HDInsight (Nodes, Hive, Spark)
\item Azure Resource Manager
\item Virtual Machines
}
}
\keyword{package}

