% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSMR-package.R
\docType{package}
\name{AzureSMR-package}
\alias{AzureSMR}
\alias{AzureSMR-package}
\title{Install packages from snapshots on the checkpoint server for reproducibility}
\description{
THe AzureSMR package connects R to the Azure Service Manager API.
}
\details{
This enables you to use and change many Azure resources, including:

\itemize{
\item Storage Blobs
\item HDInsight (Nodes, Hive, Spark)
\item Azure Resource Manager
\item Virtual Machines
}
}
\keyword{package}

