% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSMR-package.R
\docType{package}
\name{AzureSMR}
\alias{AzureSMR}
\alias{AzureSMR-package}
\title{AzureSMR}
\description{
The AzureSMR package connects R to the Azure Service Manager API.
}
\details{
This enables you to use and change many Azure resources. The following is an incomplete list of available functions:

* Create an Azure context:
  - [createAzureContext()]
* Resources and resource groups:
  - [azureListAllResources()]
  - [azureCreateResourceGroup()]
  - [azureListRG()]
* Storage accounts and blobs:
  - [azureListSA()]
  - [azureCreateStorageAccount()]
  - [azureListStorageContainers()]
  - [azureListStorageBlobs()]
  - [azureGetBlob()]
  - [azurePutBlob()]
  - [azureDeleteBlob()]
  - [azureBlobFind()]
* Virtual Machines
  - List VMs: [azureListVM()]
  - To create a virtual machine, use [azureDeployTemplate()] with a suitable template
  - Start a VM: [azureStartVM()]
  - Stop a VM: [azureStopVM()]
  - Get status: [azureVMStatus()]
* HDInsight clusters:
  - [azureListHDI()]
  - [azureCreateHDI()]
  - [azureResizeHDI()]
  - [azureDeleteHDI()]
  - [azureRunScriptAction()]
  - [azureScriptActionHistory()]
* Azure batch:
  - [azureListBatchAccounts()]
  - [azureCreateBatchAccount()]
  - [azureDeleteBatchAccount()]
  - [azureBatchGetKey()]
* Cost functions:
  - [azureDataConsumption()]
  - [azurePricingRates()]
  - [azureExpenseCalculator()]
}
\keyword{package}

