% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{AzureSparkJob}
\alias{AzureSparkJob}
\title{Submit Spark Job (Batch mode).}
\usage{
AzureSparkJob(AzureActiveContext, FILE, ClusterName, HDIAdmin, HDIPassword,
  Log = "URL", verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ClusterName}{ClusterName Object}

\item{HDIAdmin}{HDIAdmin Object}

\item{HDIPassword}{HDIPassword Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Submit Spark Job (Batch mode).
}
\seealso{
Other Spark: \code{\link{AzureSparkCMD}},
  \code{\link{AzureSparkListJobs}},
  \code{\link{AzureSparkListSessions}},
  \code{\link{AzureSparkNewSession}},
  \code{\link{AzureSparkShowURL}},
  \code{\link{AzureSparkStopSession}}
}

