% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{azureSparkListSessions}
\alias{azureSparkListSessions}
\title{List Spark Sessions.}
\usage{
azureSparkListSessions(azureActiveContext, clustername, hdiAdmin, hdiPassword,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{clustername}{clustername Object}

\item{hdiAdmin}{hdiAdmin Object}

\item{hdiPassword}{hdiPassword  Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List Spark Sessions.
}
\seealso{
Other Spark: \code{\link{azureSparkCMD}},
  \code{\link{azureSparkJob}},
  \code{\link{azureSparkListJobs}},
  \code{\link{azureSparkNewSession}},
  \code{\link{azureSparkShowURL}},
  \code{\link{azureSparkStopSession}}
}

