% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{AzureSM: AzureSparkNewSession}
\alias{AzureSM: AzureSparkNewSession}
\alias{AzureSparkNewSession}
\title{Create new Spark Session}
\usage{
AzureSparkNewSession(AzureActiveContext, ClusterName, HDIAdmin, HDIPassword,
  Kind = "spark", verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{ClusterName}{ClusterName}

\item{HDIAdmin}{HDIAdmin - HDinsight Administrator Name}

\item{HDIPassword}{HDinsight Administrator Name}

\item{Kind}{Kind Spark/PySpark/Spark}

\item{Verbose}{Print Tracing information (Default False)}
}

