% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{azureSparkShowURL}
\alias{azureSparkShowURL}
\title{Show Spark log Output.}
\usage{
azureSparkShowURL(azureActiveContext, URL)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{URL}{URL}
}
\value{
Show log in Browser
}
\description{
Show Spark log Output.
}
\seealso{
Other Spark: \code{\link{azureSparkCMD}},
  \code{\link{azureSparkJob}},
  \code{\link{azureSparkListJobs}},
  \code{\link{azureSparkListSessions}},
  \code{\link{azureSparkNewSession}},
  \code{\link{azureSparkStopSession}}
}

