% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{AzureSM: AzureSparkStopSession}
\alias{AzureSM: AzureSparkStopSession}
\alias{AzureSparkStopSession}
\title{Stop a Spark Sessions}
\usage{
AzureSparkStopSession(AzureActiveContext, ClusterName, HDIAdmin, HDIPassword,
  SessionID, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{ClusterName}{ClusterName}

\item{HDIAdmin}{HDIAdmin - HDinsight Administrator Name}

\item{HDIPassword}{HDinsight Administrator Name}

\item{SessionID}{SessionID see AzureSparkListSessions)}

\item{Verbose}{Print Tracing information (Default False)}
}

