% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureStartVM}
\alias{azureStartVM}
\title{Start a Virtual Machine.}
\usage{
azureStartVM(azureActiveContext, resourceGroup, vmName, mode = "Sync",
  subscriptionID, azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{resourceGroup}{resourceGroup Object}

\item{vmName}{Virtual Machine name}

\item{mode}{Wait for operation to complete 'Sync' (Default)}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Start a Virtual Machine.
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureListVM}}, \code{\link{azureStopVM}},
  \code{\link{azureVMStatus}}
}

