% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureStopVM}
\alias{azureStopVM}
\title{Stop a Virtual Machine.}
\usage{
azureStopVM(azureActiveContext, resourceGroup, vmName, mode = "Sync",
  subscriptionID, azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{resourceGroup}{resourceGroup Object}

\item{vmName}{vmName Object}

\item{mode}{Wait for operation to complete 'Sync' (Default)}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Stop a Virtual Machine.
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureListVM}}, \code{\link{azureStartVM}},
  \code{\link{azureVMStatus}}
}

