% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureVMStatus}
\alias{azureVMStatus}
\title{Get Status of a Virtual Machine.}
\usage{
azureVMStatus(azureActiveContext, resourceGroup, vmName, subscriptionID,
  azToken, ignore = "N", verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{resourceGroup}{Set the name of the Resource Group}

\item{vmName}{Set the name of the virtual Machine}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{ignore}{ignore}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Get Status of a Virtual Machine.
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureListVM}}, \code{\link{azureStartVM}},
  \code{\link{azureStopVM}}
}

