% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureDataLake.R
\name{createAdlFileOutputStream}
\alias{createAdlFileOutputStream}
\title{Create an adlFileOutputStream.
Create a container (\code{adlFileOutputStream}) for holding variables used by the Azure Data Lake Store data functions.}
\usage{
createAdlFileOutputStream(azureActiveContext, accountName, relativePath,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{accountName}{the account name}

\item{relativePath}{Relative path of a file/directory}

\item{verbose}{Print tracing information (default FALSE).}
}
\value{
An \code{adlFileOutputStream} object
}
\description{
Create an adlFileOutputStream.
Create a container (\code{adlFileOutputStream}) for holding variables used by the Azure Data Lake Store data functions.
}
\seealso{
Other Azure Data Lake Store functions: \code{\link{adlFileInputStreamAvailable}},
  \code{\link{adlFileInputStreamClose}},
  \code{\link{adlFileInputStreamGetPos}},
  \code{\link{adlFileInputStreamLength}},
  \code{\link{adlFileInputStreamMarkSupported}},
  \code{\link{adlFileInputStreamMark}},
  \code{\link{adlFileInputStreamReadBuffered}},
  \code{\link{adlFileInputStreamRead}},
  \code{\link{adlFileInputStreamReset}},
  \code{\link{adlFileInputStreamSeek}},
  \code{\link{adlFileInputStreamSkip}},
  \code{\link{adlFileOutputStreamClose}},
  \code{\link{adlFileOutputStreamFlush}},
  \code{\link{adlFileOutputStreamWrite}},
  \code{\link{azureDataLakeAppendBOS}},
  \code{\link{azureDataLakeAppendCore}},
  \code{\link{azureDataLakeAppend}},
  \code{\link{azureDataLakeCreate}},
  \code{\link{azureDataLakeDelete}},
  \code{\link{azureDataLakeGetFileStatus}},
  \code{\link{azureDataLakeListStatus}},
  \code{\link{azureDataLakeMkdirs}},
  \code{\link{azureDataLakeOpenBIS}},
  \code{\link{azureDataLakeReadCore}},
  \code{\link{azureDataLakeRead}},
  \code{\link{createAdlFileInputStream}},
  \code{\link{readFromService}}
}
