% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{createAdlRetryPolicy}
\alias{createAdlRetryPolicy}
\title{NOTE: Folowing points on ADLS AdlsRetryPolicy:
\enumerate{
\item Not implemented speculative reads hence not implemented \code{NoRetryPolicy}.
\item Not implemented ExponentialBackoffPolicyforMSI as its not used even in the JDK.
Create adlRetryPolicy.
Create a adlRetryPolicy (\code{adlRetryPolicy}) for holding variables used by the Azure Data Lake Store data functions.
}}
\usage{
createAdlRetryPolicy(azureActiveContext,
  retryPolicyType = retryPolicyEnum$EXPONENTIALBACKOFF, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{retryPolicyType}{the type of retryPlociy object to create.}

\item{verbose}{Print tracing information (default FALSE).}
}
\value{
An \code{adlRetryPolicy} object
}
\description{
NOTE: Folowing points on ADLS AdlsRetryPolicy:
\enumerate{
\item Not implemented speculative reads hence not implemented \code{NoRetryPolicy}.
\item Not implemented ExponentialBackoffPolicyforMSI as its not used even in the JDK.
Create adlRetryPolicy.
Create a adlRetryPolicy (\code{adlRetryPolicy}) for holding variables used by the Azure Data Lake Store data functions.
}
}
\references{
\url{https://github.com/Azure/azure-data-lake-store-java/blob/master/src/main/java/com/microsoft/azure/datalake/store/retrypolicies/RetryPolicy.java}
}
\seealso{
Other Azure Data Lake Store functions: \code{\link{adls.append.core}},
  \code{\link{adls.append.direct}},
  \code{\link{adls.append}}, \code{\link{adls.concat}},
  \code{\link{adls.create}}, \code{\link{adls.delete}},
  \code{\link{adls.file.info}},
  \code{\link{adls.fileinputstream.available}},
  \code{\link{adls.fileinputstream.close}},
  \code{\link{adls.fileinputstream.create}},
  \code{\link{adls.fileinputstream.getpos}},
  \code{\link{adls.fileinputstream.length}},
  \code{\link{adls.fileinputstream.marksupported}},
  \code{\link{adls.fileinputstream.mark}},
  \code{\link{adls.fileinputstream.readfromservice}},
  \code{\link{adls.fileinputstream.readfully}},
  \code{\link{adls.fileinputstream.read}},
  \code{\link{adls.fileinputstream.reset}},
  \code{\link{adls.fileinputstream.seek}},
  \code{\link{adls.fileinputstream.skip}},
  \code{\link{adls.fileoutputstream.close}},
  \code{\link{adls.fileoutputstream.create}},
  \code{\link{adls.fileoutputstream.flush}},
  \code{\link{adls.fileoutputstream.write}},
  \code{\link{adls.ls}}, \code{\link{adls.mkdir}},
  \code{\link{adls.read.core}},
  \code{\link{adls.read.direct}}, \code{\link{adls.read}},
  \code{\link{adls.rename}},
  \code{\link{createAdlExponentialBackoffRetryPolicy}},
  \code{\link{createAdlNonIdempotentRetryPolicy}},
  \code{\link{shouldRetry.adlExponentialBackoffRetryPolicy}},
  \code{\link{shouldRetry.adlNonIdempotentRetryPolicy}},
  \code{\link{shouldRetry}}
}
