% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContextObject.R
\name{createAzureContext}
\alias{createAzureContext}
\title{Create an AzureContext.}
\usage{
createAzureContext()
}
\description{
Create an AzureContext.
}
\seealso{
Other Context Object: \code{\link{dumpAzureContext}}
}

