% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{dataTableViewer}
\alias{dataTableViewer}
\title{This is a template to view a DT object}
\usage{
dataTableViewer(x, title = "")
}
\arguments{
\item{x}{data.frame}

\item{title}{Text to display in pane title}
}
\description{
This is a template to view a DT object
}
