% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContextObject.R
\name{DumpAzureContext}
\alias{DumpAzureContext}
\title{Dumps the contents of the AzureContext.}
\usage{
DumpAzureContext(AzureActiveContext)
}
\arguments{
\item{AzureActiveContext}{Context Object}
}
\description{
Dumps the contents of the AzureContext.
}
\seealso{
Other Context Object: \code{\link{CreateAzureContext}}
}

