% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{dumpAzureContext}
\alias{AzureListRG}
\alias{dumpAzureContext}
\title{Dumps out the contents of the AzureContext.}
\usage{
dumpAzureContext(azureActiveContext)

AzureListRG(...)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by `AzureSMR`, created by [createAzureContext()]}

\item{...}{passed to [azureListRG()]}
}
\description{
Dumps out the contents of the AzureContext.

Deprecated functions.
}

