% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read.AzureSMR.config}
\alias{read.AzureSMR.config}
\title{Reads settings from configuration file in JSON format.}
\usage{
read.AzureSMR.config(configFile = getOption("AzureSMR.config"))
}
\arguments{
\item{configFile}{location of file that contains configuration in JSON format. By default, this location is at \code{getOption("AzureSMR.config")}}
}
\description{
Reads settings from configuration file in JSON format.
}
