% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContextObject.R
\name{setAzureContext}
\alias{setAzureContext}
\title{Updates the value of an AzureContext variable.}
\usage{
setAzureContext(azureActiveContext, TID, CID, KEY, azToken, subscriptionID,
  resourceGroup, storageKey, storageAccount, container, blob, vmName, hdiAdmin,
  hdiPassword, clustername, kind, sessionID)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{TID}{Tenant ID Object}

\item{CID}{Client ID Object}

\item{KEY}{Authentication KEY Object}

\item{azToken}{Token Object}

\item{subscriptionID}{subscriptionID Object}

\item{resourceGroup}{resourceGroup Object}

\item{storageKey}{storageKey Object}

\item{storageAccount}{storageAccount Object}

\item{container}{container Object}

\item{blob}{blob Object}

\item{vmName}{vmName Object}

\item{hdiAdmin}{hdiAdmin Object}

\item{hdiPassword}{hdiPassword  Object}

\item{clustername}{clustername Object}

\item{kind}{"hadoop","spark" or "pyspark"}

\item{sessionID}{sessionID Object}
}
\description{
Updates the value of an AzureContext variable.
}

