% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{shouldRetry.adlExponentialBackoffRetryPolicy}
\alias{shouldRetry.adlExponentialBackoffRetryPolicy}
\title{Check if retry should be done based on \code{adlRetryPolicy} (adlExponentialBackoffRetryPolicy).}
\usage{
shouldRetry.adlExponentialBackoffRetryPolicy(adlRetryPolicy, httpResponseCode,
  lastException, verbose = FALSE)
}
\arguments{
\item{adlRetryPolicy}{the policy object to chek for retry}

\item{httpResponseCode}{the account name}

\item{lastException}{exception that was reported with failure}

\item{verbose}{Print tracing information (default FALSE)}
}
\value{
TRUE for retry and FALSE otherwise
}
\description{
Check if retry should be done based on \code{adlRetryPolicy} (adlExponentialBackoffRetryPolicy).
}
\seealso{
Other Azure Data Lake Store functions: \code{\link{adls.append.core}},
  \code{\link{adls.append.direct}},
  \code{\link{adls.append}}, \code{\link{adls.concat}},
  \code{\link{adls.create}}, \code{\link{adls.delete}},
  \code{\link{adls.file.info}},
  \code{\link{adls.fileinputstream.available}},
  \code{\link{adls.fileinputstream.close}},
  \code{\link{adls.fileinputstream.create}},
  \code{\link{adls.fileinputstream.getpos}},
  \code{\link{adls.fileinputstream.length}},
  \code{\link{adls.fileinputstream.marksupported}},
  \code{\link{adls.fileinputstream.mark}},
  \code{\link{adls.fileinputstream.readfromservice}},
  \code{\link{adls.fileinputstream.readfully}},
  \code{\link{adls.fileinputstream.read}},
  \code{\link{adls.fileinputstream.reset}},
  \code{\link{adls.fileinputstream.seek}},
  \code{\link{adls.fileinputstream.skip}},
  \code{\link{adls.fileoutputstream.close}},
  \code{\link{adls.fileoutputstream.create}},
  \code{\link{adls.fileoutputstream.flush}},
  \code{\link{adls.fileoutputstream.write}},
  \code{\link{adls.ls}}, \code{\link{adls.mkdir}},
  \code{\link{adls.read.core}},
  \code{\link{adls.read.direct}}, \code{\link{adls.read}},
  \code{\link{adls.rename}},
  \code{\link{createAdlExponentialBackoffRetryPolicy}},
  \code{\link{createAdlNonIdempotentRetryPolicy}},
  \code{\link{createAdlRetryPolicy}},
  \code{\link{shouldRetry.adlNonIdempotentRetryPolicy}},
  \code{\link{shouldRetry}}
}
