% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureAuthenticate.R
\name{showDeviceCodeMessageToUser}
\alias{showDeviceCodeMessageToUser}
\title{Display device code flow message to user.}
\usage{
showDeviceCodeMessageToUser(jsonResponseObject)
}
\arguments{
\item{jsonResponseObject}{JSON object that contains the message to be displayed.}
}
\value{
If successful, returns TRUE
}
\description{
Display device code flow message to user.
}
\seealso{
Other Azure resource functions: \code{\link{adlFileInputStreamCheck}},
  \code{\link{adlFileOutputStreamCheck}},
  \code{\link{azureAuthenticateOnAuthType}},
  \code{\link{azureAuthenticate}},
  \code{\link{azureCheckToken}},
  \code{\link{azureGetTokenClientCredential}},
  \code{\link{azureGetTokenDeviceCodeFetch}},
  \code{\link{azureGetTokenDeviceCode}},
  \code{\link{azureGetTokenRefreshToken}}
}
