% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureDataLake.R
\name{slurpFullFile}
\alias{slurpFullFile}
\title{Reads the whole file into buffer. Useful when reading small files.}
\usage{
slurpFullFile(adlFileInputStream, verbose = FALSE)
}
\arguments{
\item{adlFileInputStream}{the adlFileInputStream object to read from}

\item{verbose}{Print tracing information (default FALSE)}
}
\value{
number of bytes actually read
}
\description{
Reads the whole file into buffer. Useful when reading small files.
}
