% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{adlFileOutputStreamCheck}
\alias{adlFileOutputStreamCheck}
\title{Check the timestamp of a token and renew if needed.}
\usage{
adlFileOutputStreamCheck(adlFileOutputStream)
}
\description{
Check the timestamp of a token and renew if needed.
}
\seealso{
Other Azure resource functions: \code{\link{azureAuthenticateOnAuthType}},
  \code{\link{azureAuthenticate}},
  \code{\link{azureCheckToken}},
  \code{\link{azureGetTokenClientCredential}},
  \code{\link{azureGetTokenDeviceCodeFetch}},
  \code{\link{azureGetTokenDeviceCode}},
  \code{\link{azureGetTokenRefreshToken}},
  \code{\link{showDeviceCodeMessageToUser}}
}
