% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{AzureSM: AzureAuthenticate}
\alias{AzureAuthenticate}
\alias{AzureSM: AzureAuthenticate}
\title{Authenticates against Azure Active Directory application}
\usage{
AzureAuthenticate(AzureActiveContext, TID, CID, KEY, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{TID}{Tenant ID Object}

\item{CID}{Client ID Object}

\item{KEY}{Authentication KEY Object}

\item{Verbose}{Print Tracing information (Default False)}
}
\value{
Retunrs Azure Tokem and sets AzureContext Token
}
\note{
See \url{https://azure.microsoft.com/en-us/documentation/articles/resource-group-create-service-principal-portal/} to learn how to set up an Active Directory application
}
\references{
\url{https://azure.microsoft.com/en-us/documentation/articles/resource-group-create-service-principal-portal/}
}

