% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureAuthenticate.R
\name{azureAuthenticate}
\alias{azureAuthenticate}
\title{Authenticates against Azure Active directory application.}
\usage{
azureAuthenticate(azureActiveContext, TID, CID, KEY, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{TID}{The Tenant ID which was provided when the Active Directory application / service principal is created}

\item{CID}{The Client ID which was provided when the Active Directory application / service principal is created}

\item{KEY}{The Authentication Key which was provided when the Active Directory application / service principal is created}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Retunrs Azure Tokem and sets AzureContext Token
}
\description{
Authenticates against Azure Active directory application.
}
\note{
See \url{https://azure.microsoft.com/en-us/documentation/articles/resource-group-create-service-principal-portal/} to learn how to set up an Active directory application
}
\references{
\url{https://azure.microsoft.com/en-us/documentation/articles/resource-group-create-service-principal-portal/}
}
\seealso{
Other Resources: \code{\link{azureCheckToken}}
}

