% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureAuthenticate.R
\name{azureAuthenticate}
\alias{azureAuthenticate}
\title{Authenticates against Azure Active directory application.}
\usage{
azureAuthenticate(azureActiveContext, tenantID, clientID, authKey,
  verbose = FALSE, resource = "https\%3A\%2F\%2Fmanagement.azure.com\%2F")
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{tenantID}{The tenant ID provided during creation of the Active Directory application / service principal}

\item{clientID}{The client ID provided during creation of the Active Directory application / service principal}

\item{authKey}{The authentication key provided during creation of the Active Directory application / service principal}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
If successful, returns TRUE
}
\description{
Authenticates against Azure Active directory application.
}
\note{
See \url{https://azure.microsoft.com/en-us/documentation/articles/resource-group-create-service-principal-portal/} for instructions to set up an Active Directory application
}
\references{
\url{https://azure.microsoft.com/en-us/documentation/articles/resource-group-create-service-principal-portal/}
}
\seealso{
Other Azure resource functions: \code{\link{azureCheckToken}}
}
