% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureAuthenticate.R
\name{AzureAuthenticate}
\alias{AzureAuthenticate}
\title{Authenticates against Azure Active Directory application.}
\usage{
AzureAuthenticate(AzureActiveContext, TID, CID, KEY, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{TID}{Tenant ID Object}

\item{CID}{Client ID Object}

\item{KEY}{Authentication KEY Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Retunrs Azure Tokem and sets AzureContext Token
}
\description{
Authenticates against Azure Active Directory application.
}
\note{
See \url{https://azure.microsoft.com/en-us/documentation/articles/resource-group-create-service-principal-portal/} to learn how to set up an Active Directory application
}
\references{
\url{https://azure.microsoft.com/en-us/documentation/articles/resource-group-create-service-principal-portal/}
}
\seealso{
Other Resources: \code{\link{AzureCheckToken}},
  \code{\link{AzureCreateResourceGroup}},
  \code{\link{AzureDeleteResourceGroup}},
  \code{\link{AzureListAllRecources}},
  \code{\link{AzureListRG}},
  \code{\link{AzureListSubscriptions}}
}

