% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureAuthenticate.R
\name{azureAuthenticate}
\alias{azureAuthenticate}
\title{Authenticates against Azure Active directory application.}
\usage{
azureAuthenticate(azureActiveContext, tenantID, clientID, authKey,
  authType = "ClientCredential", resource = "https://management.azure.com/",
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{tenantID}{The tenant ID provided during creation of the Active Directory application / service principal}

\item{clientID}{The client ID provided during creation of the Active Directory application / service principal}

\item{authKey}{The authentication key provided during creation of the Active Directory application / service principal}

\item{authType}{Auth type for getting token: "ClientCredential", "DeviceCode"}

\item{resource}{URL of azure management portal}

\item{verbose}{If TRUE, prints verbose messages}
}
\value{
If successful, returns TRUE
}
\description{
Authenticates against Azure Active directory application.
}
\note{
See \url{https://azure.microsoft.com/en-us/documentation/articles/resource-group-create-service-principal-portal/} for instructions to set up an Active Directory application
}
\references{
\url{https://azure.microsoft.com/en-us/documentation/articles/resource-group-create-service-principal-portal/}
}
\seealso{
Other Azure resource functions: \code{\link{adlFileInputStreamCheck}},
  \code{\link{adlFileOutputStreamCheck}},
  \code{\link{azureAuthenticateOnAuthType}},
  \code{\link{azureCheckToken}},
  \code{\link{azureGetTokenClientCredential}},
  \code{\link{azureGetTokenDeviceCodeFetch}},
  \code{\link{azureGetTokenDeviceCode}},
  \code{\link{azureGetTokenRefreshToken}},
  \code{\link{showDeviceCodeMessageToUser}}
}
