% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azureblobCD}
\alias{azureblobCD}
\title{Azure blob change current directory.}
\usage{
azureblobCD(azureActiveContext, directory, container, file, storageAccount,
  storageKey, resourceGroup, subscriptionID, azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{directory}{Set directory to list}

\item{container}{container Object}

\item{file}{- Local filename to Store in Azure blob}

\item{storageAccount}{storageAccount Object}

\item{storageKey}{storageKey Object}

\item{resourceGroup}{resourceGroup Object}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Azure blob change current directory.
}
\seealso{
Other blob store functions: \code{\link{azureGetblob}},
  \code{\link{azureListSAblobs}},
  \code{\link{azurePutblob}}, \code{\link{azureblobFind}},
  \code{\link{azureblobLS}}
}

