% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azureBlobCD}
\alias{azureBlobCD}
\title{Azure blob change current directory.}
\usage{
azureBlobCD(azureActiveContext, directory, container, file, storageAccount,
  storageKey, resourceGroup, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}}

\item{directory}{Blob store directory to list for content}

\item{container}{Storage container name. See \code{\link[=azureListStorageContainers]{azureListStorageContainers()}}}

\item{file}{\itemize{
\item Local filename to store in Azure blob
}}

\item{storageAccount}{Name of the azure storage account. Storage account names must be between 3 and 24 characters in length and use numbers and lower-case letters only.}

\item{storageKey}{Storage key associated with storage account}

\item{resourceGroup}{Name of the resource group}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Azure blob change current directory.
}
\seealso{
Other Blob store functions: \code{\link{azureBlobFind}},
  \code{\link{azureBlobLS}}, \code{\link{azureDeleteBlob}},
  \code{\link{azureGetBlob}},
  \code{\link{azureListStorageBlobs}},
  \code{\link{azurePutBlob}}
}
