% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzureSM: AzureBlobCD}
\alias{AzureBlobCD}
\alias{AzureSM: AzureBlobCD}
\title{Azure Blob change current Directory}
\usage{
AzureBlobCD(AzureActiveContext, Directory, Container, File, StorageAccount,
  StorageKey, ResourceGroup, SubscriptionID, AzToken, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{Directory}{- Set Directory to Path (Supports (../ and ../..))}

\item{Container}{Storage Container}

\item{StorageAccount}{StorageAccount}

\item{StorageKey}{StorageKey}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{verbose}{Print Tracing information (Default False)}

\item{Token}{Token Object (or use AzureActiveContext)}
}

