% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azureblobCD}
\alias{azureblobCD}
\title{Azure blob change current directory.}
\usage{
azureblobCD(azureActiveContext, directory, container, file, storageAccount,
  storageKey, resourceGroup, subscriptionID, azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{directory}{Set directory to list}

\item{container}{Set the storage container}

\item{file}{- Local filename to Store in Azure blob}

\item{storageAccount}{Set the name of the azure storage account}

\item{storageKey}{Set the Storage Key associated with storage account}

\item{resourceGroup}{Set the name of the Resource Group}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Azure blob change current directory.
}
\seealso{
Other blob store functions: \code{\link{azureGetblob}},
  \code{\link{azureListSAblobs}},
  \code{\link{azurePutblob}}, \code{\link{azureblobFind}},
  \code{\link{azureblobLS}}
}

