% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzureBlobFind}
\alias{AzureBlobFind}
\title{Find File in a Storage account directory.}
\usage{
AzureBlobFind(AzureActiveContext, File, StorageAccount, StorageKey, Container,
  ResourceGroup, SubscriptionID, AzToken, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{File}{- Local FileName to Store in Azure Blob}

\item{StorageAccount}{StorageAccount Object}

\item{StorageKey}{StorageKey Object}

\item{Container}{Container Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Find File in a Storage account directory.
}
\seealso{
Other Blob store functions: \code{\link{AzureBlobCD}},
  \code{\link{AzureBlobLS}}, \code{\link{AzureGetBlob}},
  \code{\link{AzureListSABlobs}},
  \code{\link{AzurePutBlob}}
}

