% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azureBlobFind}
\alias{azureBlobFind}
\title{Find file in a Storage account directory.}
\usage{
azureBlobFind(azureActiveContext, file, storageAccount, storageKey, container,
  resourceGroup, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}}

\item{file}{\itemize{
\item Local filename to store in Azure blob
}}

\item{storageAccount}{Name of the azure storage account}

\item{storageKey}{Storage key associated with storage account}

\item{container}{Storage container name. See \code{\link[=azureListStorageContainers]{azureListStorageContainers()}}}

\item{resourceGroup}{Name of the resource group}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Find file in a Storage account directory.
}
\seealso{
Other Blob store functions: \code{\link{azureBlobCD}},
  \code{\link{azureBlobLS}}, \code{\link{azureDeleteBlob}},
  \code{\link{azureGetBlob}},
  \code{\link{azureListStorageBlobs}},
  \code{\link{azurePutBlob}}
}
