% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azureBlobFind}
\alias{azureBlobFind}
\title{Find file in a storage account directory.}
\usage{
azureBlobFind(azureActiveContext, file, storageAccount, storageKey, container,
  resourceGroup, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Either an `azureActiveContext` object or `NULL`. The blob store functions support two modes of connecting to the Azure API: authenticate using Active Directory or providing your own storage key. If this value is `NULL`, you must provide the `storageKey`.}

\item{file}{- Local filename to store in Azure blob}

\item{storageAccount}{Name of the azure storage account. Storage account names must be between 3 and 24 characters in length and use numbers and lower-case letters only.}

\item{storageKey}{Storage key associated with storage account}

\item{container}{Storage container name. See [azureListStorageContainers()]}

\item{resourceGroup}{Name of the resource group}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Find file in a storage account directory.
}
\section{Using blob store functions without authentication}{

The blob store functions support two modes of connecting to the Azure API:authenticate using Active Directory or providing your own storage key. If this value is `NULL`, you must provide the `storageKey` .
}
\seealso{
Other Blob store functions: \code{\link{azureBlobCD}},
  \code{\link{azureBlobLS}}, \code{\link{azureDeleteBlob}},
  \code{\link{azureGetBlob}},
  \code{\link{azureListStorageBlobs}},
  \code{\link{azurePutBlob}}
}

