% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azureBlobLS}
\alias{azureBlobLS}
\title{List blob blobs in a storage account directory.}
\usage{
azureBlobLS(azureActiveContext, directory, recursive = FALSE, storageAccount,
  storageKey, container, resourceGroup, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}}

\item{directory}{Blob store directory to list for content}

\item{recursive}{If TRUE, list blob store directories recursively}

\item{storageAccount}{Name of the azure storage account. Storage account names must be between 3 and 24 characters in length and use numbers and lower-case letters only.}

\item{storageKey}{Storage key associated with storage account}

\item{container}{Storage container name. See \code{\link[=azureListStorageContainers]{azureListStorageContainers()}}}

\item{resourceGroup}{Name of the resource group}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List blob blobs in a storage account directory.
}
\seealso{
Other Blob store functions: \code{\link{azureBlobCD}},
  \code{\link{azureBlobFind}},
  \code{\link{azureDeleteBlob}},
  \code{\link{azureGetBlob}},
  \code{\link{azureListStorageBlobs}},
  \code{\link{azurePutBlob}}
}
