% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azureblobLS}
\alias{azureblobLS}
\title{List blob files in a Storage account directory.}
\usage{
azureblobLS(azureActiveContext, directory, recursive = FALSE, storageAccount,
  storageKey, container, resourceGroup, subscriptionID, azToken,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{directory}{Set directory to list}

\item{recursive}{List directories recursively (Default FALSE)}

\item{storageAccount}{Name of the azure storage account}

\item{storageKey}{Storage key associated with storage account}

\item{container}{Storage container name. See \code{\link{azureListStorageContainers}}}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Set the subscriptionID.  This is obtained automatically by \code{\link{azureAuthenticate}} when only a single subscriptionID is available via Active Directory}

\item{azToken}{Azure authentication token, obtained by \code{\link{azureAuthenticate}}}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List blob files in a Storage account directory.
}
\seealso{
Other blob store functions: \code{\link{azureBlobCD}},
  \code{\link{azureBlobFind}}, \code{\link{azureGetBlob}},
  \code{\link{azureListStorageBlobs}},
  \code{\link{azurePutBlob}}
}

