% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azureblobLS}
\alias{azureblobLS}
\title{List blob files in a Storage account directory.}
\usage{
azureblobLS(azureActiveContext, directory, recursive = FALSE, storageAccount,
  storageKey, container, resourceGroup, subscriptionID, azToken,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{directory}{Set directory to list}

\item{recursive}{List directories recursively (Default FALSE)}

\item{storageAccount}{Set the name of the azure storage account}

\item{storageKey}{Set the Storage Key associated with storage account}

\item{container}{Set the storage container}

\item{resourceGroup}{Set the name of the Resource Group}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List blob files in a Storage account directory.
}
\seealso{
Other blob store functions: \code{\link{azureGetblob}},
  \code{\link{azureListSAblobs}},
  \code{\link{azurePutblob}}, \code{\link{azureblobCD}},
  \code{\link{azureblobFind}}
}

