% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{AzureCancelDeploy}
\alias{AzureCancelDeploy}
\title{Cancel Template Deployment.}
\usage{
AzureCancelDeploy(AzureActiveContext, DeplName, ResourceGroup, SubscriptionID,
  AzToken, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{DeplName}{DeplName}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Cancel Template Deployment.
}
\seealso{
Other Template functions: \code{\link{AzureDeleteDeploy}},
  \code{\link{AzureDeployStatus}},
  \code{\link{AzureDeployTemplate}}
}

