% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{azureCancelDeploy}
\alias{azureCancelDeploy}
\title{Cancel Template Deployment.}
\usage{
azureCancelDeploy(azureActiveContext, deplname, resourceGroup, subscriptionID,
  azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{deplname}{deplname}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Set the subscriptionID.  This is obtained automatically by \code{\link{azureAuthenticate}} when only a single subscriptionID is available via Active Directory}

\item{azToken}{Azure authentication token, obtained by \code{\link{azureAuthenticate}}}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Cancel Template Deployment.
}
\seealso{
Other Template functions: \code{\link{azureDeleteDeploy}},
  \code{\link{azureDeployStatus}},
  \code{\link{azureDeployTemplate}}
}
