% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureAuthenticate.R
\name{AzureCheckToken}
\alias{AzureCheckToken}
\title{Check the timestamp of a Token and Renew if needed.}
\usage{
AzureCheckToken(AzureActiveContext)
}
\arguments{
\item{AzureActiveContext}{Context Object}
}
\description{
Check the timestamp of a Token and Renew if needed.
}
\seealso{
Other Resources: \code{\link{AzureAuthenticate}}
}

