% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{AzureCheckToken}
\alias{AzureCheckToken}
\title{Check the timestamp of a Token and Renew if needed.}
\usage{
AzureCheckToken(AzureActiveContext)
}
\arguments{
\item{AzureActiveContext}{Context Object}
}
\description{
Check the timestamp of a Token and Renew if needed.
}
\seealso{
Other Resources: \code{\link{AzureAuthenticate}},
  \code{\link{AzureCreateResourceGroup}},
  \code{\link{AzureDeleteResourceGroup}},
  \code{\link{AzureListAllRecources}},
  \code{\link{AzureListRG}},
  \code{\link{AzureListSubscriptions}}
}

