% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureAuthenticate.R
\name{azureCheckToken}
\alias{azureCheckToken}
\title{Check the timestamp of a Token and Renew if needed.}
\usage{
azureCheckToken(azureActiveContext)
}
\arguments{
\item{azureActiveContext}{Context Object}
}
\description{
Check the timestamp of a Token and Renew if needed.
}
\seealso{
Other Resources: \code{\link{azureAuthenticate}}
}

