% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{azureCreateHDI}
\alias{azureCreateHDI}
\title{Create Specifed HDInsight Cluster.}
\usage{
azureCreateHDI(azureActiveContext, clustername, location, kind = "spark",
  storageAccount, storageKey, version = "3.4", workers = 2, adminUser,
  adminPassword, sshUser, sshPassword, hiveServer, hiveDB, hiveUser,
  hivePassword, resourceGroup, azToken, subscriptionID, mode = "Sync",
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{clustername}{resourceGroup Object (or use azureActiveContext)}

\item{location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{kind}{"hadoop","spark" or "pyspark"}

\item{storageAccount}{storageAccount Object}

\item{storageKey}{Storage Key}

\item{version}{version}

\item{workers}{Number of worker nodes}

\item{adminUser}{Admin user name}

\item{adminPassword}{Admin user password}

\item{sshUser}{SSH user name}

\item{sshPassword}{SSH user password}

\item{hiveServer}{Hive server}

\item{hiveDB}{Hive DB}

\item{hiveUser}{Hive user name}

\item{hivePassword}{Hive user password}

\item{resourceGroup}{resourceGroup Object}

\item{azToken}{Token Object}

\item{subscriptionID}{subscriptionID Object}

\item{mode}{mode}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Success message
}
\description{
Create Specifed HDInsight Cluster.
}
\seealso{
Other HDInsight functions: \code{\link{azureDeleteHDI}},
  \code{\link{azureHDIConf}}, \code{\link{azureListHDI}},
  \code{\link{azureResizeHDI}},
  \code{\link{azureRunScriptAction}},
  \code{\link{azureScriptActionHistory}}
}

