% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{AzureCreateHDI}
\alias{AzureCreateHDI}
\title{Create Specifed HDInsight Cluster.}
\usage{
AzureCreateHDI(AzureActiveContext, ClusterName, Location, Kind = "spark",
  StorageAcc, SKey, Version = "3.4", Workers = 2, AdminUser, AdminPassword,
  SSHUser, SSHPassword, HiveServer, HiveDB, HiveUser, HivePassword,
  ResourceGroup, AzToken, SubscriptionID, Mode = "Sync", verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ClusterName}{ClusterName Object}

\item{Location}{- Location String}

\item{Kind}{- Kind (spark/hadoop) DEFAULT{spark}}

\item{Workers}{- # of Workers (Default 2)}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Success Message
}
\description{
Create Specifed HDInsight Cluster.
}
\seealso{
Other HDInsight: \code{\link{AzureDeleteHDI}},
  \code{\link{AzureHDIConf}}, \code{\link{AzureListHDI}},
  \code{\link{AzureResizeHDI}},
  \code{\link{AzureRunScriptAction}},
  \code{\link{AzureScriptActionHistory}}
}

