% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{azureCreateHDI}
\alias{azureCreateHDI}
\title{Create Specifed HDInsight Cluster.}
\usage{
azureCreateHDI(azureActiveContext, clustername, location, kind = "spark",
  storageAccount, storageKey, version = "3.4", componentVersion = "1.6.2",
  workers = 2, adminUser, adminPassword, sshUser, sshPassword, hiveServer,
  hiveDB, hiveUser, hivePassword, resourceGroup, vmSize = "Large",
  subscriptionID, mode = "Sync", verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}}

\item{clustername}{Cluster name, used for HDI and Spark clusters. See \code{\link[=azureCreateHDI]{azureCreateHDI()}}}

\item{location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{kind}{HDinsight kind: "hadoop","spark" or "rserver"}

\item{storageAccount}{Name of the azure storage account}

\item{storageKey}{Storage key associated with storage account}

\item{version}{HDinsight version}

\item{componentVersion}{Spark componentVersion. Default : 1.6.2}

\item{workers}{Define the number of worker nodes}

\item{adminUser}{Admin user name}

\item{adminPassword}{Admin user password}

\item{sshUser}{SSH user name}

\item{sshPassword}{SSH user password}

\item{hiveServer}{URI address of the Hive server}

\item{hiveDB}{Hive DB name}

\item{hiveUser}{Hive user name}

\item{hivePassword}{Hive user password}

\item{resourceGroup}{Name of the resource group}

\item{vmSize}{Size of nodes: "Large", "Small", "Standard_D14_V2", etc.}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{mode}{Provisioning mode, "Sync" or "Async". Use "Async" to immediately return to R session after submission of request}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Success message
}
\description{
Create Specifed HDInsight Cluster.
}
\note{
See \url{https://docs.microsoft.com/en-us/azure/hdinsight/hdinsight-component-versioning} to learn about HDInsight Versions
}
\seealso{
Other HDInsight functions: \code{\link{azureDeleteHDI}},
  \code{\link{azureHDIConf}}, \code{\link{azureListHDI}},
  \code{\link{azureResizeHDI}},
  \code{\link{azureRunScriptAction}},
  \code{\link{azureScriptActionHistory}}
}
