% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{azureCreateHDI}
\alias{azureCreateHDI}
\title{Create Specifed HDInsight Cluster.}
\usage{
azureCreateHDI(azureActiveContext, clustername, location, kind = "spark",
  storageAccount, storageKey, version = "3.4", workers = 2, adminUser,
  adminPassword, sshUser, sshPassword, hiveServer, hiveDB, hiveUser,
  hivePassword, resourceGroup, azToken, subscriptionID, mode = "Sync",
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{clustername}{set the Name of cluster}

\item{location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{kind}{Set the HDinsight kind "hadoop","spark" or "pyspark"}

\item{storageAccount}{Set the name of the azure storage account}

\item{storageKey}{set the Storage Key manually for the associaetd storage account}

\item{version}{HDinsight version see}

\item{workers}{Define the number of worker nodes}

\item{adminUser}{Admin user name}

\item{adminPassword}{Admin user password}

\item{sshUser}{set the SSH user name}

\item{sshPassword}{set the SSH user password}

\item{hiveServer}{URI address of the Hive server}

\item{hiveDB}{Set the name of the Hive DB}

\item{hiveUser}{Set the hive user name}

\item{hivePassword}{Set the Hive user password}

\item{resourceGroup}{Set the name of the Resource Group}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{mode}{Set Provisioning mode - Default Sync (Syncronous), set to ASYNC to return to session after submission}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Success message
}
\description{
Create Specifed HDInsight Cluster.
}
\note{
See \url{https://docs.microsoft.com/en-us/azure/hdinsight/hdinsight-component-versioning} to learn about HDInsight Versions
}
\seealso{
Other HDInsight functions: \code{\link{azureDeleteHDI}},
  \code{\link{azureHDIConf}}, \code{\link{azureListHDI}},
  \code{\link{azureResizeHDI}},
  \code{\link{azureRunScriptAction}},
  \code{\link{azureScriptActionHistory}}
}

