% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{AzureCreateHDI}
\alias{AzureCreateHDI}
\title{Create Specifed HDInsight Cluster.}
\usage{
AzureCreateHDI(AzureActiveContext, ClusterName, Location, Kind = "spark",
  StorageAcc, SKey, Version = "3.4", Workers = 2, AdminUser, AdminPassword,
  SSHUser, SSHPassword, HiveServer, HiveDB, HiveUser, HivePassword,
  ResourceGroup, AzToken, subscriptionID, Mode = "Sync", verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ClusterName}{ResourceGroup Object (or use AzureActiveContext)}

\item{Location}{Location string}

\item{Kind}{"hadoop","spark" or "pyspark"}

\item{SKey}{Storage Key}

\item{Version}{Version}

\item{Workers}{Number of worker nodes}

\item{AdminUser}{Admin user name}

\item{AdminPassword}{Admin user password}

\item{SSHUser}{SSH user name}

\item{SSHPassword}{SSH user password}

\item{HiveServer}{Hive server}

\item{HiveDB}{Hive DB}

\item{HiveUser}{Hive user name}

\item{HivePassword}{Hive user password}

\item{ResourceGroup}{ResourceGroup Object}

\item{AzToken}{Token Object}

\item{Mode}{Mode}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Success message
}
\description{
Create Specifed HDInsight Cluster.
}
\seealso{
Other HDInsight: \code{\link{AzureDeleteHDI}},
  \code{\link{AzureHDIConf}}, \code{\link{AzureListHDI}},
  \code{\link{AzureResizeHDI}},
  \code{\link{AzureRunScriptAction}},
  \code{\link{AzureScriptActionHistory}}
}

