% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{azureCreateResourceGroup}
\alias{azureCreateResourceGroup}
\title{Create a resourceGroup.}
\usage{
azureCreateResourceGroup(azureActiveContext, resourceGroup, location,
  subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{resourceGroup}{Name of the resource group}

\item{location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of Resources
}
\description{
Create a resourceGroup.
}
\seealso{
Other Resource group functions: \code{\link{azureDeleteResourceGroup}},
  \code{\link{azureListAllResources}},
  \code{\link{azureListRG}},
  \code{\link{azureListSubscriptions}}
}
