% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{AzureCreateResourceGroup}
\alias{AzureCreateResourceGroup}
\title{Create a ResourceGroup.}
\usage{
AzureCreateResourceGroup(AzureActiveContext, ResourceGroup, Location,
  SubscriptionID, AzToken, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{Location}{Location string}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of Resources
}
\description{
Create a ResourceGroup.
}
\seealso{
Other Resource group functions: \code{\link{AzureDeleteResourceGroup}},
  \code{\link{AzureListAllResources}},
  \code{\link{AzureListRG}},
  \code{\link{AzureListSubscriptions}}
}

