% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{azureCreateResourceGroup}
\alias{azureCreateResourceGroup}
\title{Create a resourceGroup.}
\usage{
azureCreateResourceGroup(azureActiveContext, resourceGroup, location,
  subscriptionID, azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{resourceGroup}{Set the name of the Resource Group}

\item{location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of Resources
}
\description{
Create a resourceGroup.
}
\seealso{
Other Resource group functions: \code{\link{azureDeleteResourceGroup}},
  \code{\link{azureListAllResources}},
  \code{\link{azureListRG}},
  \code{\link{azureListSubscriptions}}
}

