% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{azureCreateresourceGroup}
\alias{azureCreateresourceGroup}
\title{Create a resourceGroup.}
\usage{
azureCreateresourceGroup(azureActiveContext, resourceGroup, location,
  subscriptionID, azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{resourceGroup}{resourceGroup Object}

\item{location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of Resources
}
\description{
Create a resourceGroup.
}
\seealso{
Other Resource group functions: \code{\link{azureDeleteresourceGroup}},
  \code{\link{azureListAllResources}},
  \code{\link{azureListRG}},
  \code{\link{azureListSubscriptions}}
}

