% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzureSM: AzureCreateSAContainer}
\alias{AzureCreateSAContainer}
\alias{AzureSM: AzureCreateSAContainer}
\title{Create Storage Containers in a specified Storage Account}
\usage{
AzureCreateSAContainer(AzureActiveContext, Container, StorageAccount,
  StorageKey, ResourceGroup, AzToken, SubscriptionID, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{Container}{- Storage Container}

\item{StorageAccount}{StorageAccount}

\item{StorageKey}{StorageKey}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{verbose}{Print Tracing information (Default False)}

\item{Token}{Token Object (or use AzureActiveContext)}
}

