% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzureCreateSAContainer}
\alias{AzureCreateSAContainer}
\title{Create Storage Containers in a specified Storage Account.}
\usage{
AzureCreateSAContainer(AzureActiveContext, Container, StorageAccount,
  StorageKey, ResourceGroup, AzToken, SubscriptionID, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{Container}{Container Object}

\item{StorageAccount}{StorageAccount Object}

\item{StorageKey}{StorageKey Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{AzToken}{Token Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object}
}
\description{
Create Storage Containers in a specified Storage Account.
}
\seealso{
Other Storage account: \code{\link{AzureBlobCD}},
  \code{\link{AzureBlobFind}}, \code{\link{AzureBlobLS}},
  \code{\link{AzureCreateStorageAccount}},
  \code{\link{AzureDeleteSAContainer}},
  \code{\link{AzureDeleteStorageAccount}},
  \code{\link{AzureGetBlob}},
  \code{\link{AzureListSABlobs}},
  \code{\link{AzureListSAContainers}},
  \code{\link{AzureListSA}}, \code{\link{AzurePutBlob}},
  \code{\link{AzureSAGetKey}}
}

