% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureStorageAccount.R
\name{AzureCreateStorageAccount}
\alias{AzureCreateStorageAccount}
\title{Create an Azure Storage Account.}
\usage{
AzureCreateStorageAccount(AzureActiveContext, StorageAccount, AzToken,
  ResourceGroup, SubscriptionID, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{StorageAccount}{StorageAccount Object}

\item{AzToken}{Token Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Create an Azure Storage Account.
}
\seealso{
Other Storage account: \code{\link{AzureBlobCD}},
  \code{\link{AzureBlobFind}}, \code{\link{AzureBlobLS}},
  \code{\link{AzureCreateSAContainer}},
  \code{\link{AzureDeleteSAContainer}},
  \code{\link{AzureDeleteStorageAccount}},
  \code{\link{AzureGetBlob}},
  \code{\link{AzureListSABlobs}},
  \code{\link{AzureListSAContainers}},
  \code{\link{AzureListSA}}, \code{\link{AzurePutBlob}},
  \code{\link{AzureSAGetKey}}
}

