% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContainer.R
\name{azureCreateStorageContainer}
\alias{azureCreateStorageContainer}
\title{Create Storage containers in a specified Storage Account.}
\usage{
azureCreateStorageContainer(azureActiveContext, container, storageAccount,
  storageKey, resourceGroup, azToken, subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{container}{container Object}

\item{storageAccount}{storageAccount Object}

\item{storageKey}{storageKey Object}

\item{resourceGroup}{resourceGroup Object}

\item{azToken}{Token Object}

\item{subscriptionID}{subscriptionID Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Create Storage containers in a specified Storage Account.
}
\seealso{
Other container functions: \code{\link{azureDeleteStorageContainer}},
  \code{\link{azureListStorageContainers}}
}

