% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContainer.R
\name{azureCreateStorageContainer}
\alias{azureCreateStorageContainer}
\title{Create Storage containers in a specified Storage Account.}
\usage{
azureCreateStorageContainer(azureActiveContext, container, storageAccount,
  storageKey, resourceGroup, subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}}

\item{container}{Storage container name. See \code{\link[=azureListStorageContainers]{azureListStorageContainers()}}}

\item{storageAccount}{Name of the azure storage account. Storage account names must be between 3 and 24 characters in length and use numbers and lower-case letters only.}

\item{storageKey}{Storage key associated with storage account}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Create Storage containers in a specified Storage Account.
}
\seealso{
Other Container functions: \code{\link{azureDeleteStorageContainer}},
  \code{\link{azureListStorageContainers}}
}
