% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureDataLake.R
\name{azureDataLakeListStatus}
\alias{azureDataLakeListStatus}
\title{Azure Data Lake LISTSTATUS for specified relativePath of an azure data lake account.}
\usage{
azureDataLakeListStatus(azureActiveContext, azureDataLakeAccount,
  relativePath = "", verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Provide an \code{azureActiveContext} object used for authentication.}

\item{azureDataLakeAccount}{Provide the name of the Azure Data Lake account.}

\item{relativePath}{Provide a relative path of the directory.}

\item{verbose}{Print tracing information (default FALSE).}
}
\value{
Returns a data frame.
}
\description{
Azure Data Lake LISTSTATUS for specified relativePath of an azure data lake account.
}
\seealso{
Other Azure Data Lake Store functions: \code{\link{azureDataLakeAppend}},
  \code{\link{azureDataLakeCreate}},
  \code{\link{azureDataLakeDelete}},
  \code{\link{azureDataLakeGetFileStatus}},
  \code{\link{azureDataLakeMkdirs}},
  \code{\link{azureDataLakeOpen}}
}
