% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBatch.R
\name{azureDeleteBatchAccount}
\alias{azureDeleteBatchAccount}
\title{Delete an azure batch account.}
\usage{
azureDeleteBatchAccount(azureActiveContext, batchAccount, resourceGroup,
  subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by `AzureSMR`, created by [createAzureContext()]}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by [azureAuthenticate()] when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Delete an azure batch account.
}
\references{
https://docs.microsoft.com/en-us/rest/api/batchmanagement/batchaccount#BatchAccount_Delete
}
\seealso{
Other Batch account functions: \code{\link{azureBatchGetKey}},
  \code{\link{azureCreateBatchAccount}},
  \code{\link{azureListBatchAccounts}}
}

