% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{azureDeleteDeploy}
\alias{azureDeleteDeploy}
\title{Delete Template Deployment.}
\usage{
azureDeleteDeploy(azureActiveContext, deplname, resourceGroup, subscriptionID,
  azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{deplname}{deplname}

\item{resourceGroup}{resourceGroup Object}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Delete Template Deployment.
}
\seealso{
Other Template functions: \code{\link{azureCancelDeploy}},
  \code{\link{azureDeployStatus}},
  \code{\link{azureDeployTemplate}}
}

