% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{AzureDeleteDeploy}
\alias{AzureDeleteDeploy}
\title{Delete Template Deployment.}
\usage{
AzureDeleteDeploy(AzureActiveContext, DeplName, ResourceGroup, SubscriptionID,
  AzToken, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{DeplName}{DeplName}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Delete Template Deployment.
}
\seealso{
Other Template functions: \code{\link{AzureCancelDeploy}},
  \code{\link{AzureDeployStatus}},
  \code{\link{AzureDeployTemplate}}
}

