% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{AzureDeleteDeploy}
\alias{AzureDeleteDeploy}
\title{Delete Template Deployment.}
\usage{
AzureDeleteDeploy(AzureActiveContext, DeplName, ResourceGroup, SubscriptionID,
  AzToken, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}
}
\description{
Delete Template Deployment.
}

