% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{azureDeleteDeploy}
\alias{azureDeleteDeploy}
\title{Delete Template Deployment.}
\usage{
azureDeleteDeploy(azureActiveContext, deplname, resourceGroup, subscriptionID,
  azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{deplname}{deplname}

\item{resourceGroup}{Set the name of the Resource Group}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Delete Template Deployment.
}
\seealso{
Other Template functions: \code{\link{azureCancelDeploy}},
  \code{\link{azureDeployStatus}},
  \code{\link{azureDeployTemplate}}
}

