% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{AzureSM: AzureDeleteHDI}
\alias{AzureDeleteHDI}
\alias{AzureSM: AzureDeleteHDI}
\title{Delete Specifed HDInsight Cluster}
\usage{
AzureDeleteHDI(AzureActiveContext, ClusterName, AzToken, SubscriptionID,
  ResourceGroup, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{ClusterName}{ResourceGroup Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{verbose}{Print Tracing information (Default False)}

\item{Token}{Token Object (or use AzureActiveContext)}
}
\value{
Returns Dataframe of HDInsight Clusters information
}

