% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{azureDeleteHDI}
\alias{azureDeleteHDI}
\title{Delete Specifed HDInsight Cluster.}
\usage{
azureDeleteHDI(azureActiveContext, clustername, azToken, subscriptionID,
  resourceGroup, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{clustername}{clustername Object}

\item{azToken}{Token Object}

\item{subscriptionID}{subscriptionID Object}

\item{resourceGroup}{resourceGroup Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of HDInsight Clusters information
}
\description{
Delete Specifed HDInsight Cluster.
}
\seealso{
Other HDInsight functions: \code{\link{azureCreateHDI}},
  \code{\link{azureHDIConf}}, \code{\link{azureListHDI}},
  \code{\link{azureResizeHDI}},
  \code{\link{azureRunScriptAction}},
  \code{\link{azureScriptActionHistory}}
}

