% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{azureDeleteHDI}
\alias{azureDeleteHDI}
\title{Delete Specifed HDInsight Cluster.}
\usage{
azureDeleteHDI(azureActiveContext, clustername, azToken, subscriptionID,
  resourceGroup, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{clustername}{Set the clustername}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{resourceGroup}{Set the name of the Resource Group}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of HDInsight Clusters information
}
\description{
Delete Specifed HDInsight Cluster.
}
\seealso{
Other HDInsight functions: \code{\link{azureCreateHDI}},
  \code{\link{azureHDIConf}}, \code{\link{azureListHDI}},
  \code{\link{azureResizeHDI}},
  \code{\link{azureRunScriptAction}},
  \code{\link{azureScriptActionHistory}}
}

