% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{AzureDeleteHDI}
\alias{AzureDeleteHDI}
\title{Delete Specifed HDInsight Cluster.}
\usage{
AzureDeleteHDI(AzureActiveContext, ClusterName, AzToken, SubscriptionID,
  ResourceGroup, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ClusterName}{ClusterName Object}

\item{AzToken}{Token Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of HDInsight Clusters information
}
\description{
Delete Specifed HDInsight Cluster.
}
\seealso{
Other HDInsight functions: \code{\link{AzureCreateHDI}},
  \code{\link{AzureHDIConf}}, \code{\link{AzureListHDI}},
  \code{\link{AzureResizeHDI}},
  \code{\link{AzureRunScriptAction}},
  \code{\link{AzureScriptActionHistory}}
}

