% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{AzureDeleteResourceGroup}
\alias{AzureDeleteResourceGroup}
\title{Delete a ResourceGroup with all Resources.}
\usage{
AzureDeleteResourceGroup(AzureActiveContext, ResourceGroup, SubscriptionID,
  AzToken, Type, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object}

\item{Type}{Type}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of Resources
}
\description{
Delete a ResourceGroup with all Resources.
}
\seealso{
Other Resource group functions: \code{\link{AzureCreateResourceGroup}},
  \code{\link{AzureListAllResources}},
  \code{\link{AzureListRG}},
  \code{\link{AzureListSubscriptions}}
}

