% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{azureDeleteresourceGroup}
\alias{azureDeleteresourceGroup}
\title{Delete a resourceGroup with all Resources.}
\usage{
azureDeleteresourceGroup(azureActiveContext, resourceGroup, subscriptionID,
  azToken, type, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{resourceGroup}{Set the name of the Resource Group}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{type}{filter by resource type}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of Resources
}
\description{
Delete a resourceGroup with all Resources.
}
\seealso{
Other Resource group functions: \code{\link{azureCreateresourceGroup}},
  \code{\link{azureListAllResources}},
  \code{\link{azureListRG}},
  \code{\link{azureListSubscriptions}}
}

