% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{azureDeleteresourceGroup}
\alias{azureDeleteresourceGroup}
\title{Delete a resourceGroup with all Resources.}
\usage{
azureDeleteresourceGroup(azureActiveContext, resourceGroup, subscriptionID,
  azToken, type, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{resourceGroup}{resourceGroup Object}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{type}{type}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of Resources
}
\description{
Delete a resourceGroup with all Resources.
}
\seealso{
Other Resource group functions: \code{\link{azureCreateresourceGroup}},
  \code{\link{azureListAllResources}},
  \code{\link{azureListRG}},
  \code{\link{azureListSubscriptions}}
}

