% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContainer.R
\name{AzureDeleteSAContainer}
\alias{AzureDeleteSAContainer}
\title{Delete Storage Container in a specified Storage Account.}
\usage{
AzureDeleteSAContainer(AzureActiveContext, Container, StorageAccount,
  StorageKey, ResourceGroup, AzToken, SubscriptionID, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{Container}{Container Object}

\item{StorageAccount}{StorageAccount Object}

\item{StorageKey}{StorageKey Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{AzToken}{Token Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Delete Storage Container in a specified Storage Account.
}
\seealso{
Other Storage account: \code{\link{AzureCreateSAContainer}},
  \code{\link{AzureListSAContainers}}
}

