% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzureDeleteStorageAccount}
\alias{AzureDeleteStorageAccount}
\title{Delete an Azure Storage Account.}
\usage{
AzureDeleteStorageAccount(AzureActiveContext, StorageAccount, AzToken,
  ResourceGroup, SubscriptionID, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{StorageAccount}{StorageAccount Object}

\item{AzToken}{Token Object (or use AzureActiveContext)}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}
}
\description{
Delete an Azure Storage Account.
}
\seealso{
Other Storage account: \code{\link{AzureBlobCD}},
  \code{\link{AzureBlobFind}}, \code{\link{AzureBlobLS}},
  \code{\link{AzureCreateSAContainer}},
  \code{\link{AzureCreateStorageAccount}},
  \code{\link{AzureDeleteSAContainer}},
  \code{\link{AzureGetBlob}},
  \code{\link{AzureListSABlobs}},
  \code{\link{AzureListSAContainers}},
  \code{\link{AzureListSA}}, \code{\link{AzurePutBlob}},
  \code{\link{AzureSAGetKey}}
}

