% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureStorageAccount.R
\name{AzureDeleteStorageAccount}
\alias{AzureDeleteStorageAccount}
\title{Delete an Azure Storage Account.}
\usage{
AzureDeleteStorageAccount(AzureActiveContext, StorageAccount, AzToken,
  ResourceGroup, SubscriptionID, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{StorageAccount}{StorageAccount Object}

\item{AzToken}{Token Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Delete an Azure Storage Account.
}
\seealso{
Other Storage account functions: \code{\link{AzureCreateStorageAccount}},
  \code{\link{AzureListSA}}, \code{\link{AzureSAGetKey}}
}

