% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureStorageAccount.R
\name{azureDeletestorageAccount}
\alias{azureDeletestorageAccount}
\title{Delete an Azure Storage Account.}
\usage{
azureDeletestorageAccount(azureActiveContext, storageAccount, azToken,
  resourceGroup, subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{storageAccount}{storageAccount Object}

\item{azToken}{Token Object}

\item{resourceGroup}{resourceGroup Object}

\item{subscriptionID}{subscriptionID Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Delete an Azure Storage Account.
}
\seealso{
Other Storage account functions: \code{\link{azureCreateStorageAccount}},
  \code{\link{azureListSA}}, \code{\link{azureSAGetKey}}
}

