% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureStorageAccount.R
\name{azureDeletestorageAccount}
\alias{azureDeletestorageAccount}
\title{Delete an Azure Storage Account.}
\usage{
azureDeletestorageAccount(azureActiveContext, storageAccount, azToken,
  resourceGroup, subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{storageAccount}{Name of the azure storage account}

\item{azToken}{Azure authentication token, obtained by \code{\link{azureAuthenticate}}}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Set the subscriptionID.  This is obtained automatically by \code{\link{azureAuthenticate}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Delete an Azure Storage Account.
}
\seealso{
Other Storage account functions: \code{\link{azureCreateStorageAccount}},
  \code{\link{azureListSA}}, \code{\link{azureSAGetKey}}
}
