% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContainer.R
\name{azureDeleteStorageContainer}
\alias{azureDeleteStorageContainer}
\title{Delete Storage container in a specified Storage Account.}
\usage{
azureDeleteStorageContainer(azureActiveContext, container, storageAccount,
  storageKey, resourceGroup, azToken, subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{container}{Set the storage container}

\item{storageAccount}{Set the name of the azure storage account}

\item{storageKey}{Set the Storage Key associated with storage account}

\item{resourceGroup}{Set the name of the Resource Group}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Delete Storage container in a specified Storage Account.
}
\seealso{
Other container functions: \code{\link{azureCreateStorageContainer}},
  \code{\link{azureListStorageContainers}}
}

