% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContainer.R
\name{azureDeleteStorageContainer}
\alias{azureDeleteStorageContainer}
\title{Delete Storage container in a specified Storage Account.}
\usage{
azureDeleteStorageContainer(azureActiveContext, container, storageAccount,
  storageKey, resourceGroup, subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by `AzureSMR`, created by [createAzureContext()]}

\item{container}{Storage container name. See [azureListStorageContainers()]}

\item{storageAccount}{Name of the azure storage account. Storage account names must be between 3 and 24 characters in length and use numbers and lower-case letters only.}

\item{storageKey}{Storage key associated with storage account}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by [azureAuthenticate()] when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Delete Storage container in a specified Storage Account.
}
\seealso{
Other Container functions: \code{\link{azureCreateStorageContainer}},
  \code{\link{azureListStorageContainers}}
}

