% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureDeleteVM}
\alias{azureDeleteVM}
\title{Delete a Virtual Machine.}
\usage{
azureDeleteVM(azureActiveContext, resourceGroup, vmName, subscriptionID,
  azToken, mode = "Sync", verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{resourceGroup}{resourceGroup Object}

\item{vmName}{Virtual Machine name}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{mode}{Wait for operation to complete 'Sync' (Default)}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Delete a Virtual Machine.
}
\seealso{
Other Virtual machine functions: \code{\link{azureListVM}},
  \code{\link{azureStartVM}}, \code{\link{azureStopVM}},
  \code{\link{azureVMStatus}}
}

