% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{AzureDeleteVM}
\alias{AzureDeleteVM}
\title{Delete a Virtual Machine.}
\usage{
AzureDeleteVM(AzureActiveContext, ResourceGroup, VMName, SubscriptionID,
  AzToken, Mode = "Sync", verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{VMName}{Virtual Machine Name}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object}

\item{Mode}{Wait for operation to complete 'Sync' (Default)}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Delete a Virtual Machine.
}
\seealso{
Other Virtual machine functions: \code{\link{AzureListVM}},
  \code{\link{AzureStartVM}}, \code{\link{AzureStopVM}},
  \code{\link{AzureVMStatus}}
}

