% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{AzureSM: AzureDeleteVM}
\alias{AzureDeleteVM}
\alias{AzureSM: AzureDeleteVM}
\title{Delete a Virtual Machine}
\usage{
AzureDeleteVM(AzureActiveContext, ResourceGroup, VMName, SubscriptionID,
  AzToken, Mode = "Sync", Verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{VMName}{Virtual Machine Name}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{Verbose}{Print Tracing information (Default False)}

\item{Token}{Token Object (or use AzureActiveContext)}
}

