% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureDeleteVM}
\alias{azureDeleteVM}
\title{Delete a Virtual Machine.}
\usage{
azureDeleteVM(azureActiveContext, resourceGroup, vmName, subscriptionID,
  mode = "Sync", verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{resourceGroup}{Name of the resource group}

\item{vmName}{Virtual Machine name}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{mode}{Wait for operation to complete 'Sync' (Default)}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Delete a Virtual Machine.
}
\seealso{
Other Virtual machine functions: \code{\link{azureGetAllVMstatus}},
  \code{\link{azureListScaleSetNetwork}},
  \code{\link{azureListScaleSetVM}},
  \code{\link{azureListScaleSets}},
  \code{\link{azureListVM}}, \code{\link{azureStartVM}},
  \code{\link{azureStopVM}}, \code{\link{azureVMStatus}}
}
