% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureDeleteVM}
\alias{azureDeleteVM}
\title{Delete a Virtual Machine.}
\usage{
azureDeleteVM(azureActiveContext, resourceGroup, vmName, subscriptionID,
  azToken, mode = "Sync", verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{resourceGroup}{Name of the resource group}

\item{vmName}{Virtual Machine name}

\item{subscriptionID}{Set the subscriptionID.  This is obtained automatically by \code{\link{azureAuthenticate}} when only a single subscriptionID is available via Active Directory}

\item{azToken}{Azure authentication token, obtained by \code{\link{azureAuthenticate}}}

\item{mode}{Wait for operation to complete 'Sync' (Default)}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Delete a Virtual Machine.
}
\seealso{
Other Virtual machine functions: \code{\link{azureListVM}},
  \code{\link{azureStartVM}}, \code{\link{azureStopVM}},
  \code{\link{azureVMStatus}}
}

