% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{AzureDeployStatus}
\alias{AzureDeployStatus}
\title{Check Template DeployStatus.}
\usage{
AzureDeployStatus(AzureActiveContext, DeplName, ResourceGroup, SubscriptionID,
  AzToken, verbose = FALSE)
}
\description{
Check Template DeployStatus.
}

