% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{azureDeployStatus}
\alias{azureDeployStatus}
\title{Check template deployment Status.}
\usage{
azureDeployStatus(azureActiveContext, deplname, resourceGroup, subscriptionID,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}}

\item{deplname}{Deployment name}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Check template deployment Status.
}
\seealso{
Other Template functions: \code{\link{azureCancelDeploy}},
  \code{\link{azureDeleteDeploy}},
  \code{\link{azureDeployTemplate}}
}
