% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{azureDeployTemplate}
\alias{azureDeployTemplate}
\title{Deploy an Azure Resource Manager Template.}
\usage{
azureDeployTemplate(azureActiveContext, deplname, templateURL, paramURL,
  templateJSON, paramJSON, mode = "Sync", resourceGroup, subscriptionID,
  azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{deplname}{deplname}

\item{templateURL}{templateURL}

\item{paramURL}{paramURL}

\item{templateJSON}{templateJSON}

\item{paramJSON}{paramJSON}

\item{mode}{'Sync' or 'Async'}

\item{resourceGroup}{resourceGroup Object}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Deploy an Azure Resource Manager Template.
}
\seealso{
Other Template functions: \code{\link{azureCancelDeploy}},
  \code{\link{azureDeleteDeploy}},
  \code{\link{azureDeployStatus}}
}

