% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{azureDeployTemplate}
\alias{azureDeployTemplate}
\title{Deploy an Azure Resource Manager Template.}
\usage{
azureDeployTemplate(azureActiveContext, deplname, templateURL, paramURL,
  templateJSON, paramJSON, mode = "Sync", resourceGroup, subscriptionID,
  azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{deplname}{deplname}

\item{templateURL}{templateURL}

\item{paramURL}{paramURL}

\item{templateJSON}{templateJSON}

\item{paramJSON}{paramJSON}

\item{mode}{'Sync' or 'Async'}

\item{resourceGroup}{Set the name of the Resource Group}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Deploy an Azure Resource Manager Template.
}
\seealso{
Other Template functions: \code{\link{azureCancelDeploy}},
  \code{\link{azureDeleteDeploy}},
  \code{\link{azureDeployStatus}}
}

