% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{AzureDeployTemplate}
\alias{AzureDeployTemplate}
\title{Deploy an Azure Resource Manager Template.}
\usage{
AzureDeployTemplate(AzureActiveContext, DeplName, TemplateURL, ParamURL,
  TemplateJSON, ParamJSON, Mode = "Sync", ResourceGroup, SubscriptionID,
  AzToken, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{DeplName}{DeplName}

\item{TemplateURL}{TemplateURL}

\item{ParamURL}{ParamURL}

\item{TemplateJSON}{TemplateJSON}

\item{ParamJSON}{ParamJSON}

\item{Mode}{'Sync' or 'Async'}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Deploy an Azure Resource Manager Template.
}
\seealso{
Other Template functions: \code{\link{AzureCancelDeploy}},
  \code{\link{AzureDeleteDeploy}},
  \code{\link{AzureDeployStatus}}
}

