% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{AzureSM: AzureDeployTemplate}
\alias{AzureDeployTemplate}
\alias{AzureSM: AzureDeployTemplate}
\title{Submit a Azure Resource ManagerTemplate}
\usage{
AzureDeployTemplate(AzureActiveContext, DeplName, TemplateURL, ParamURL,
  TemplateJSON, ParamJSON, Mode = "Sync", ResourceGroup, SubscriptionID,
  AzToken, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{DeplName}{DeplName}

\item{TemplateURL}{TemplateURL}

\item{ParamURL}{ParamURL}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{verbose}{Print Tracing information (Default False)}

\item{Token}{Token Object (or use AzureActiveContext)}
}

