% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{azureDeployTemplate}
\alias{azureDeployTemplate}
\title{Deploy a resource from an Azure Resource Manager (ARM) template.}
\usage{
azureDeployTemplate(azureActiveContext, deplname, templateURL, paramURL,
  templateJSON, paramJSON, mode = c("Sync", "Async"), resourceGroup,
  subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{deplname}{Deployment name}

\item{templateURL}{URL that contains the ARM template to deploy. You must specify either \code{templateURL} OR \code{templateJSON}}

\item{paramURL}{URL that contains the template parameters. You must specify either \code{paramULR} OR \code{paramJSON}}

\item{templateJSON}{character vector that contains the ARM template to deploy. You must specify either \code{templateJSON} OR \code{templateURL}}

\item{paramJSON}{paramJSON}

\item{mode}{Provisioning mode, "Sync" or "Async". Use "Async" to immediately return to R session after submission of request}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{If TRUE, prints verbose messages}
}
\description{
Deploy a resource from a template. See https://github.com/Azure/azure-quickstart-templates
}
\seealso{
Other Template functions: \code{\link{azureCancelDeploy}},
  \code{\link{azureDeleteDeploy}},
  \code{\link{azureDeployStatus}}
}
