% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{AzureDeployTemplate}
\alias{AzureDeployTemplate}
\title{Submit a Azure Resource ManagerTemplate.}
\usage{
AzureDeployTemplate(AzureActiveContext, DeplName, TemplateURL, ParamURL,
  TemplateJSON, ParamJSON, Mode = "Sync", ResourceGroup, SubscriptionID,
  AzToken, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Submit a Azure Resource ManagerTemplate.
}
\seealso{
Other Template functions: \code{\link{AzureCancelDeploy}},
  \code{\link{AzureDeleteDeploy}},
  \code{\link{AzureDeployStatus}}
}

