% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureTemplate.R
\name{azureDeployTemplate}
\alias{azureDeployTemplate}
\title{Deploy an Azure Resource Manager Template.}
\usage{
azureDeployTemplate(azureActiveContext, deplname, templateURL, paramURL,
  templateJSON, paramJSON, mode = "Sync", resourceGroup, subscriptionID,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{deplname}{deplname}

\item{templateURL}{templateURL}

\item{paramURL}{paramURL}

\item{templateJSON}{templateJSON}

\item{paramJSON}{paramJSON}

\item{mode}{Provisioning mode, "Sync" or "Async". Use "Async" to immediately return to R session after submission of request}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Set the subscriptionID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Deploy an Azure Resource Manager Template.
}
\seealso{
Other Template functions: \code{\link{azureCancelDeploy}},
  \code{\link{azureDeleteDeploy}},
  \code{\link{azureDeployStatus}}
}
