% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureCost.R
\name{azureExpenseCalculator}
\alias{azureExpenseCalculator}
\title{Calculate cost of using a specific instance of Azure for certain period.}
\usage{
azureExpenseCalculator(azureActiveContext, instance = "", timeStart, timeEnd,
  granularity, currency, locale, offerId, region, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}}

\item{instance}{Instance name that one would like to check expe
nse. It is by default empty, which returns data consumption for
all instances under subscription.}

\item{timeStart}{Start time.}

\item{timeEnd}{End time.}

\item{granularity}{Aggregation granularity. Can be either "Daily" or
"Hourly".}

\item{currency}{Currency in which price rating is measured.}

\item{locale}{Locality information of subscription.}

\item{offerId}{Offer ID of the subscription. Detailed information can be
found at https://azure.microsoft.com/en-us/support/legal/offer-details/}

\item{region}{region information about the subscription.}
}
\value{
Total cost measured in the given currency of the specified Azure
instance in the period.
}
\description{
Calculate cost of using a specific instance of Azure for certain period.
}
\note{
Note if difference between \code{timeStart} and \code{timeEnd} is
less than the finest granularity, e.g., "Hourly" (we notice this is a
usual case when one needs to be aware of the charges of a job that takes
less than an hour), the expense will be estimated based solely on computation
hour. That is, the total expense is the multiplication of computation hour
and pricing rate of the requested instance.
}
