% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzureSM: AzureGetBlob}
\alias{AzureGetBlob}
\alias{AzureSM: AzureGetBlob}
\title{Get contents from a specifed Storage Blob}
\usage{
AzureGetBlob(AzureActiveContext, StorageAccount, StorageKey, Container, Blob,
  Type = "Text", ResourceGroup, SubscriptionID, AzToken, Verbose = TRUE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{StorageAccount}{StorageAccount}

\item{StorageKey}{StorageKey}

\item{Container}{Storage Container}

\item{Blob}{Blob Name}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{Verbose}{Print Tracing information (Default False)}

\item{Blob}{Type ("Text" or "Raw")}

\item{Token}{Token Object (or use AzureActiveContext)}
}

