% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azureGetBlob}
\alias{azureGetBlob}
\title{Get contents from a specifed Storage blob.}
\usage{
azureGetBlob(azureActiveContext, blob, directory, type = "text",
  storageAccount, storageKey, container, resourceGroup, subscriptionID, azToken,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{blob}{blob Object}

\item{directory}{Set directory to list}

\item{type}{"Text" or "Raw"}

\item{storageAccount}{storageAccount Object}

\item{storageKey}{storageKey Object}

\item{container}{container Object}

\item{resourceGroup}{resourceGroup Object}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Get contents from a specifed Storage blob.
}
\seealso{
Other blob store functions: \code{\link{azureBlobCD}},
  \code{\link{azureBlobFind}},
  \code{\link{azureListStorageBlobs}},
  \code{\link{azurePutBlob}}, \code{\link{azureblobLS}}
}

