% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzureSM: AzureGetBlob}
\alias{AzureGetBlob}
\alias{AzureSM: AzureGetBlob}
\title{Get contents from a specifed Storage Blob}
\usage{
AzureGetBlob(AzureActiveContext, Blob, Directory, Type = "text",
  StorageAccount, StorageKey, Container, ResourceGroup, SubscriptionID, AzToken,
  verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{Blob}{Blob Directory (leave for current directory)}

\item{StorageAccount}{StorageAccount}

\item{StorageKey}{StorageKey}

\item{Container}{Storage Container}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{verbose}{Print Tracing information (Default False)}

\item{Blob}{Blob Name}

\item{Blob}{Type ("Text" or "Raw")}

\item{Token}{Token Object (or use AzureActiveContext)}
}

