% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azureGetblob}
\alias{azureGetblob}
\title{Get contents from a specifed Storage blob.}
\usage{
azureGetblob(azureActiveContext, blob, directory, type = "text",
  storageAccount, storageKey, container, resourceGroup, subscriptionID, azToken,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{blob}{Set the blob name}

\item{directory}{Set directory to list}

\item{type}{"Text" or "Raw"}

\item{storageAccount}{Set the name of the azure storage account}

\item{storageKey}{Set the Storage Key associated with storage account}

\item{container}{Set the storage container}

\item{resourceGroup}{Set the name of the Resource Group}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Get contents from a specifed Storage blob.
}
\seealso{
Other blob store functions: \code{\link{azureListSAblobs}},
  \code{\link{azurePutblob}}, \code{\link{azureblobCD}},
  \code{\link{azureblobFind}}, \code{\link{azureblobLS}}
}

