% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{AzureSM: AzureHDIConf}
\alias{AzureHDIConf}
\alias{AzureSM: AzureHDIConf}
\title{Get Configuration Information for a specified ClusterName}
\usage{
AzureHDIConf(AzureActiveContext, ClusterName, ResourceGroup, SUBID, ATI, Name,
  Type, Location, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{ClusterName}{ResourceGroup Object (or use AzureActiveContext)}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{verbose}{Print Tracing information (Default False)}

\item{Token}{Token Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}
}
\value{
Returns Dataframe of HDInsight Clusters information
}

