% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{AzureHDIConf}
\alias{AzureHDIConf}
\title{Get Configuration Information for a specified ClusterName.}
\usage{
AzureHDIConf(AzureActiveContext, ClusterName, ResourceGroup, SUBID, ATI, Name,
  Type, Location, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ClusterName}{ClusterName Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of HDInsight Clusters information
}
\description{
Get Configuration Information for a specified ClusterName.
}
\seealso{
Other HDInsight: \code{\link{AzureCreateHDI}},
  \code{\link{AzureDeleteHDI}}, \code{\link{AzureListHDI}},
  \code{\link{AzureResizeHDI}},
  \code{\link{AzureRunScriptAction}},
  \code{\link{AzureScriptActionHistory}}
}

