% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{azureHDIConf}
\alias{azureHDIConf}
\title{Get configuration information for a specified cluster name.}
\usage{
azureHDIConf(azureActiveContext, clustername, resourceGroup, subscriptionID,
  name, type, location, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{clustername}{Cluster name, used for HDI and Spark clusters. See \code{\link[=azureCreateHDI]{azureCreateHDI()}}}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{name}{filter by resource name}

\item{type}{filter by resource type}

\item{location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{verbose}{If TRUE, prints verbose messages}
}
\value{
Returns Dataframe of HDInsight Clusters information
}
\description{
Get configuration information for a specified cluster name.
}
\seealso{
Other HDInsight functions: \code{\link{azureCreateHDI}},
  \code{\link{azureDeleteHDI}}, \code{\link{azureListHDI}},
  \code{\link{azureResizeHDI}},
  \code{\link{azureRunScriptAction}},
  \code{\link{azureScriptActionHistory}}
}
