% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{AzureHDIConf}
\alias{AzureHDIConf}
\title{Get Configuration Information for a specified ClusterName.}
\usage{
AzureHDIConf(AzureActiveContext, ClusterName, ResourceGroup, SubscriptionID,
  AzToken, Name, Type, Location, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ClusterName}{ClusterName Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object}

\item{Name}{Name}

\item{Type}{Type}

\item{Location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of HDInsight Clusters information
}
\description{
Get Configuration Information for a specified ClusterName.
}
\seealso{
Other HDInsight functions: \code{\link{AzureCreateHDI}},
  \code{\link{AzureDeleteHDI}}, \code{\link{AzureListHDI}},
  \code{\link{AzureResizeHDI}},
  \code{\link{AzureRunScriptAction}},
  \code{\link{AzureScriptActionHistory}}
}

