% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHive.R
\name{azureHiveSQL}
\alias{azureHiveSQL}
\title{Submit SQL command to Hive Service.}
\usage{
azureHiveSQL(azureActiveContext, CMD, clustername, hdiAdmin, hdiPassword,
  path = "wasb:///tmp/", verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{CMD}{SQl COmmand String}

\item{clustername}{clustername Object}

\item{hdiAdmin}{hdiAdmin Object}

\item{hdiPassword}{hdiPassword  Object}

\item{path}{path}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Submit SQL command to Hive Service.
}
\seealso{
Other Hive functions: \code{\link{azureHiveStatus}}
}

