% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHive.R
\name{azureHiveStatus}
\alias{azureHiveStatus}
\title{Get Status of a HDI Hive Service / version.}
\usage{
azureHiveStatus(azureActiveContext, clustername, hdiAdmin, hdiPassword,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by `AzureSMR`, created by [createAzureContext()]}

\item{clustername}{Cluster name, used for HDI and Spark clusters. See [azureCreateHDI()]}

\item{hdiAdmin}{HDInsight admin username. See [azureCreateHDI()]}

\item{hdiPassword}{HDInsight admin password. See [azureCreateHDI()]}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Get Status of a HDI Hive Service / version.
}
\seealso{
Other Hive functions: \code{\link{azureHiveSQL}}
}

