% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHive.R
\name{AzureSM: AzureHiveStatus}
\alias{AzureHiveStatus}
\alias{AzureSM: AzureHiveStatus}
\title{Get Status of a HDI Hive Service/Version}
\usage{
AzureHiveStatus(AzureActiveContext, ClusterName, HDIAdmin, HDIPassword,
  verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{ClusterName}{ClusterName}

\item{HDIAdmin}{HDIAdmin - HDinsight Administrator Name}

\item{HDIPassword}{HDinsight Administrator Name}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{Verbose}{Print Tracing information (Default False)}
}

