% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHive.R
\name{azureHiveStatus}
\alias{azureHiveStatus}
\title{Get Status of a HDI Hive Service/version.}
\usage{
azureHiveStatus(azureActiveContext, clustername, hdiAdmin, hdiPassword,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{clustername}{clustername Object}

\item{hdiAdmin}{hdiAdmin Object}

\item{hdiPassword}{hdiPassword  Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Get Status of a HDI Hive Service/version.
}
\seealso{
Other Hive functions: \code{\link{azureHiveSQL}}
}

