% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHive.R
\name{AzureHiveStatus}
\alias{AzureHiveStatus}
\title{Get Status of a HDI Hive Service/Version.}
\usage{
AzureHiveStatus(AzureActiveContext, ClusterName, HDIAdmin, HDIPassword,
  verbose = FALSE)
}
\description{
Get Status of a HDI Hive Service/Version.
}

