% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{AzureSM: AzureListHDI}
\alias{AzureListHDI}
\alias{AzureSM: AzureListHDI}
\title{Get all HDInsight Clusters in default Subscription or details for a specified ClusterName}
\usage{
AzureListHDI(AzureActiveContext, ResourceGroup, ClusterName = "*", SUBID, ATI,
  Name, Type, Location, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{ClusterName}{ResourceGroup Object (or use AzureActiveContext)}

\item{Token}{Token Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{Verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of HDInsight Clusters
}

