% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{AzureListHDI}
\alias{AzureListHDI}
\title{Get all HDInsight Clusters in default Subscription or details for a specified ClusterName.}
\usage{
AzureListHDI(AzureActiveContext, ResourceGroup, ClusterName = "*",
  SubscriptionID, ATI, Name, Type, Location, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{ClusterName}{ResourceGroup Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of HDInsight Clusters
}
\description{
Get all HDInsight Clusters in default Subscription or details for a specified ClusterName.
}
\seealso{
Other HDInsight: \code{\link{AzureCreateHDI}},
  \code{\link{AzureDeleteHDI}}, \code{\link{AzureHDIConf}},
  \code{\link{AzureResizeHDI}},
  \code{\link{AzureRunScriptAction}},
  \code{\link{AzureScriptActionHistory}}
}

