% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{azureListRG}
\alias{azureListRG}
\title{Get all Resource Groups in default Subscription.}
\usage{
azureListRG(azureActiveContext, subscriptionID, azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of resourceGroups
}
\description{
Get all Resource Groups in default Subscription.
}
\seealso{
Other Resource group functions: \code{\link{azureCreateResourceGroup}},
  \code{\link{azureDeleteResourceGroup}},
  \code{\link{azureListAllResources}},
  \code{\link{azureListSubscriptions}}
}

