% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{azureListRG}
\alias{azureListRG}
\title{Get all Resource Groups in default Subscription.}
\usage{
azureListRG(azureActiveContext, subscriptionID, azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of resourceGroups
}
\description{
Get all Resource Groups in default Subscription.
}
\seealso{
Other Resource group functions: \code{\link{azureCreateresourceGroup}},
  \code{\link{azureDeleteresourceGroup}},
  \code{\link{azureListAllResources}},
  \code{\link{azureListSubscriptions}}
}

