% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{AzureListRG}
\alias{AzureListRG}
\title{Get all Resource Groups in default Subscription.}
\usage{
AzureListRG(AzureActiveContext, SubscriptionID, AzToken, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of ResourceGroups
}
\description{
Get all Resource Groups in default Subscription.
}
\seealso{
Other Resource group functions: \code{\link{AzureCreateResourceGroup}},
  \code{\link{AzureDeleteResourceGroup}},
  \code{\link{AzureListAllResources}},
  \code{\link{AzureListSubscriptions}}
}

