% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{AzureSM: AzureListRG}
\alias{AzureListRG}
\alias{AzureSM: AzureListRG}
\title{Get all Resource Groups in default Subscription}
\usage{
AzureListRG(AzureActiveContext, SUBID, AT, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{Token}{Token Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{Verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of ResourceGroups
}

