% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{AzureListRG}
\alias{AzureListRG}
\title{Get all Resource Groups in default Subscription.}
\usage{
AzureListRG(AzureActiveContext, subscriptionID, AT, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of ResourceGroups
}
\description{
Get all Resource Groups in default Subscription.
}
\seealso{
Other Resources: \code{\link{AzureAuthenticate}},
  \code{\link{AzureCheckToken}},
  \code{\link{AzureCreateResourceGroup}},
  \code{\link{AzureDeleteResourceGroup}},
  \code{\link{AzureListAllRecources}},
  \code{\link{AzureListSubscriptions}}

Other Resources: \code{\link{AzureAuthenticate}},
  \code{\link{AzureCheckToken}},
  \code{\link{AzureCreateResourceGroup}},
  \code{\link{AzureDeleteResourceGroup}},
  \code{\link{AzureListAllRecources}},
  \code{\link{AzureListSubscriptions}}
}

