% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzureListSA}
\alias{AzureListSA}
\title{List Storage accounts.}
\usage{
AzureListSA(AzureActiveContext, ResourceGroup, SubscriptionID, AzToken,
  verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object (or use AzureActiveContext)}
}
\description{
List Storage accounts.
}
\seealso{
Other Storage account: \code{\link{AzureBlobCD}},
  \code{\link{AzureBlobFind}}, \code{\link{AzureBlobLS}},
  \code{\link{AzureCreateSAContainer}},
  \code{\link{AzureCreateStorageAccount}},
  \code{\link{AzureDeleteSAContainer}},
  \code{\link{AzureDeleteStorageAccount}},
  \code{\link{AzureGetBlob}},
  \code{\link{AzureListSABlobs}},
  \code{\link{AzureListSAContainers}},
  \code{\link{AzurePutBlob}}, \code{\link{AzureSAGetKey}}
}

