% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzureSM: AzureListSA}
\alias{AzureListSA}
\alias{AzureSM: AzureListSA}
\title{List Storage accounts}
\usage{
AzureListSA(AzureActiveContext, ResourceGroup, SubscriptionID, AzToken,
  verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{AzToken}{Token Object (or use AzureActiveContext)}

\item{verbose}{Print Tracing information (Default False)}
}

