% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureStorageAccount.R
\name{azureListSA}
\alias{azureListSA}
\title{List Storage accounts.}
\usage{
azureListSA(azureActiveContext, resourceGroup, subscriptionID, azToken,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{resourceGroup}{resourceGroup Object}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List Storage accounts.
}
\seealso{
Other Storage account functions: \code{\link{azureCreatestorageAccount}},
  \code{\link{azureDeletestorageAccount}},
  \code{\link{azureSAGetKey}}
}

