% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureStorageAccount.R
\name{AzureListSA}
\alias{AzureListSA}
\title{List Storage accounts.}
\usage{
AzureListSA(AzureActiveContext, ResourceGroup, SubscriptionID, AzToken,
  verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List Storage accounts.
}
\seealso{
Other Storage account functions: \code{\link{AzureCreateStorageAccount}},
  \code{\link{AzureDeleteStorageAccount}},
  \code{\link{AzureSAGetKey}}
}

