% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureStorageAccount.R
\name{azureListSA}
\alias{azureListSA}
\title{List storage accounts.}
\usage{
azureListSA(azureActiveContext, resourceGroup, subscriptionID,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List storage accounts.
}
\seealso{
Other Storage account functions: \code{\link{azureCreateStorageAccount}},
  \code{\link{azureDeletestorageAccount}},
  \code{\link{azureSAGetKey}}
}
