% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzureSM: AzureListSABlobs}
\alias{AzureListSABlobs}
\alias{AzureSM: AzureListSABlobs}
\title{List Storage Containers for Specified Storage Account}
\usage{
AzureListSABlobs(AzureActiveContext, StorageAccount, StorageKey, Container,
  ResourceGroup, SubscriptionID, AzToken)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{StorageAccount}{StorageAccount}

\item{StorageKey}{StorageKey}

\item{Container}{Storage Container}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{AzToken}{Token Object (or use AzureActiveContext)}

\item{Verbose}{Print Tracing information (Default False)}
}

