% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azureListSAblobs}
\alias{azureListSAblobs}
\title{List Storage containers for Specified Storage Account.}
\usage{
azureListSAblobs(azureActiveContext, storageAccount, storageKey, container,
  resourceGroup, subscriptionID, azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{storageAccount}{storageAccount Object}

\item{storageKey}{storageKey Object}

\item{container}{container Object}

\item{resourceGroup}{resourceGroup Object}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List Storage containers for Specified Storage Account.
}
\seealso{
Other blob store functions: \code{\link{azureGetblob}},
  \code{\link{azurePutblob}}, \code{\link{azureblobCD}},
  \code{\link{azureblobFind}}, \code{\link{azureblobLS}}
}

