% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzureSM: AzureListSAContainers}
\alias{AzureListSAContainers}
\alias{AzureSM: AzureListSAContainers}
\title{List Storage Containers for Specified Storage Account}
\usage{
AzureListSAContainers(AzureActiveContext, StorageAccount, StorageKey,
  ResourceGroup, AzToken, SubscriptionID, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{StorageAccount}{StorageAccount}

\item{StorageKey}{StorageKey}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{Token}{Token Object (or use AzureActiveContext)}

\item{Verbose}{Print Tracing information (Default False)}
}

