% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureListScaleSetVM}
\alias{azureListScaleSetVM}
\title{List VMs within a scale set}
\usage{
azureListScaleSetVM(azureActiveContext, scaleSet, resourceGroup, location,
  subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{scaleSet}{name of the scale refer to (azureListScaleSets)}

\item{resourceGroup}{Name of the resource group}

\item{location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{subscriptionID}{Set the subscriptionID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List VMs within a scale set
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureListScaleSetNetwork}},
  \code{\link{azureListScaleSets}},
  \code{\link{azureListVM}}, \code{\link{azureStartVM}},
  \code{\link{azureStopVM}}, \code{\link{azureVMStatus}}
}
