% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azureListStorageBlobs}
\alias{azureListStorageBlobs}
\title{List storage blobs for specified storage account.}
\usage{
azureListStorageBlobs(azureActiveContext, storageAccount, storageKey, container,
  resourceGroup, maxresults, prefix, delimiter, marker, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{storageAccount}{Name of the azure storage account}

\item{storageKey}{Storage key associated with storage account}

\item{container}{Storage container name. See \code{azureListStorageContainers}}

\item{resourceGroup}{Name of the resource group}

\item{maxresults}{Optional. Specifies the maximum number of blobs to return, including all BlobPrefix elements. If the request does not specify maxresults or specifies a value greater than 5,000, the server will return up to 5,000 items.  Setting maxresults to a value less than or equal to zero results in error response code 400 (Bad Request).}

\item{prefix}{Optional. Filters the results to return only blobs whose names begin with the specified prefix.}

\item{delimiter}{Optional. When the request includes this parameter, the operation returns a BlobPrefix element in the response body that acts as a placeholder for all blobs whose names begin with the same substring up to the appearance of the delimiter character. The delimiter may be a single character or a string.}

\item{marker}{Optional. A string value that identifies the portion of the list to be returned with the next list operation. The operation returns a marker value within the response body if the list returned was not complete. The marker value may then be used in a subsequent call to request the next set of list items.  The marker value is opaque to the client.}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns a data frame. This data frame has an attribute called "marker" that can be used with the "marker" argument to return the next set of values.
}
\description{
List storage blobs for specified storage account.
}
\seealso{
Other blob store functions: \code{\link{azureBlobCD}},
  \code{\link{azureBlobFind}}, \code{\link{azureBlobLS}},
  \code{\link{azureDeleteBlob}},
  \code{\link{azureGetBlob}}, \code{\link{azurePutBlob}}
}
