% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azureListStorageBlobs}
\alias{azureListStorageBlobs}
\title{List Storage containers for Specified Storage Account.}
\usage{
azureListStorageBlobs(azureActiveContext, storageAccount, storageKey, container,
  resourceGroup, subscriptionID, azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{storageAccount}{storageAccount Object}

\item{storageKey}{storageKey Object}

\item{container}{container Object}

\item{resourceGroup}{resourceGroup Object}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List Storage containers for Specified Storage Account.
}
\seealso{
Other blob store functions: \code{\link{azureBlobCD}},
  \code{\link{azureBlobFind}}, \code{\link{azureGetBlob}},
  \code{\link{azurePutBlob}}, \code{\link{azureblobLS}}
}

