% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContainer.R
\name{azureListStorageContainers}
\alias{azureListStorageContainers}
\title{List Storage containers for Specified Storage Account.}
\usage{
azureListStorageContainers(azureActiveContext, storageAccount, storageKey,
  resourceGroup, azToken, subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{storageAccount}{Name of the azure storage account}

\item{storageKey}{Storage key associated with storage account}

\item{resourceGroup}{Name of the resource group}

\item{azToken}{Azure authentication token, obtained by \code{azureAuthenticate}}

\item{subscriptionID}{Set the subscriptionID.  This is obtained automatically by \code{azureAuthenticate} when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List Storage containers for Specified Storage Account.
}
\seealso{
Other container functions: \code{\link{azureCreateStorageContainer}},
  \code{\link{azureDeleteStorageContainer}}
}
