% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContainer.R
\name{azureListStorageContainers}
\alias{azureListStorageContainers}
\title{List Storage containers for Specified Storage Account.}
\usage{
azureListStorageContainers(azureActiveContext, storageAccount, storageKey,
  resourceGroup, subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{storageAccount}{Name of the azure storage account. Storage account names must be between 3 and 24 characters in length and use numbers and lower-case letters only.}

\item{storageKey}{Storage key associated with storage account}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{If TRUE, prints verbose messages}
}
\description{
List Storage containers for Specified Storage Account.
}
\seealso{
Other Container functions: \code{\link{azureCreateStorageContainer}},
  \code{\link{azureDeleteStorageContainer}}
}
