% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContainer.R
\name{azureListStorageContainers}
\alias{azureListStorageContainers}
\title{List Storage containers for Specified Storage Account.}
\usage{
azureListStorageContainers(azureActiveContext, storageAccount, storageKey,
  resourceGroup, azToken, subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{storageAccount}{storageAccount Object}

\item{storageKey}{storageKey Object}

\item{resourceGroup}{resourceGroup Object}

\item{azToken}{Token Object}

\item{subscriptionID}{subscriptionID Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List Storage containers for Specified Storage Account.
}
\seealso{
Other container functions: \code{\link{azureCreateStorageContainer}},
  \code{\link{azureDeleteStorageContainer}}
}

