% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{AzureListSubscriptions}
\alias{AzureListSubscriptions}
\title{Get available Subscriptions.}
\usage{
AzureListSubscriptions(AzureActiveContext, AzToken, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{AzToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of SubscriptionID sets AzureContext SubscriptionID
}
\description{
Get available Subscriptions.
}
\seealso{
Other Resource group functions: \code{\link{AzureCreateResourceGroup}},
  \code{\link{AzureDeleteResourceGroup}},
  \code{\link{AzureListAllResources}},
  \code{\link{AzureListRG}}
}

