% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{AzureListSubscriptions}
\alias{AzureListSubscriptions}
\title{Get available Subscriptions.}
\usage{
AzureListSubscriptions(AzureActiveContext, ATI, verbose = FALSE)
}
\arguments{
\item{Azure}{Context Object}

\item{Token}{Token Object (or use AzureActiveContext)}

\item{Verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of SubscriptionID sets AzureContext SubscriptionID
}
\description{
Get available Subscriptions
}
\seealso{
Other Resources: \code{\link{AzureAuthenticate}},
  \code{\link{AzureCheckToken}},
  \code{\link{AzureCreateResourceGroup}},
  \code{\link{AzureDeleteResourceGroup}},
  \code{\link{AzureListAllRecources}},
  \code{\link{AzureListRG}}
}

