% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{azureListSubscriptions}
\alias{azureListSubscriptions}
\title{Get available Subscriptions.}
\usage{
azureListSubscriptions(azureActiveContext, azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{azToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of subscriptionID sets AzureContext subscriptionID
}
\description{
Get available Subscriptions.
}
\seealso{
Other Resource group functions: \code{\link{azureCreateresourceGroup}},
  \code{\link{azureDeleteresourceGroup}},
  \code{\link{azureListAllResources}},
  \code{\link{azureListRG}}
}

