% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureResources.R
\name{azureListSubscriptions}
\alias{azureListSubscriptions}
\title{Get available Subscriptions.}
\usage{
azureListSubscriptions(azureActiveContext, azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{azToken}{Azure authentication token, obtained by \code{\link{azureAuthenticate}}}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of subscriptionID sets AzureContext subscriptionID
}
\description{
Get available Subscriptions.
}
\seealso{
Other Resource group functions: \code{\link{azureCreateResourceGroup}},
  \code{\link{azureDeleteResourceGroup}},
  \code{\link{azureListAllResources}},
  \code{\link{azureListRG}}
}
