% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureListVM}
\alias{azureListVM}
\title{List VMs in a Subscription.}
\usage{
azureListVM(azureActiveContext, resourceGroup, location, subscriptionID,
  azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{resourceGroup}{Set the name of the Resource Group}

\item{location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List VMs in a Subscription.
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureStartVM}}, \code{\link{azureStopVM}},
  \code{\link{azureVMStatus}}
}

