% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{AzureSM: AzureListVM}
\alias{AzureListVM}
\alias{AzureSM: AzureListVM}
\title{List VMs in a Subscription}
\usage{
AzureListVM(AzureActiveContext, ResourceGroup, Location, SubscriptionID,
  AzToken, Verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{Location}{Azure Resource Location}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{Verbose}{Print Tracing information (Default False)}

\item{Token}{Token Object (or use AzureActiveContext)}
}

