% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{AzureListVM}
\alias{AzureListVM}
\title{List VMs in a Subscription.}
\usage{
AzureListVM(AzureActiveContext, ResourceGroup, Location, SubscriptionID,
  AzToken, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{Location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{AzToken}{Token Object (or use AzureActiveContext)}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List VMs in a Subscription.
}
\seealso{
Other Virtual Machine: \code{\link{AzureDeleteVM}},
  \code{\link{AzureStartVM}}, \code{\link{AzureStopVM}},
  \code{\link{AzureVMStatus}}
}

