% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureListVM}
\alias{azureListVM}
\title{List VMs in a Subscription.}
\usage{
azureListVM(azureActiveContext, resourceGroup, location, subscriptionID,
  azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{resourceGroup}{Name of the resource group}

\item{location}{Azure region, e.g. 'westeurope' or 'southcentralus'}

\item{subscriptionID}{Set the subscriptionID.  This is obtained automatically by \code{azureAuthenticate} when only a single subscriptionID is available via Active Directory}

\item{azToken}{Azure authentication token, obtained by \code{azureAuthenticate}}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List VMs in a Subscription.
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureListScaleSetNetwork}},
  \code{\link{azureListScaleSetVM}},
  \code{\link{azureListScaleSets}},
  \code{\link{azureStartVM}}, \code{\link{azureStopVM}},
  \code{\link{azureVMStatus}}
}
