% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureCost.R
\name{azurePricingRates}
\alias{azurePricingRates}
\title{Get pricing details of resources under a subscription.}
\usage{
azurePricingRates(azureActiveContext, currency, locale, offerId, region,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}}

\item{currency}{Currency in which price rating is measured.}

\item{locale}{Locality information of subscription.}

\item{offerId}{Offer ID of the subscription. For more information see https://azure.microsoft.com/en-us/support/legal/offer-details/}

\item{region}{region information about the subscription.}
}
\description{
Get pricing details of resources under a subscription.
}
\note{
The pricing rates function wraps API calls to Azure RateCard and currently the API supports only the Pay-As-You-Go offer scheme.
}
\seealso{
Other Cost functions: \code{\link{azureDataConsumption}},
  \code{\link{azureExpenseCalculator}}
}
