% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azurePutblob}
\alias{azurePutblob}
\title{Write contents to a specifed Storage blob.}
\usage{
azurePutblob(azureActiveContext, blob, contents = "", file = "", directory,
  storageAccount, storageKey, container, resourceGroup, subscriptionID, azToken,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{blob}{blob Object}

\item{contents}{- Object to Store or Value}

\item{file}{- Local filename to Store in Azure blob}

\item{directory}{Set directory to list}

\item{storageAccount}{storageAccount Object}

\item{storageKey}{storageKey Object}

\item{container}{container Object}

\item{resourceGroup}{resourceGroup Object}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Write contents to a specifed Storage blob.
}
\seealso{
Other blob store functions: \code{\link{azureGetblob}},
  \code{\link{azureListSAblobs}},
  \code{\link{azureblobCD}}, \code{\link{azureblobFind}},
  \code{\link{azureblobLS}}
}

