% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzurePutBlob}
\alias{AzurePutBlob}
\title{Write contents to a specifed Storage Blob.}
\usage{
AzurePutBlob(AzureActiveContext, Blob, Contents = "", File = "", Directory,
  StorageAccount, StorageKey, Container, ResourceGroup, SubscriptionID, AzToken,
  verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{Blob}{Blob Object}

\item{Contents}{- Object to Store or Value}

\item{File}{- Local FileName to Store in Azure Blob}

\item{Directory}{Set Directory to list}

\item{StorageAccount}{StorageAccount Object}

\item{StorageKey}{StorageKey Object}

\item{Container}{Container Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Write contents to a specifed Storage Blob.
}
\seealso{
Other Blob store functions: \code{\link{AzureBlobCD}},
  \code{\link{AzureBlobFind}}, \code{\link{AzureBlobLS}},
  \code{\link{AzureGetBlob}},
  \code{\link{AzureListSABlobs}}
}

