% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzureSM: AzurePutBlob}
\alias{AzurePutBlob}
\alias{AzureSM: AzurePutBlob}
\title{Write contents to a specifed Storage Blob}
\usage{
AzurePutBlob(AzureActiveContext, StorageAccount, StorageKey, Container, Blob,
  ResourceGroup, Contents = "", File = "", SubscriptionID, AzToken,
  Verbose = TRUE)
}
\arguments{
\item{AzureActiveContext}{-Azure Context Object}

\item{StorageAccount}{- StorageAccount}

\item{StorageKey}{- StorageKey}

\item{Container}{- Storage Container}

\item{Blob}{- Blob Name}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{Contents}{- Object to Store or Value}

\item{File}{- Local FileName to Store in Azure Blob}

\item{SubscriptionID}{- SubscriptionID Object (or use AzureActiveContext)}

\item{Verbose}{- Print Tracing information (Default False)}

\item{Token}{- Token Object (or use AzureActiveContext)}
}

