% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{azurePutBlob}
\alias{azurePutBlob}
\title{Write contents to a specifed Storage blob.}
\usage{
azurePutBlob(azureActiveContext, blob, contents = "", file = "", directory,
  storageAccount, storageKey, container, resourceGroup, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{blob}{Blob name}

\item{contents}{- Object or value to store}

\item{file}{- Local filename to store in Azure blob}

\item{directory}{Blob store directory to list for content}

\item{storageAccount}{Name of the azure storage account}

\item{storageKey}{Storage key associated with storage account}

\item{container}{Storage container name. See \code{\link{azureListStorageContainers}}}

\item{resourceGroup}{Name of the resource group}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Write contents to a specifed Storage blob.
}
\seealso{
Other blob store functions: \code{\link{azureBlobCD}},
  \code{\link{azureBlobFind}}, \code{\link{azureBlobLS}},
  \code{\link{azureDeleteBlob}},
  \code{\link{azureGetBlob}},
  \code{\link{azureListStorageBlobs}}
}
