% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{azureResizeHDI}
\alias{azureResizeHDI}
\title{Resize a HDInsight Cluster role.}
\usage{
azureResizeHDI(azureActiveContext, clustername, role = "worker", size = 2,
  mode = "Sync", azToken, subscriptionID, resourceGroup, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{clustername}{Set the clustername}

\item{role}{role type: 'worker', 'head' or 'Edge'}

\item{size}{Desired size of role type}

\item{mode}{'Sync' or 'Async'}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{resourceGroup}{Set the name of the Resource Group}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Resize a HDInsight Cluster role.
}
\seealso{
Other HDInsight functions: \code{\link{azureCreateHDI}},
  \code{\link{azureDeleteHDI}}, \code{\link{azureHDIConf}},
  \code{\link{azureListHDI}},
  \code{\link{azureRunScriptAction}},
  \code{\link{azureScriptActionHistory}}
}

