% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{AzureResizeHDI}
\alias{AzureResizeHDI}
\title{Resize a HDInsight Cluster Role.}
\usage{
AzureResizeHDI(AzureActiveContext, ClusterName, Role = "worker", Size = 2,
  Mode = "Sync", AzToken, SubscriptionID, ResourceGroup, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ClusterName}{ClusterName Object}

\item{Role}{Role Type: 'worker', 'head' or 'Edge'}

\item{Size}{Desired size of Role Type}

\item{Mode}{'Sync' or 'Async'}

\item{AzToken}{Token Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Resize a HDInsight Cluster Role.
}
\seealso{
Other HDInsight functions: \code{\link{AzureCreateHDI}},
  \code{\link{AzureDeleteHDI}}, \code{\link{AzureHDIConf}},
  \code{\link{AzureListHDI}},
  \code{\link{AzureRunScriptAction}},
  \code{\link{AzureScriptActionHistory}}
}

