% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{azureResizeHDI}
\alias{azureResizeHDI}
\title{Resize a HDInsight Cluster role.}
\usage{
azureResizeHDI(azureActiveContext, clustername, role = "worker", size = 2,
  mode = "Sync", azToken, subscriptionID, resourceGroup, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{clustername}{clustername Object}

\item{role}{role type: 'worker', 'head' or 'Edge'}

\item{size}{Desired size of role type}

\item{mode}{'Sync' or 'Async'}

\item{azToken}{Token Object}

\item{subscriptionID}{subscriptionID Object}

\item{resourceGroup}{resourceGroup Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Resize a HDInsight Cluster role.
}
\seealso{
Other HDInsight functions: \code{\link{azureCreateHDI}},
  \code{\link{azureDeleteHDI}}, \code{\link{azureHDIConf}},
  \code{\link{azureListHDI}},
  \code{\link{azureRunScriptAction}},
  \code{\link{azureScriptActionHistory}}
}

