% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{azureResizeHDI}
\alias{azureResizeHDI}
\title{Resize a HDInsight cluster role.}
\usage{
azureResizeHDI(azureActiveContext, clustername, role = c("workernode",
  "headnode", "edgenode"), size = 2, mode = c("Sync", "Async"),
  subscriptionID, resourceGroup, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{clustername}{Cluster name, used for HDI and Spark clusters. See \code{\link[=azureCreateHDI]{azureCreateHDI()}}}

\item{role}{role type: 'worker', 'head' or 'edge'}

\item{size}{Numeric: the number of nodes for this type of role}

\item{mode}{Provisioning mode, "Sync" or "Async". Use "Async" to immediately return to R session after submission of request}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{resourceGroup}{Name of the resource group}

\item{verbose}{If TRUE, prints verbose messages}
}
\description{
Resize a HDInsight cluster role.
}
\seealso{
Other HDInsight functions: \code{\link{azureCreateHDI}},
  \code{\link{azureDeleteHDI}}, \code{\link{azureHDIConf}},
  \code{\link{azureListHDI}},
  \code{\link{azureRunScriptAction}},
  \code{\link{azureScriptActionHistory}}
}
