% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{AzureSM: AzureResizeHDI}
\alias{AzureResizeHDI}
\alias{AzureSM: AzureResizeHDI}
\title{Resize a HDInsight CLuster Role}
\usage{
AzureResizeHDI(AzureActiveContext, ClusterName, Role = "worker", Size = 2,
  Mode = "Sync", AzToken, SubscriptionID, ResourceGroup, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{ClusterName}{ResourceGroup Object (or use AzureActiveContext)}

\item{Role}{Role Type (Worker, Head, Edge)}

\item{Size}{Desired size of Role Type}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{MODE}{MODE Sync/Async}

\item{Token}{Token Object (or use AzureActiveContext)}

\item{Verbose}{Print Tracing information (Default False)}
}

