% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{AzureRunScriptAction}
\alias{AzureRunScriptAction}
\title{Run Script Action on HDI Cluster.}
\usage{
AzureRunScriptAction(AzureActiveContext, ScriptName = "script1", ScriptURL,
  HeadNode = TRUE, WorkerNode = FALSE, EdgeNode = FALSE, ClusterName,
  ResourceGroup, Parameters = "", AzToken, SubscriptionID, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ScriptName}{- Identifier for Custom action scrript operation}

\item{ScriptURL}{- URL to custom action script (Sring)}

\item{HeadNode}{- install on head nodes (default FALSE)}

\item{WorkerNode}{- install on worker nodes (default FALSE)}

\item{EdgeNode}{- install on worker nodes (default FALSE)}

\item{ClusterName}{ClusterName Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{AzToken}{Token Object (or use AzureActiveContext)}

\item{SubscriptionID}{SubscriptionID Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Success Message
}
\description{
Run Script Action on HDI Cluster.
}
\seealso{
Other HDInsight: \code{\link{AzureCreateHDI}},
  \code{\link{AzureDeleteHDI}}, \code{\link{AzureHDIConf}},
  \code{\link{AzureListHDI}}, \code{\link{AzureResizeHDI}},
  \code{\link{AzureScriptActionHistory}}
}

