% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{azureRunScriptAction}
\alias{azureRunScriptAction}
\title{Run script action on HDI cluster.}
\usage{
azureRunScriptAction(azureActiveContext, scriptname, scriptURL,
  headNode = TRUE, workerNode = FALSE, edgeNode = FALSE, clustername,
  resourceGroup, parameters = "", subscriptionID, wait = TRUE,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{scriptname}{Identifier for Custom action script operation}

\item{scriptURL}{URL to custom action script}

\item{headNode}{install on head nodes}

\item{workerNode}{install on worker nodes}

\item{edgeNode}{install on worker nodes}

\item{clustername}{Cluster name, used for HDI and Spark clusters. See \code{\link[=azureCreateHDI]{azureCreateHDI()}}}

\item{resourceGroup}{Name of the resource group}

\item{parameters}{parameters}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{wait}{If TRUE, runs script action synchronously, i.e. waits for successfull completion. If FALSE, submits the action asynchronously}

\item{verbose}{If TRUE, prints verbose messages}
}
\value{
Returns Success Message
}
\description{
Run script action on HDI cluster.
}
\references{
https://docs.microsoft.com/en-us/rest/api/hdinsight/hdinsight-cluster#run-script-actions-on-a-running-cluster-linux-cluster-only
}
\seealso{
Other HDInsight functions: \code{\link{azureCreateHDI}},
  \code{\link{azureDeleteHDI}}, \code{\link{azureHDIConf}},
  \code{\link{azureListHDI}}, \code{\link{azureResizeHDI}},
  \code{\link{azureScriptActionHistory}}
}
