% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureStorageAccount.R
\name{azureSAGetKey}
\alias{azureSAGetKey}
\title{Get the Storage Keys for Specified Storage Account.}
\usage{
azureSAGetKey(azureActiveContext, storageAccount, azToken, resourceGroup,
  subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{storageAccount}{storageAccount}

\item{azToken}{Token Object (or use azureActiveContext)}

\item{resourceGroup}{resourceGroup Object}

\item{subscriptionID}{subscriptionID Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Get the Storage Keys for Specified Storage Account.
}
\seealso{
Other Storage account functions: \code{\link{azureCreatestorageAccount}},
  \code{\link{azureDeletestorageAccount}},
  \code{\link{azureListSA}}
}

