% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureStorageAccount.R
\name{AzureSAGetKey}
\alias{AzureSAGetKey}
\title{Get the Storage Keys for Specified Storage Account.}
\usage{
AzureSAGetKey(AzureActiveContext, StorageAccount, AzToken, ResourceGroup,
  SubscriptionID, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{StorageAccount}{StorageAccount}

\item{AzToken}{Token Object (or use AzureActiveContext)}

\item{ResourceGroup}{ResourceGroup Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Get the Storage Keys for Specified Storage Account.
}
\seealso{
Other Storage account functions: \code{\link{AzureCreateStorageAccount}},
  \code{\link{AzureDeleteStorageAccount}},
  \code{\link{AzureListSA}}
}

