% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureStorageAccount.R
\name{azureSAGetKey}
\alias{azureSAGetKey}
\title{Get the Storage Keys for Specified Storage Account.}
\usage{
azureSAGetKey(azureActiveContext, storageAccount, azToken, resourceGroup,
  subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{storageAccount}{storageAccount}

\item{azToken}{Token Object (or use azureActiveContext)}

\item{resourceGroup}{Name of the resource group}

\item{subscriptionID}{Set the subscriptionID.  This is obtained automatically by \code{\link{azureAuthenticate}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Get the Storage Keys for Specified Storage Account.
}
\seealso{
Other Storage account functions: \code{\link{azureCreateStorageAccount}},
  \code{\link{azureDeletestorageAccount}},
  \code{\link{azureListSA}}
}
