% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureBlob.R
\name{AzureSM: AzureSAGetKey}
\alias{AzureCreateStorageAccount}
\alias{AzureDeleteStorageAccount}
\alias{AzureSAGetKey}
\alias{AzureSM: AzureCreateStorageAccount}
\alias{AzureSM: AzureDeleteStorageAccount}
\alias{AzureSM: AzureSAGetKey}
\title{Get the Storage Keys for Specified Storage Account}
\usage{
AzureSAGetKey(AzureActiveContext, StorageAccount, AzToken, ResourceGroup,
  SubscriptionID, verbose = FALSE)

AzureCreateStorageAccount(AzureActiveContext, StorageAccount, AzToken,
  ResourceGroup, SubscriptionID, verbose = FALSE)

AzureDeleteStorageAccount(AzureActiveContext, StorageAccount, AzToken,
  ResourceGroup, SubscriptionID, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{StorageAccount}{StorageAccount}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{verbose}{Print Tracing information (Default False)}

\item{Token}{Token Object (or use AzureActiveContext)}

\item{AzureActiveContext}{Azure Context Object}

\item{StorageAccount}{StorageAccount}

\item{Token}{Token Object (or use AzureActiveContext)}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{verbose}{Print Tracing information (Default False)}

\item{AzureActiveContext}{Azure Context Object}

\item{StorageAccount}{StorageAccount}

\item{Token}{Token Object (or use AzureActiveContext)}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{verbose}{Print Tracing information (Default False)}
}

