% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{azureScriptActionHistory}
\alias{azureScriptActionHistory}
\title{Get all HDInsight Script Action History for a specified clustername.}
\usage{
azureScriptActionHistory(azureActiveContext, resourceGroup, clustername = "*",
  subscriptionID, azToken, name, type, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{resourceGroup}{resourceGroup Object}

\item{clustername}{clustername Object}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{name}{name}

\item{type}{type}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of HDInsight Clusters
}
\description{
Get all HDInsight Script Action History for a specified clustername.
}
\seealso{
Other HDInsight functions: \code{\link{azureCreateHDI}},
  \code{\link{azureDeleteHDI}}, \code{\link{azureHDIConf}},
  \code{\link{azureListHDI}}, \code{\link{azureResizeHDI}},
  \code{\link{azureRunScriptAction}}
}

