% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{AzureScriptActionHistory}
\alias{AzureScriptActionHistory}
\title{Get all HDInsight Script Action Historyfor a specified ClusterName.}
\usage{
AzureScriptActionHistory(AzureActiveContext, ResourceGroup, ClusterName = "*",
  SubscriptionID, ATI, Name, Type, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{ClusterName}{ClusterName Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of HDInsight Clusters
}
\description{
Get all HDInsight Script Action Historyfor a specified ClusterName.
}
\seealso{
Other HDInsight: \code{\link{AzureCreateHDI}},
  \code{\link{AzureDeleteHDI}}, \code{\link{AzureHDIConf}},
  \code{\link{AzureListHDI}}, \code{\link{AzureResizeHDI}},
  \code{\link{AzureRunScriptAction}}
}

