% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{AzureScriptActionHistory}
\alias{AzureScriptActionHistory}
\title{Get all HDInsight Script Action History for a specified ClusterName.}
\usage{
AzureScriptActionHistory(AzureActiveContext, ResourceGroup, ClusterName = "*",
  SubscriptionID, AzToken, Name, Type, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{ClusterName}{ClusterName Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object}

\item{Name}{Name}

\item{Type}{Type}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of HDInsight Clusters
}
\description{
Get all HDInsight Script Action History for a specified ClusterName.
}
\seealso{
Other HDInsight functions: \code{\link{AzureCreateHDI}},
  \code{\link{AzureDeleteHDI}}, \code{\link{AzureHDIConf}},
  \code{\link{AzureListHDI}}, \code{\link{AzureResizeHDI}},
  \code{\link{AzureRunScriptAction}}
}

