% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureHDI.R
\name{azureScriptActionHistory}
\alias{azureScriptActionHistory}
\title{Get all HDInsight Script Action History for a specified clustername.}
\usage{
azureScriptActionHistory(azureActiveContext, resourceGroup, clustername = "*",
  subscriptionID, azToken, name, type, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{resourceGroup}{Name of the resource group}

\item{clustername}{Cluster name, used for HDI and Spark clusters. See \code{azureCreateHDI}}

\item{subscriptionID}{Set the subscriptionID.  This is obtained automatically by \code{azureAuthenticate} when only a single subscriptionID is available via Active Directory}

\item{azToken}{Azure authentication token, obtained by \code{azureAuthenticate}}

\item{name}{filter by resource name}

\item{type}{filter by resource type}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
Returns Dataframe of HDInsight Clusters
}
\description{
Get all HDInsight Script Action History for a specified clustername.
}
\seealso{
Other HDInsight functions: \code{\link{azureCreateHDI}},
  \code{\link{azureDeleteHDI}}, \code{\link{azureHDIConf}},
  \code{\link{azureListHDI}}, \code{\link{azureResizeHDI}},
  \code{\link{azureRunScriptAction}}
}
