% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSMR-package.R
\docType{package}
\name{AzureSMR}
\alias{AzureSMR}
\alias{AzureSMR-package}
\alias{AzureSMR-package}
\title{AzureSMR}
\description{
The AzureSMR package connects R to the Azure Service Manager API.
}
\details{
This enables you to use and change many Azure resources. The following is an incomplete list of available functions:
\itemize{
\item Create an Azure context:
\itemize{
\item \code{\link[=createAzureContext]{createAzureContext()}}
}
\item Resources and resource groups:
\itemize{
\item \code{\link[=azureListAllResources]{azureListAllResources()}}
\item \code{\link[=azureCreateResourceGroup]{azureCreateResourceGroup()}}
\item \code{\link[=azureListRG]{azureListRG()}}
}
\item Storage accounts and blobs:
\itemize{
\item \code{\link[=azureListSA]{azureListSA()}}
\item \code{\link[=azureCreateStorageAccount]{azureCreateStorageAccount()}}
\item \code{\link[=azureListStorageContainers]{azureListStorageContainers()}}
\item \code{\link[=azureListStorageBlobs]{azureListStorageBlobs()}}
}
\item Virtual Machines
\itemize{
\item To create a virtual machine, use \code{\link[=azureDeployTemplate]{azureDeployTemplate()}} with a suitable template
\item Start a VM: \code{\link[=azureStartVM]{azureStartVM()}}
\item Stop a VM: \code{\link[=azureStopVM]{azureStopVM()}}
\item Get status: \code{\link[=azureVMStatus]{azureVMStatus()}}
\item List VMs: \code{\link[=azureListVM]{azureListVM()}}
}
\item HDInsight (Nodes, Hive, Spark)
\item Azure Resource Manager
}
}
\keyword{package}
