% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSMR-package.R
\docType{package}
\name{AzureSMR}
\alias{AzureSMR}
\alias{AzureSMR-package}
\alias{AzureSMR-package}
\title{AzureSMR}
\description{
The AzureSMR package connects R to the Azure Service Manager API.
}
\details{
This enables you to use and change many Azure resources. The following is an incomplete list of available functions:
\itemize{
\item Create an Azure context:
\itemize{
\item \code{\link[=createAzureContext]{createAzureContext()}}
}
\item Resources and resource groups:
\itemize{
\item \code{\link[=azureListAllResources]{azureListAllResources()}}
\item \code{\link[=azureCreateResourceGroup]{azureCreateResourceGroup()}}
\item \code{\link[=azureListRG]{azureListRG()}}
}
\item Storage accounts and blobs:
\itemize{
\item \code{\link[=azureListSA]{azureListSA()}}
\item \code{\link[=azureCreateStorageAccount]{azureCreateStorageAccount()}}
\item \code{\link[=azureListStorageContainers]{azureListStorageContainers()}}
\item \code{\link[=azureListStorageBlobs]{azureListStorageBlobs()}}
\item \code{\link[=azureGetBlob]{azureGetBlob()}}
\item \code{\link[=azurePutBlob]{azurePutBlob()}}
\item \code{\link[=azureDeleteBlob]{azureDeleteBlob()}}
\item \code{\link[=azureBlobFind]{azureBlobFind()}}
}
\item Virtual Machines
\itemize{
\item List VMs: \code{\link[=azureListVM]{azureListVM()}}
\item To create a virtual machine, use \code{\link[=azureDeployTemplate]{azureDeployTemplate()}} with a suitable template
\item Start a VM: \code{\link[=azureStartVM]{azureStartVM()}}
\item Stop a VM: \code{\link[=azureStopVM]{azureStopVM()}}
\item Get status: \code{\link[=azureVMStatus]{azureVMStatus()}}
}
\item HDInsight clusters:
\itemize{
\item \code{\link[=azureListHDI]{azureListHDI()}}
\item \code{\link[=azureCreateHDI]{azureCreateHDI()}}
\item \code{\link[=azureResizeHDI]{azureResizeHDI()}}
\item \code{\link[=azureDeleteHDI]{azureDeleteHDI()}}
\item \code{\link[=azureRunScriptAction]{azureRunScriptAction()}}
\item \code{\link[=azureScriptActionHistory]{azureScriptActionHistory()}}
}
\item Azure batch:
\itemize{
\item \code{\link[=azureListBatchAccounts]{azureListBatchAccounts()}}
\item \code{\link[=azureCreateBatchAccount]{azureCreateBatchAccount()}}
\item \code{\link[=azureDeleteBatchAccount]{azureDeleteBatchAccount()}}
\item \code{\link[=azureBatchGetKey]{azureBatchGetKey()}}
}
\item Cost functions:
\itemize{
\item \code{\link[=azureDataConsumption]{azureDataConsumption()}}
\item \code{\link[=azurePricingRates]{azurePricingRates()}}
\item \code{\link[=azureExpenseCalculator]{azureExpenseCalculator()}}
}
\item Azure Data Lake Store functions:
\itemize{
\item \code{\link[=adls.ls]{adls.ls()}}
\item \code{\link[=adls.file.info]{adls.file.info()}}
\item \code{\link[=adls.mkdir]{adls.mkdir()}}
\item \code{\link[=adls.create]{adls.create()}}
\item \code{\link[=adls.delete]{adls.delete()}}
\item \code{\link[=adls.rename]{adls.rename()}}
\item \code{\link[=adls.concat]{adls.concat()}}
\item \code{\link[=adls.append]{adls.append()}}
\item \code{\link[=adls.fileoutputstream.write]{adls.fileoutputstream.write()}}
\item \code{\link[=adls.fileoutputstream.flush]{adls.fileoutputstream.flush()}}
\item \code{\link[=adls.fileoutputstream.close]{adls.fileoutputstream.close()}}
\item \code{\link[=adls.read]{adls.read()}}
\item \code{\link[=adls.fileinputstream.read]{adls.fileinputstream.read()}}
\item \code{\link[=adls.fileinputstream.seek]{adls.fileinputstream.seek()}}
\item \code{\link[=adls.fileinputstream.skip]{adls.fileinputstream.skip()}}
\item \code{\link[=adls.fileinputstream.available]{adls.fileinputstream.available()}}
\item \code{\link[=adls.fileinputstream.length]{adls.fileinputstream.length()}}
\item \code{\link[=adls.fileinputstream.getpos]{adls.fileinputstream.getpos()}}
\item \code{\link[=adls.fileinputstream.close]{adls.fileinputstream.close()}}
}
}
}
\keyword{package}
