% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{azureSparkCMD}
\alias{azureSparkCMD}
\title{Send Spark Statements/comamnds (REPL/Interactive mode).}
\usage{
azureSparkCMD(azureActiveContext, CMD, clustername, hdiAdmin, hdiPassword,
  sessionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{CMD}{CMD}

\item{clustername}{clustername Object}

\item{hdiAdmin}{hdiAdmin Object}

\item{hdiPassword}{hdiPassword  Object}

\item{sessionID}{sessionID Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Send Spark Statements/comamnds (REPL/Interactive mode).
}
\seealso{
Other Spark: \code{\link{azureSparkJob}},
  \code{\link{azureSparkListJobs}},
  \code{\link{azureSparkListSessions}},
  \code{\link{azureSparkNewSession}},
  \code{\link{azureSparkShowURL}},
  \code{\link{azureSparkStopSession}}
}

