% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{AzureSparkCMD}
\alias{AzureSparkCMD}
\title{Send Spark Statements/comamnds (REPL/Interactive mode).}
\usage{
AzureSparkCMD(AzureActiveContext, CMD, ClusterName, HDIAdmin, HDIPassword,
  SessionID, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{CMD}{CMD}

\item{ClusterName}{ClusterName Object}

\item{HDIAdmin}{HDIAdmin Object}

\item{HDIPassword}{HDIPassword Object}

\item{SessionID}{SessionID Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Send Spark Statements/comamnds (REPL/Interactive mode).
}
\seealso{
Other Spark: \code{\link{AzureSparkJob}},
  \code{\link{AzureSparkListJobs}},
  \code{\link{AzureSparkListSessions}},
  \code{\link{AzureSparkNewSession}},
  \code{\link{AzureSparkShowURL}},
  \code{\link{AzureSparkStopSession}}
}

