% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{azureSparkCMD}
\alias{azureSparkCMD}
\title{Send Spark Statements/comamnds (REPL/Interactive mode).}
\usage{
azureSparkCMD(azureActiveContext, CMD, clustername, hdiAdmin, hdiPassword,
  sessionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by `AzureSMR`, created by [createAzureContext()]}

\item{CMD}{CMD}

\item{clustername}{Cluster name, used for HDI and Spark clusters. See [azureCreateHDI()]}

\item{hdiAdmin}{HDInsight admin username. See [azureCreateHDI()]}

\item{hdiPassword}{HDInsight admin password. See [azureCreateHDI()]}

\item{sessionID}{Spark sessionID. See [azureSparkCMD()]}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Send Spark Statements/comamnds (REPL/Interactive mode).
}
\seealso{
Other Spark functions: \code{\link{azureSparkJob}},
  \code{\link{azureSparkListJobs}},
  \code{\link{azureSparkListSessions}},
  \code{\link{azureSparkNewSession}},
  \code{\link{azureSparkShowURL}},
  \code{\link{azureSparkStopSession}}
}

