% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{AzureSM: AzureSparkCMD}
\alias{AzureSM: AzureSparkCMD}
\alias{AzureSparkCMD}
\title{Send Spark Statements/comamnds (REPL/Interactive mode)}
\usage{
AzureSparkCMD(AzureActiveContext, CMD, ClusterName, HDIAdmin, HDIPassword,
  SessionID, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{CMD}{CMD}

\item{ClusterName}{ClusterName}

\item{HDIAdmin}{HDIAdmin - HDinsight Administrator Name}

\item{HDIPassword}{HDinsight Administrator Name}

\item{SessionID}{SessionID}

\item{Verbose}{Print Tracing information (Default False)}
}

