% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{AzureSM: AzureSparkJob}
\alias{AzureSM: AzureSparkJob}
\alias{AzureSparkJob}
\title{Submit Spark Job (Batch mode)}
\usage{
AzureSparkJob(AzureActiveContext, FILE, ClusterName, HDIAdmin, HDIPassword,
  Log = "URL", verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{FILE}{FILE}

\item{ClusterName}{ClusterName}

\item{HDIAdmin}{HDIAdmin - HDinsight Administrator Name}

\item{HDIPassword}{HDinsight Administrator Name}

\item{Verbose}{Print Tracing information (Default False)}
}

