% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{azureSparkJob}
\alias{azureSparkJob}
\title{Submit Spark Job (Batch mode).}
\usage{
azureSparkJob(azureActiveContext, FILE, clustername, hdiAdmin, hdiPassword,
  log = "URL", verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{FILE}{file}

\item{clustername}{clustername Object}

\item{hdiAdmin}{hdiAdmin Object}

\item{hdiPassword}{hdiPassword  Object}

\item{log}{log}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Submit Spark Job (Batch mode).
}
\seealso{
Other Spark: \code{\link{azureSparkCMD}},
  \code{\link{azureSparkListJobs}},
  \code{\link{azureSparkListSessions}},
  \code{\link{azureSparkNewSession}},
  \code{\link{azureSparkShowURL}},
  \code{\link{azureSparkStopSession}}
}

