% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{azureSparkJob}
\alias{azureSparkJob}
\title{Submit Spark Job (Batch mode).}
\usage{
azureSparkJob(azureActiveContext, FILE, clustername, hdiAdmin, hdiPassword,
  log = "URL", verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by `AzureSMR`, created by [createAzureContext()]}

\item{FILE}{file}

\item{clustername}{Cluster name, used for HDI and Spark clusters. See [azureCreateHDI()]}

\item{hdiAdmin}{HDInsight admin username. See [azureCreateHDI()]}

\item{hdiPassword}{HDInsight admin password. See [azureCreateHDI()]}

\item{log}{log}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Submit Spark Job (Batch mode).
}
\seealso{
Other Spark functions: \code{\link{azureSparkCMD}},
  \code{\link{azureSparkListJobs}},
  \code{\link{azureSparkListSessions}},
  \code{\link{azureSparkNewSession}},
  \code{\link{azureSparkShowURL}},
  \code{\link{azureSparkStopSession}}
}

