% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{AzureSparkListJobs}
\alias{AzureSparkListJobs}
\title{List Spark Jobs (Batch mode).}
\usage{
AzureSparkListJobs(AzureActiveContext, ClusterName, HDIAdmin, HDIPassword, KIND,
  verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ClusterName}{ClusterName Object}

\item{HDIAdmin}{HDIAdmin Object}

\item{HDIPassword}{HDIPassword Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
(manually direct output to Blob fule /SQL in script)
}
\description{
List Spark Jobs (Batch mode).
}
\seealso{
Other Spark: \code{\link{AzureSparkCMD}},
  \code{\link{AzureSparkJob}},
  \code{\link{AzureSparkListSessions}},
  \code{\link{AzureSparkNewSession}},
  \code{\link{AzureSparkShowURL}},
  \code{\link{AzureSparkStopSession}}
}

