% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{AzureSM: AzureSparkListJobs}
\alias{AzureSM: AzureSparkListJobs}
\alias{AzureSparkListJobs}
\title{List Spark Jobs (Batch mode)}
\usage{
AzureSparkListJobs(AzureActiveContext, ClusterName, HDIAdmin, HDIPassword, KIND,
  verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{ClusterName}{ClusterName}

\item{HDIAdmin}{HDIAdmin - HDinsight Administrator Name}

\item{HDIPassword}{HDinsight Administrator Name}

\item{Verbose}{Print Tracing information (Default False)}
}
\value{
(manually direct output to Blob fule /SQL in script)
}

