% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{azureSparkListJobs}
\alias{azureSparkListJobs}
\title{List Spark Jobs (Batch mode).}
\usage{
azureSparkListJobs(azureActiveContext, clustername, hdiAdmin, hdiPassword,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{clustername}{clustername Object}

\item{hdiAdmin}{hdiAdmin Object}

\item{hdiPassword}{hdiPassword  Object}

\item{verbose}{Print Tracing information (Default False)}
}
\value{
(manually direct output to blob fule /SQL in script)
}
\description{
List Spark Jobs (Batch mode).
}
\seealso{
Other Spark: \code{\link{azureSparkCMD}},
  \code{\link{azureSparkJob}},
  \code{\link{azureSparkListSessions}},
  \code{\link{azureSparkNewSession}},
  \code{\link{azureSparkShowURL}},
  \code{\link{azureSparkStopSession}}
}

