% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{AzureSM: AzureSparkListSessions}
\alias{AzureSM: AzureSparkListSessions}
\alias{AzureSparkListSessions}
\title{List Spark Sessions}
\usage{
AzureSparkListSessions(AzureActiveContext, ClusterName, HDIAdmin, HDIPassword,
  verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{ClusterName}{ClusterName}

\item{HDIAdmin}{HDIAdmin - HDinsight Administrator Name}

\item{HDIPassword}{HDinsight Administrator Name}

\item{Verbose}{Print Tracing information (Default False)}
}

