% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{AzureSparkListSessions}
\alias{AzureSparkListSessions}
\title{List Spark Sessions.}
\usage{
AzureSparkListSessions(AzureActiveContext, ClusterName, HDIAdmin, HDIPassword,
  verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ClusterName}{ClusterName Object}

\item{HDIAdmin}{HDIAdmin Object}

\item{HDIPassword}{HDIPassword Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List Spark Sessions.
}
\seealso{
Other Spark: \code{\link{AzureSparkCMD}},
  \code{\link{AzureSparkJob}},
  \code{\link{AzureSparkListJobs}},
  \code{\link{AzureSparkNewSession}},
  \code{\link{AzureSparkShowURL}},
  \code{\link{AzureSparkStopSession}}
}

