% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{azureSparkListSessions}
\alias{azureSparkListSessions}
\title{List Spark Sessions.}
\usage{
azureSparkListSessions(azureActiveContext, clustername, hdiAdmin, hdiPassword,
  verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{clustername}{Set the clustername}

\item{hdiAdmin}{Set the HDInsight admin username}

\item{hdiPassword}{Set the HDInsight admin Password}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
List Spark Sessions.
}
\seealso{
Other Spark: \code{\link{azureSparkCMD}},
  \code{\link{azureSparkJob}},
  \code{\link{azureSparkListJobs}},
  \code{\link{azureSparkNewSession}},
  \code{\link{azureSparkShowURL}},
  \code{\link{azureSparkStopSession}}
}

