% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{AzureSparkNewSession}
\alias{AzureSparkNewSession}
\title{Create new Spark Session.}
\usage{
AzureSparkNewSession(AzureActiveContext, ClusterName, HDIAdmin, HDIPassword,
  Kind = "spark", verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ClusterName}{ClusterName}

\item{HDIAdmin}{HDIAdmin - HDinsight Administrator Name}

\item{HDIPassword}{HDinsight Administrator Name}

\item{Kind}{Kind Spark/PySpark/Spark}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Create new Spark Session.
}
\seealso{
Other Spark: \code{\link{AzureSparkCMD}},
  \code{\link{AzureSparkJob}},
  \code{\link{AzureSparkListJobs}},
  \code{\link{AzureSparkListSessions}},
  \code{\link{AzureSparkShowURL}},
  \code{\link{AzureSparkStopSession}}
}

