% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{AzureSparkShowURL}
\alias{AzureSparkShowURL}
\title{Show Spark Log Output.}
\usage{
AzureSparkShowURL(AzureActiveContext, URL)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{URL}{URL}
}
\value{
Show Log in Browser
}
\description{
Show Spark Log Output.
}
\seealso{
Other Spark: \code{\link{AzureSparkCMD}},
  \code{\link{AzureSparkJob}},
  \code{\link{AzureSparkListJobs}},
  \code{\link{AzureSparkListSessions}},
  \code{\link{AzureSparkNewSession}},
  \code{\link{AzureSparkStopSession}}
}

