% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{AzureSM: AzureSparkShowURL}
\alias{AzureSM: AzureSparkShowURL}
\alias{AzureSparkShowURL}
\title{Show Spark Log Output}
\usage{
AzureSparkShowURL(AzureActiveContext, URL)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{URL}{(From Batch Job)}
}
\value{
Show Log in Browser
}

