% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{azureSparkStopSession}
\alias{azureSparkStopSession}
\title{Stop a Spark Sessions.}
\usage{
azureSparkStopSession(azureActiveContext, clustername, hdiAdmin, hdiPassword,
  sessionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{clustername}{Set the clustername}

\item{hdiAdmin}{Set the HDInsight admin username}

\item{hdiPassword}{Set the HDInsight admin Password}

\item{sessionID}{Set the Spark sessionID}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Stop a Spark Sessions.
}
\seealso{
Other Spark: \code{\link{azureSparkCMD}},
  \code{\link{azureSparkJob}},
  \code{\link{azureSparkListJobs}},
  \code{\link{azureSparkListSessions}},
  \code{\link{azureSparkNewSession}},
  \code{\link{azureSparkShowURL}}
}

