% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{azureSparkStopSession}
\alias{azureSparkStopSession}
\title{Stop a Spark Sessions.}
\usage{
azureSparkStopSession(azureActiveContext, clustername, hdiAdmin, hdiPassword,
  sessionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{clustername}{clustername Object}

\item{hdiAdmin}{hdiAdmin Object}

\item{hdiPassword}{hdiPassword  Object}

\item{sessionID}{sessionID Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Stop a Spark Sessions.
}
\seealso{
Other Spark: \code{\link{azureSparkCMD}},
  \code{\link{azureSparkJob}},
  \code{\link{azureSparkListJobs}},
  \code{\link{azureSparkListSessions}},
  \code{\link{azureSparkNewSession}},
  \code{\link{azureSparkShowURL}}
}

