% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureSpark.R
\name{AzureSparkStopSession}
\alias{AzureSparkStopSession}
\title{Stop a Spark Sessions.}
\usage{
AzureSparkStopSession(AzureActiveContext, ClusterName, HDIAdmin, HDIPassword,
  SessionID, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ClusterName}{ClusterName Object}

\item{HDIAdmin}{HDIAdmin Object}

\item{HDIPassword}{HDIPassword Object}

\item{SessionID}{SessionID Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Stop a Spark Sessions.
}
\seealso{
Other Spark: \code{\link{AzureSparkCMD}},
  \code{\link{AzureSparkJob}},
  \code{\link{AzureSparkListJobs}},
  \code{\link{AzureSparkListSessions}},
  \code{\link{AzureSparkNewSession}},
  \code{\link{AzureSparkShowURL}}
}

