% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{AzureStartVM}
\alias{AzureStartVM}
\title{Start a Virtual Machine.}
\usage{
AzureStartVM(AzureActiveContext, ResourceGroup, VMName, Mode = "Sync",
  SubscriptionID, AzToken, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{VMName}{Virtual Machine Name}

\item{Mode}{Wait for operation to complete 'Sync' (Default)}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Start a Virtual Machine.
}
\seealso{
Other Virtual machine functions: \code{\link{AzureDeleteVM}},
  \code{\link{AzureListVM}}, \code{\link{AzureStopVM}},
  \code{\link{AzureVMStatus}}
}

