% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureStartVM}
\alias{azureStartVM}
\title{Start a Virtual Machine.}
\usage{
azureStartVM(azureActiveContext, resourceGroup, vmName, mode = "Sync",
  subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by `AzureSMR`, created by [createAzureContext()]}

\item{resourceGroup}{Name of the resource group}

\item{vmName}{Virtual Machine name}

\item{mode}{Wait for operation to complete 'Sync' (Default)}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by [azureAuthenticate()] when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Start a Virtual Machine.
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureGetAllVMstatus}},
  \code{\link{azureListScaleSetNetwork}},
  \code{\link{azureListScaleSetVM}},
  \code{\link{azureListScaleSets}},
  \code{\link{azureListVM}}, \code{\link{azureStopVM}},
  \code{\link{azureVMInfo}}, \code{\link{azureVMStatus}}
}

