% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureStartVM}
\alias{azureStartVM}
\title{Start a Virtual Machine.}
\usage{
azureStartVM(azureActiveContext, resourceGroup, vmName, mode = "Sync",
  subscriptionID, azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{resourceGroup}{Name of the resource group}

\item{vmName}{Virtual Machine name}

\item{mode}{Wait for operation to complete 'Sync' (Default)}

\item{subscriptionID}{Set the subscriptionID.  This is obtained automatically by \code{\link{azureAuthenticate}} when only a single subscriptionID is available via Active Directory}

\item{azToken}{Azure authentication token, obtained by \code{\link{azureAuthenticate}}}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Start a Virtual Machine.
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureListVM}}, \code{\link{azureStopVM}},
  \code{\link{azureVMStatus}}
}

