% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureStopVM}
\alias{azureStopVM}
\title{Stop a Virtual Machine.}
\usage{
azureStopVM(azureActiveContext, resourceGroup, vmName, mode = "Sync",
  subscriptionID, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}}

\item{resourceGroup}{Name of the resource group}

\item{vmName}{Name of the virtual machine}

\item{mode}{Wait for operation to complete 'Sync' (Default)}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Stop a Virtual Machine.
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureListScaleSetNetwork}},
  \code{\link{azureListScaleSetVM}},
  \code{\link{azureListScaleSets}},
  \code{\link{azureListVM}}, \code{\link{azureStartVM}},
  \code{\link{azureVMStatus}}
}
