% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureStopVM}
\alias{azureStopVM}
\title{Stop a Virtual Machine.}
\usage{
azureStopVM(azureActiveContext, resourceGroup, vmName, mode = "Sync",
  subscriptionID, azToken, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{resourceGroup}{Set the name of the Resource Group}

\item{vmName}{Set the name of the virtual Machine}

\item{mode}{Wait for operation to complete 'Sync' (Default)}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Stop a Virtual Machine.
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureListVM}}, \code{\link{azureStartVM}},
  \code{\link{azureVMStatus}}
}

