% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{AzureStopVM}
\alias{AzureStopVM}
\title{Stop a Virtual Machine.}
\usage{
AzureStopVM(AzureActiveContext, ResourceGroup, VMName, Mode = "Sync",
  SubscriptionID, AzToken, verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{VMName}{VMName Object}

\item{Mode}{Wait for operation to complete "Sync" (Default)}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Stop a Virtual Machine.
}
\seealso{
Other Virtual Machine: \code{\link{AzureDeleteVM}},
  \code{\link{AzureListVM}}, \code{\link{AzureStartVM}},
  \code{\link{AzureVMStatus}}
}

