% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureVMInfo}
\alias{azureVMInfo}
\title{Get detailed information (e.g., name, OS, size, etc.) of a Virtual Machine.}
\usage{
azureVMInfo(azureActiveContext, resourceGroup, vmName, subscriptionID,
  ignore = "N", verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by `AzureSMR`, created by [createAzureContext()]}

\item{resourceGroup}{Name of the resource group}

\item{vmName}{Name of the virtual machine}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by [azureAuthenticate()] when only a single subscriptionID is available via Active Directory}

\item{ignore}{ignore}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Get detailed information (e.g., name, OS, size, etc.) of a Virtual Machine.
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureGetAllVMstatus}},
  \code{\link{azureListScaleSetNetwork}},
  \code{\link{azureListScaleSetVM}},
  \code{\link{azureListScaleSets}},
  \code{\link{azureListVM}}, \code{\link{azureStartVM}},
  \code{\link{azureStopVM}}, \code{\link{azureVMStatus}}
}

