% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureVMStatus}
\alias{azureVMStatus}
\title{Get Status of a Virtual Machine.}
\usage{
azureVMStatus(azureActiveContext, resourceGroup, vmName, subscriptionID,
  ignore = "N", verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{resourceGroup}{Name of the resource group}

\item{vmName}{Name of the virtual machine}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by \code{\link[=azureAuthenticate]{azureAuthenticate()}} when only a single subscriptionID is available via Active Directory}

\item{ignore}{ignore}

\item{verbose}{If TRUE, prints verbose messages}
}
\description{
Get Status of a Virtual Machine.
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureGetAllVMstatus}},
  \code{\link{azureListScaleSetNetwork}},
  \code{\link{azureListScaleSetVM}},
  \code{\link{azureListScaleSets}},
  \code{\link{azureListVM}}, \code{\link{azureStartVM}},
  \code{\link{azureStopVM}}, \code{\link{azureVMInfo}}
}
