% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{AzureVMStatus}
\alias{AzureVMStatus}
\title{Get Status of a Virtual Machine.}
\usage{
AzureVMStatus(AzureActiveContext, ResourceGroup, VMName, SubscriptionID,
  AzToken, Ignore = "N", verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Context Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{VMName}{VMName Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{AzToken}{Token Object (or use AzureActiveContext)}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Get Status of a Virtual Machine.
}
\seealso{
Other Virtual Machine: \code{\link{AzureDeleteVM}},
  \code{\link{AzureListVM}}, \code{\link{AzureStartVM}},
  \code{\link{AzureStopVM}}
}

