% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{AzureSM: AzureVMStatus}
\alias{AzureSM: AzureVMStatus}
\alias{AzureVMStatus}
\title{Get Status of a Virtual Machine}
\usage{
AzureVMStatus(AzureActiveContext, ResourceGroup, VMName, SubscriptionID,
  AzToken, Ignore = "N", Verbose = FALSE)
}
\arguments{
\item{AzureActiveContext}{Azure Context Object}

\item{ResourceGroup}{ResourceGroup Object (or use AzureActiveContext)}

\item{VMName}{Virtual Machine Name}

\item{SubscriptionID}{SubscriptionID Object (or use AzureActiveContext)}

\item{Verbose}{Print Tracing information (Default False)}

\item{Token}{Token Object (or use AzureActiveContext)}
}

