% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureVMStatus}
\alias{azureVMStatus}
\title{Get Status of a Virtual Machine.}
\usage{
azureVMStatus(azureActiveContext, resourceGroup, vmName, subscriptionID,
  azToken, ignore = "N", verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{Context Object}

\item{resourceGroup}{resourceGroup Object}

\item{vmName}{vmName Object}

\item{subscriptionID}{subscriptionID Object}

\item{azToken}{Token Object}

\item{ignore}{ignore}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Get Status of a Virtual Machine.
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureListVM}}, \code{\link{azureStartVM}},
  \code{\link{azureStopVM}}
}

