% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureVM.R
\name{azureVMStatus}
\alias{azureVMStatus}
\title{Get Status of a Virtual Machine.}
\usage{
azureVMStatus(azureActiveContext, resourceGroup, vmName, subscriptionID,
  azToken, ignore = "N", verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{resourceGroup}{Name of the resource group}

\item{vmName}{Name of the virtual Machine}

\item{subscriptionID}{Set the subscriptionID.  This is obtained automatically by \code{\link{azureAuthenticate}} when only a single subscriptionID is available via Active Directory}

\item{azToken}{Azure authentication token, obtained by \code{\link{azureAuthenticate}}}

\item{ignore}{ignore}

\item{verbose}{Print Tracing information (Default False)}
}
\description{
Get Status of a Virtual Machine.
}
\seealso{
Other Virtual machine functions: \code{\link{azureDeleteVM}},
  \code{\link{azureListScaleSetNetwork}},
  \code{\link{azureListScaleSetVM}},
  \code{\link{azureListScaleSets}},
  \code{\link{azureListVM}}, \code{\link{azureStartVM}},
  \code{\link{azureStopVM}}
}

