% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureDataLake.R
\name{createAdlFileOutputStream}
\alias{createAdlFileOutputStream}
\title{Create an adlFileOutputStream.
Create a container (\code{adlFileOutputStream}) for holding variables used by the Azure Data Lake Store data functions.}
\usage{
createAdlFileOutputStream(azureActiveContext, accountFQDN)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{accountFQDN}{the account FDQN}
}
\value{
An \code{adlFileOutputStream} object
}
\description{
Create an adlFileOutputStream.
Create a container (\code{adlFileOutputStream}) for holding variables used by the Azure Data Lake Store data functions.
}
\seealso{
Other Azure Data Lake Store functions: \code{\link{adlFileOutputStreamWrite}},
  \code{\link{azureDataLakeAppend}},
  \code{\link{azureDataLakeCreate}},
  \code{\link{azureDataLakeDelete}},
  \code{\link{azureDataLakeGetFileStatus}},
  \code{\link{azureDataLakeListStatus}},
  \code{\link{azureDataLakeMkdirs}},
  \code{\link{azureDataLakeRead}}
}
