% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{createAdlRetryPolicy}
\alias{createAdlRetryPolicy}
\title{Create adlRetryPolicy.
Create a adlRetryPolicy (\code{adlRetryPolicy}) for holding variables used by the Azure Data Lake Store data functions.}
\usage{
createAdlRetryPolicy(azureActiveContext,
  retryPolicyType = retryPolicyEnum$EXPONENTIALBACKOFF, verbose = FALSE)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{retryPolicyType}{the type of retryPlociy object to create.}

\item{verbose}{Print tracing information (default FALSE).}
}
\value{
An \code{adlRetryPolicy} object
}
\description{
Create adlRetryPolicy.
Create a adlRetryPolicy (\code{adlRetryPolicy}) for holding variables used by the Azure Data Lake Store data functions.
}
\references{
\url{https://github.com/Azure/azure-data-lake-store-java/blob/master/src/main/java/com/microsoft/azure/datalake/store/retrypolicies/RetryPolicy.java}
}
\seealso{
Other Azure Data Lake Store functions: \code{\link{adlFileInputStreamAvailable}},
  \code{\link{adlFileInputStreamClose}},
  \code{\link{adlFileInputStreamGetPos}},
  \code{\link{adlFileInputStreamLength}},
  \code{\link{adlFileInputStreamMarkSupported}},
  \code{\link{adlFileInputStreamMark}},
  \code{\link{adlFileInputStreamReadBuffered}},
  \code{\link{adlFileInputStreamRead}},
  \code{\link{adlFileInputStreamReset}},
  \code{\link{adlFileInputStreamSeek}},
  \code{\link{adlFileInputStreamSkip}},
  \code{\link{adlFileOutputStreamClose}},
  \code{\link{adlFileOutputStreamFlush}},
  \code{\link{adlFileOutputStreamWrite}},
  \code{\link{azureDataLakeAppendBOS}},
  \code{\link{azureDataLakeAppendCore}},
  \code{\link{azureDataLakeAppend}},
  \code{\link{azureDataLakeCreate}},
  \code{\link{azureDataLakeDelete}},
  \code{\link{azureDataLakeGetFileStatus}},
  \code{\link{azureDataLakeListStatus}},
  \code{\link{azureDataLakeMkdirs}},
  \code{\link{azureDataLakeOpenBIS}},
  \code{\link{azureDataLakeReadCore}},
  \code{\link{azureDataLakeRead}},
  \code{\link{createAdlExponentialBackoffRetryPolicy}},
  \code{\link{createAdlFileInputStream}},
  \code{\link{createAdlFileOutputStream}},
  \code{\link{createAdlNonIdempotentRetryPolicy}},
  \code{\link{readFromService}},
  \code{\link{shouldRetry.adlExponentialBackoffRetryPolicy}},
  \code{\link{shouldRetry.adlNonIdempotentRetryPolicy}},
  \code{\link{shouldRetry}}
}
