% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContextObject.R
\name{CreateAzureContext}
\alias{CreateAzureContext}
\title{Create an AzureContext.}
\usage{
CreateAzureContext()
}
\description{
Create an AzureContext.
}
\seealso{
Other Context Object: \code{\link{DumpAzureContext}}
}

