% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContextObject.R
\name{createAzureContext}
\alias{createAzureContext}
\title{Create an Azure Context.}
\usage{
createAzureContext(tenantID, clientID, authKey, configFile)
}
\arguments{
\item{tenantID}{The tenant ID provided during creation of the Active Directory application / service principal}

\item{clientID}{The client ID provided during creation of the Active Directory application / service principal}

\item{authKey}{The authentication key provided during creation of the Active Directory application / service principal}
}
\value{
An `azureActiveContext` object
}
\description{
Create a container (`azureActiveContext`) for holding variables used by the `AzureSMR` package.  If `tenantID`, `clientID` and `authKey` are provided, attempts to authenticate the session against Azure.
}
\details{
See the Azure documentation (\url{https://azure.microsoft.com/en-us/documentation/articles/resource-group-create-service-principal-portal/}) for information to configure an Active Directory application.
}
\seealso{
[setAzureContext()], [azureAuthenticate()], [read.AzureSMR.config)]

Other azureActiveContext functions: \code{\link{setAzureContext}}
}

