% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContextObject.R
\name{createAzureContext}
\alias{createAzureContext}
\title{Create an AzureContext}
\usage{
createAzureContext(TID, CID, KEY)
}
\arguments{
\item{TID}{The Tenant ID which was provided when the Active Directory application / service principal is created}

\item{CID}{The Client ID which was provided when the Active Directory application / service principal is created}

\item{KEY}{The Authentication Key which was provided when the Active Directory application / service principal is created}
}
\description{
Create a container(AzureContext) for holding variables used by the AzureSMR package.
If the Tenant ID, Client ID and Authenication Key is provided the function will attempt to
authenicate the session.
}
\note{
See \url{https://azure.microsoft.com/en-us/documentation/articles/resource-group-create-service-principal-portal/} to learn how to set up an Active directory application
}
\references{
\url{https://azure.microsoft.com/en-us/documentation/articles/resource-group-create-service-principal-portal/}
}
\seealso{
Other Context Object: \code{\link{dumpAzureContext}}
}

