% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContextObject.R
\name{createAzureContext}
\alias{createAzureContext}
\title{Create an Azure Context.}
\usage{
createAzureContext(tenantID, clientID, authKey)
}
\arguments{
\item{tenantID}{The Tenant ID provided during creation of the Active Directory application / service principal}

\item{clientID}{The Client ID provided during creation of the Active Directory application / service principal}

\item{authKey}{The Authentication Key provided during creation of the Active Directory application / service principal}
}
\value{
An \code{azureActiveContext} object
}
\description{
Create a container (\code{azureActiveContext}) for holding variables used by the \code{AzureSMR} package.  If \code{tenantID}, \code{clientID} and \code{authKey} are provided, attempts to authenticate the session against Azure.
}
\details{
See the Azure documentation (\url{https://azure.microsoft.com/en-us/documentation/articles/resource-group-create-service-principal-portal/}) for information to configure an Active Directory application.
}
\seealso{
\code{\link[=setAzureContext]{setAzureContext()}}, \code{\link[=azureAuthenticate]{azureAuthenticate()}}

Other azureActiveContext functions: \code{\link{setAzureContext}}
}
