% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContextObject.R
\name{DumpAzureContext}
\alias{DumpAzureContext}
\title{AzureSM: Dumps the contents of the AzureContext}
\usage{
DumpAzureContext(AzEnv)
}
\arguments{
\item{Azure}{Context Object}
}
\description{
DumpAzureContext
}
\details{
Dumps the contents of the AzureContext
}

