% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContextObject.R
\name{dumpAzureContext}
\alias{dumpAzureContext}
\title{Dumps out the contents of the AzureContext.}
\usage{
dumpAzureContext(azureActiveContext)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}
}
\description{
Dumps out the contents of the AzureContext.
}
\seealso{
Other Context Object: \code{\link{createAzureContext}}
}

