% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContextObject.R
\name{dumpAzureContext}
\alias{dumpAzureContext}
\title{Dumps the contents of the AzureContext.}
\usage{
dumpAzureContext(azureActiveContext)
}
\arguments{
\item{azureActiveContext}{Context Object}
}
\description{
Dumps the contents of the AzureContext.
}
\seealso{
Other Context Object: \code{\link{createAzureContext}}
}

