% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.azureActiveContext}
\alias{as.azureActiveContext}
\alias{is.azureActiveContext}
\title{azureActiveContext object.}
\usage{
as.azureActiveContext(x)

is.azureActiveContext(x)
}
\arguments{
\item{x}{Object to create, test or print}

\item{...}{Ignored}
}
\description{
Functions for creating and displaying information about azureActiveContext objects.
}
\seealso{
\code{\link[=createAzureContext]{createAzureContext()}}
}
