% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.azureActiveContext}
\alias{as.azureActiveContext}
\alias{is.azureActiveContext}
\alias{as.adlFileOutputStream}
\alias{is.adlFileOutputStream}
\alias{as.adlFileInputStream}
\alias{is.adlFileInputStream}
\alias{as.adlRetryPolicy}
\alias{is.adlRetryPolicy}
\title{azureActiveContext object.}
\usage{
as.azureActiveContext(x)

is.azureActiveContext(x)

as.adlFileOutputStream(x)

is.adlFileOutputStream(x)

as.adlFileInputStream(x)

is.adlFileInputStream(x)

as.adlRetryPolicy(x)

is.adlRetryPolicy(x)
}
\arguments{
\item{x}{the Object to create, test or print}
}
\description{
Functions for creating and displaying information about azureActiveContext objects.

Functions for creating and displaying information about adlFileOutputStream objects.

Functions for creating and displaying information about adlFileInputStream objects.

Functions for creating and displaying information about adlRetryPolicy objects.
}
\seealso{
\code{\link[=createAzureContext]{createAzureContext()}}

\code{\link[=createAdlFileOutputStream]{createAdlFileOutputStream()}}

\code{\link[=createAdlFileInputStream]{createAdlFileInputStream()}}

\code{\link[=createAdlFileInputStream]{createAdlFileInputStream()}}
}
