% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{dumpAzureContext}
\alias{dumpAzureContext}
\alias{AzureListRG}
\title{Dumps out the contents of the AzureContext.}
\usage{
dumpAzureContext(azureActiveContext)

AzureListRG(...)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by \code{AzureSMR}, created by \code{\link[=createAzureContext]{createAzureContext()}}}

\item{...}{passed to \code{\link[=azureListRG]{azureListRG()}}}
}
\description{
Dumps out the contents of the AzureContext.

Deprecated functions.
}
