% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContextObject.R
\name{setAzureContext}
\alias{setAzureContext}
\title{Updates the value of an AzureContext variable.}
\usage{
setAzureContext(azureActiveContext, TID, CID, KEY, azToken, subscriptionID,
  resourceGroup, storageKey, storageAccount, container, blob, vmName, hdiAdmin,
  hdiPassword, clustername, kind, sessionID)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{TID}{The Tenant ID which was provided when the Active Directory application / service principal is created}

\item{CID}{The Client ID which was provided when the Active Directory application / service principal is created}

\item{KEY}{The Authentication Key which was provided when the Active Directory application / service principal is created}

\item{azToken}{Azure Token Object - Obtained by AzureAuthenticate}

\item{subscriptionID}{Set the subscriptionID Obtained automatically by AzureAuthenticate when only one is available}

\item{resourceGroup}{Set the name of the Resource Group}

\item{storageKey}{Set the Storage Key associated with storage account}

\item{storageAccount}{Set the name of the azure storage account}

\item{container}{Set the storage container}

\item{blob}{Set the blob name}

\item{vmName}{Set the name of the virtual Machine}

\item{hdiAdmin}{Set the HDInsight admin username}

\item{hdiPassword}{Set the HDInsight admin Password}

\item{clustername}{Set the clustername}

\item{kind}{Set the HDinsight kind "hadoop","spark" or "pyspark"}

\item{sessionID}{Set the Spark sessionID}
}
\description{
Updates the value of an AzureContext variable.
}

