% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContextObject.R
\name{setAzureContext}
\alias{setAzureContext}
\title{Updates azureActiveContext object.}
\usage{
setAzureContext(azureActiveContext, tenantID, clientID, authKey, subscriptionID,
  resourceGroup, storageKey, storageAccount, container, blob, vmName, hdiAdmin,
  hdiPassword, clustername, kind, sessionID)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by `AzureSMR`, created by [createAzureContext()]}

\item{tenantID}{The tenant ID provided during creation of the Active Directory application / service principal}

\item{clientID}{The client ID provided during creation of the Active Directory application / service principal}

\item{authKey}{The authentication key provided during creation of the Active Directory application / service principal}

\item{subscriptionID}{Subscription ID.  This is obtained automatically by [azureAuthenticate()] when only a single subscriptionID is available via Active Directory}

\item{resourceGroup}{Name of the resource group}

\item{storageKey}{Storage key associated with storage account}

\item{storageAccount}{Name of the azure storage account. Storage account names must be between 3 and 24 characters in length and use numbers and lower-case letters only.}

\item{container}{Storage container name. See [azureListStorageContainers()]}

\item{blob}{Blob name}

\item{vmName}{Name of the virtual machine}

\item{hdiAdmin}{HDInsight admin username. See [azureCreateHDI()]}

\item{hdiPassword}{HDInsight admin password. See [azureCreateHDI()]}

\item{clustername}{Cluster name, used for HDI and Spark clusters. See [azureCreateHDI()]}

\item{kind}{HDinsight kind: "hadoop","spark" or "rserver". See [azureCreateHDI()]}

\item{sessionID}{Spark sessionID. See [azureSparkCMD()]}
}
\description{
Updates the value of an `azureActiveContext` object, created by [createAzureContext()]
}
\seealso{
Other azureActiveContext functions: \code{\link{createAzureContext}}
}

