% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContextObject.R
\name{SetAzureContext}
\alias{SetAzureContext}
\title{AzureSM: Updates the value of an AzureContext variable}
\usage{
SetAzureContext(AzureActiveContext, TID, CID, KEY, Token, SubscriptionID,
  ResourceGroup, StorageKey, StorageAccount, Container, Blob, VMName, HDIAdmin,
  HDIPassword, ClusterName, Kind, SessionID)
}
\arguments{
\item{TID}{Tenant ID Object}

\item{CID}{Client ID Object}

\item{KEY}{Authentication KEY Object}

\item{Token}{Token Object}

\item{SubscriptionID}{SubscriptionID Object}

\item{ResourceGroup}{ResourceGroup Object}

\item{StorageKey}{StorageKey Object}

\item{StorageAccount}{StorageAccount Object}

\item{Container}{Container Object}

\item{Blob}{Blob Object}

\item{VMName}{VMName Object}

\item{HDIAdmin}{HDIAdmin Object}

\item{HDIPassword}{HDIPassword Object}

\item{ClusterName}{ClusterName Object}

\item{Kind}{Kind Object}

\item{SessionID}{SessionID Object}

\item{Azure}{Context Object}

\item{Log}{Log Object#'}
}
\description{
AzureSM: SetAzureContext
}
\details{
Updates the value of an AzureContext variable
}
\author{
Alan Weaver
}

