% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AzureContextObject.R
\name{setAzureContext}
\alias{setAzureContext}
\title{Updates azureActiveContext object.}
\usage{
setAzureContext(azureActiveContext, tenantID, clientID, authKey, azToken,
  subscriptionID, resourceGroup, storageKey, storageAccount, container, blob,
  vmName, hdiAdmin, hdiPassword, clustername, kind, sessionID)
}
\arguments{
\item{azureActiveContext}{A container used for caching variables used by AzureSMR}

\item{tenantID}{The Tenant ID provided during creation of the Active Directory application / service principal}

\item{clientID}{The Client ID provided during creation of the Active Directory application / service principal}

\item{authKey}{The Authentication Key provided during creation of the Active Directory application / service principal}

\item{azToken}{Azure authentication token, obtained by \code{\link{azureAuthenticate}}}

\item{subscriptionID}{Set the subscriptionID.  This is obtained automatically by \code{\link{azureAuthenticate}} when only a single subscriptionID is available via Active Directory}

\item{resourceGroup}{Name of the resource group}

\item{storageKey}{Storage key associated with storage account}

\item{storageAccount}{Name of the azure storage account}

\item{container}{Storage container name. See \code{\link{azureListStorageContainers}}}

\item{blob}{Blob name}

\item{vmName}{Name of the virtual Machine}

\item{hdiAdmin}{HDInsight admin username}

\item{hdiPassword}{HDInsight admin password}

\item{clustername}{Cluster name, used for HDI and Spark clusters. See \code{\link{azureCreateHDI}}}

\item{kind}{HDinsight kind: "hadoop","spark" or "pyspark"}

\item{sessionID}{Spark sessionID. See \code{\link{azureSparkCMD}}}
}
\description{
Updates the value of an azureActiveContext object, created by \code{\link{createAzureContext}}
}
\seealso{
Other azureActiveContext functions: \code{\link{createAzureContext}}
}
